@@ -9,76 +9,107 @@
     "url": "https://github.com/aol/moloch.git"
   },
   "dependencies": {
-    "async": "^2.0.1",
-    "body-parser": "^1.15.2",
-    "bootstrap": "^3.3.7",
     "compression": "^1.6.2",
-    "elasticsearch": "^11.0.1",
-    "express": "^4.14.0",
-    "font-awesome": "^4.6.3",
     "fs-ext": "^0.5.0",
-    "iniparser": "http://github.com/awick/node-iniparser/tarball/master",
-    "jade": "^1.11.0",
-    "pug": "2.0.0-beta6",
+    "helmet": "^3.15.1",
     "method-override": "^2.3.6",
-    "moment": "^2.15.1",
-    "morgan": "^1.7.0",
     "multer": "^1.2.0",
     "passport": "^0.3.2",
     "passport-http": "^0.3.0",
     "peek-stream": "^1.1.1",
+    "pngjs": "^3.3.3",
+    "save-svg-as-png": "^1.4.5",
     "serve-favicon": "^2.3.0",
-    "stylus": "^0.54.5",
     "through2": "^2.0.1"
   },
   "scripts": {
+    "addtestuser": "node addUser.js -c ../tests/config.test.ini -n testuser admin admin admin --admin --packetSearch",
+    "build": "npm run bundle:min",
     "start": "node viewer.js -c ../config.ini",
-    "start:test": "node viewer.js -c ../tests/config.test.ini -n test & webpack -d --inline --watch --display-error-details",
-    "start:testuser": "node viewer.js -c ../tests/config.test.ini -n testuser & webpack -d --inline --watch --display-error-details",
-    "addtestuser": "node addUser.js -c ../tests/config.test.ini -n testuser admin admin admin --admin",
-    "start:dev": "node viewer.js -c config.dev.ini & webpack -d --inline --watch --display-error-details",
+    "start:dev": "node viewer.js -c config.dev.ini & webpack -d --watch --progress --hide-modules --display-error-details --config vueapp/build/webpack.dev.conf.js",
+    "start:test": "node viewer.js --debug -c ../tests/config.test.ini -n test & webpack -d --watch --progress --hide-modules --display-error-details --config vueapp/build/webpack.dev.conf.js",
+    "start:testuser": "node viewer.js -c ../tests/config.test.ini -n testuser & webpack -d --watch --progress --hide-modules --display-error-details --config vueapp/build/webpack.dev.conf.js",
+    "start:testuserdebug": "node viewer.js -c ../tests/config.test.ini -n testuser --debug & webpack -d --watch --progress --hide-modules --display-error-details --config vueapp/build/webpack.dev.conf.js",
+    "start:multiviewer": "node multies.js -c ../tests/config.test.ini -n all",
+    "start:testmultiviewer": "node viewer.js -c ../tests/config.test.ini -n testmulties & webpack -d --watch --progress --hide-modules --display-error-details --config vueapp/build/webpack.dev.conf.js",
+    "bundle": "webpack --progress --hide-modules -d --display-error-details --bail --config vueapp/build/webpack.dev.conf.js",
+    "bundle:min": "node vueapp/build/build.js",
     "prestart": "npm run bundle:min",
-    "build": "npm run test && npm run bundle:min",
-    "lint": "jshint app components",
-    "bundle": "webpack --progress --display-error-details --bail",
-    "bundle:min": "webpack -p --progress --bail",
-    "test": "karma start --single-run",
-    "test:watch": "karma start --auto-watch",
-    "pretest": "npm run lint",
-    "pretest:watch": "npm run lint"
+    "lint": "eslint --ext .js,.vue vueapp/src",
+    "prelint": "jshint viewer.js db.js config.js multies.js"
   },
   "devDependencies": {
-    "angular": "^1.5.8",
-    "angular-animate": "^1.5.8",
-    "angular-bind-notifier": "^1.1.7",
-    "angular-bootstrap-colorpicker": "^3.0.25",
-    "angular-filter": "^0.5.11",
-    "angular-mocks": "^1.5.8",
-    "angular-resource": "^1.5.8",
-    "angular-route": "^1.5.8",
-    "angular-sanitize": "^1.6.1",
-    "angular-ui-bootstrap": "2.3.0",
-    "babel-core": "^6.14.0",
-    "babel-loader": "^6.2.5",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "css-loader": "^0.24.0",
-    "file-loader": "^0.9.0",
-    "html-loader": "^0.4.3",
-    "jasmine-core": "^2.4.1",
-    "jquery": "^3.1.0",
-    "json-loader": "^0.5.4",
-    "karma": "^1.2.0",
-    "karma-jasmine": "^1.0.2",
-    "karma-phantomjs-launcher": "^1.0.2",
-    "karma-spec-reporter": "0.0.26",
-    "karma-webpack": "^1.8.0",
-    "ngdraggable": "^0.1.10",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.2"
+    "autoprefixer": "^7.1.2",
+    "babel-core": "^6.22.1",
+    "babel-eslint": "^10.0.3",
+    "babel-helper-vue-jsx-merge-props": "^2.0.3",
+    "babel-loader": "^7.1.1",
+    "babel-plugin-syntax-jsx": "^6.18.0",
+    "babel-plugin-transform-runtime": "^6.22.0",
+    "babel-plugin-transform-vue-jsx": "^3.5.0",
+    "babel-preset-env": "^1.3.2",
+    "babel-preset-stage-2": "^6.22.0",
+    "bootstrap": "^4.3.1",
+    "bootstrap-vue": "^2.0.0-rc.11",
+    "chalk": "^2.0.1",
+    "copy-webpack-plugin": "4.5.1",
+    "css-loader": "^2.1.1",
+    "d3": "^5.7.0",
+    "eslint": "^6.2.2",
+    "eslint-config-standard": "^12.0.0-alpha.0",
+    "eslint-friendly-formatter": "^4.0.1",
+    "eslint-loader": "^3.0.0",
+    "eslint-plugin-import": "^2.13.0",
+    "eslint-plugin-node": "^6.0.1",
+    "eslint-plugin-promise": "^3.8.0",
+    "eslint-plugin-standard": "^3.1.0",
+    "eslint-plugin-vue": "^5.2.3",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "file-loader": "^1.1.4",
+    "friendly-errors-webpack-plugin": "^1.6.1",
+    "html-webpack-plugin": "^3.2.0",
+    "jquery": "^3.4.1",
+    "js-yaml": "^3.13.1",
+    "jshint": "^2.9.6",
+    "json-loader": "^0.5.7",
+    "lodash": "^4.17.13",
+    "lodash.mergewith": "^4.6.2",
+    "moment-timezone": "^0.5.14",
+    "node-notifier": "^5.1.2",
+    "optimize-css-assets-webpack-plugin": "^5.0.1",
+    "ora": "^1.2.0",
+    "portfinder": "^1.0.13",
+    "postcss-import": "^11.0.0",
+    "postcss-loader": "2.1.4",
+    "postcss-url": "^7.2.1",
+    "qs": "^6.5.1",
+    "rimraf": "^2.6.0",
+    "sanitize-html": "^1.18.2",
+    "semver": "^5.3.0",
+    "shelljs": "^0.7.6",
+    "style-loader": "^0.19.0",
+    "uglifyjs-webpack-plugin": "1.2.5",
+    "url-loader": "^1.0.1",
+    "vue": "^2.6.10",
+    "vue-bootstrap-datetimepicker": "^5.0.1",
+    "vue-clickaway": "^2.2.2",
+    "vue-clipboard2": "^0.3.0",
+    "vue-color": "^2.4.6",
+    "vue-loader": "^15.7.0",
+    "vue-style-loader": "^3.0.1",
+    "webpack": "^4.30.0",
+    "webpack-bundle-analyzer": "^3.3.2",
+    "webpack-cli": "^3.3.1",
+    "webpack-dev-server": "^3.3.1",
+    "webpack-merge": "4.1.2"
   },
-  "optionalDependencies": {
-    "png": "http://github.com/awick/node-png/tarball/master"
-  }
+  "engines": {
+    "node": ">= 6.0.0",
+    "npm": ">= 3.0.0"
+  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "not ie <= 8"
+  ]
 }
