@@ -1,4 +1,4 @@
-/* parser generated by jison 0.4.15 */
+/* parser generated by jison 0.4.18 */
 /*
   Returns a Parser object of the following structure:
 
@@ -71,9 +71,9 @@
     recoverable: (boolean: TRUE when the parser has a error recovery rule available for this particular error)
   }
 */
-var parser = (function(){
+var molochparser = (function(){
 var o=function(k,v,o,l){for(o=o||{},l=k.length;l--;o[k[l]]=v);return o},$V0=[1,6],$V1=[1,3],$V2=[1,4],$V3=[1,5],$V4=[1,8],$V5=[1,9],$V6=[5,18,19,23],$V7=[14,15,16,17];
-var parser = {trace: function trace() { },
+var parser = {trace: function trace () { },
 yy: {},
 symbols_: {"error":2,"expressions":3,"e":4,"EOF":5,"OP":6,"lt":7,"lte":8,"gt":9,"gte":10,"==":11,"!=":12,"VALUE":13,"STR":14,"QUOTEDSTR":15,"REGEXSTR":16,"LIST":17,"&&":18,"||":19,"!":20,"-":21,"(":22,")":23,"EXISTS":24,"$accept":0,"$end":1},
 terminals_: {2:"error",5:"EOF",7:"lt",8:"lte",9:"gt",10:"gte",11:"==",12:"!=",14:"STR",15:"QUOTEDSTR",16:"REGEXSTR",17:"LIST",18:"&&",19:"||",20:"!",21:"-",22:"(",23:")",24:"EXISTS"},
@@ -134,11 +134,13 @@ break;
 },
 table: [{3:1,4:2,14:$V0,20:$V1,21:$V2,22:$V3},{1:[3]},{5:[1,7],18:$V4,19:$V5},{4:10,14:$V0,20:$V1,21:$V2,22:$V3},{4:11,14:$V0,20:$V1,21:$V2,22:$V3},{4:12,14:$V0,20:$V1,21:$V2,22:$V3},{6:15,7:[1,16],8:[1,17],9:[1,18],10:[1,19],11:[1,13],12:[1,14]},{1:[2,1]},{4:20,14:$V0,20:$V1,21:$V2,22:$V3},{4:21,14:$V0,20:$V1,21:$V2,22:$V3},o($V6,[2,14]),o($V6,[2,15]),{18:$V4,19:$V5,23:[1,22]},o($V7,[2,6],{24:[1,23]}),o($V7,[2,7],{24:[1,24]}),{13:25,14:[1,26],15:[1,27],16:[1,28],17:[1,29]},o($V7,[2,2]),o($V7,[2,3]),o($V7,[2,4]),o($V7,[2,5]),o($V6,[2,12]),o([5,19,23],[2,13],{18:$V4}),o($V6,[2,16]),o($V6,[2,17]),o($V6,[2,18]),o($V6,[2,19]),o($V6,[2,8]),o($V6,[2,9]),o($V6,[2,10]),o($V6,[2,11])],
 defaultActions: {7:[2,1]},
-parseError: function parseError(str, hash) {
+parseError: function parseError (str, hash) {
     if (hash.recoverable) {
         this.trace(str);
     } else {
-        throw new Error(str);
+        var error = new Error(str);
+        error.hash = hash;
+        throw error;
     }
 },
 parse: function parse(input) {
@@ -171,14 +173,14 @@ parse: function parse(input) {
         lstack.length = lstack.length - n;
     }
     _token_stack:
-        function lex() {
+        var lex = function () {
             var token;
             token = lexer.lex() || EOF;
             if (typeof token !== 'number') {
                 token = self.symbols_[token] || token;
             }
             return token;
-        }
+        };
     var symbol, preErrorSymbol, state, action, a, r, yyval = {}, p, len, newState, expected;
     while (true) {
         state = stack[stack.length - 1];
@@ -283,36 +285,92 @@ parse: function parse(input) {
 var    util           = require('util');
 var    moment         = require('moment');
 
+/* Build a list of all the field infos for ip field types.
+ * Can specify if a port field needs to be available for the type or not
+ */
+function getIpInfoList(yy, needPort)
+{
+  let ors = [];
+  let completed = {};
+
+  for (field in yy.fieldsMap) {
+    let info = yy.fieldsMap[field];
+
+    // If ip itself or not an ip field stop
+    if (field === "ip" || info.type !== "ip")
+      continue;
+
+    // Already completed
+    if (completed[info.dbField])
+      continue;
+    completed[info.dbField] = 1;
+
+    // If port specified then skip ips without ports
+    if (needPort && !info.portField) {
+      continue;
+    }
+
+    if (info.requiredRight && yy[info.requiredRight] !== true) {
+      continue;
+    }
+
+    ors.push(info);
+  }
+
+  return ors;
+}
+
+/* Do all the magic around ip field parsing.
+ * Supports many formats such as
+ * ip
+ * ip/cidr
+ * ip1 = ip1.0.0.0/24
+ * ip1.ip2 = ip1.ip2.0.0/16
+ * ip1.ip2.ip3 = ip1.ip2.ip3.0/8
+ * All of the above with a :port(v4) or .port(v6) at the end
+ * Arrays of all of the above
+ */
 function parseIpPort(yy, field, ipPortStr) {
   var dbField = yy.fieldsMap[field].dbField;
 
-  function singleIp(dbField, ip1, ip2, port) {
+  // Have just a single Ip, create obj for it
+  function singleIp(exp, dbField, ip, port) {
     var obj;
 
-    if (ip1 !== undefined) {
-      if (ip1 === ip2) {
-        obj = {term: {}};
-        obj.term[dbField] = ip1>>>0;
-      } else {
-        obj = {range: {}};
-        obj.range[dbField] = {from: ip1>>>0, to: ip2>>>0};
-      }
+    if (ip !== undefined) {
+      obj = {term: {}};
+      obj.term[dbField] = ip;
     }
 
     if (port !== -1) {
-      if (yy.fieldsMap[field].portField) {
+      if (yy.fieldsMap[exp].portField) {
         obj = {bool: {must: [obj, {term: {}}]}};
-        obj.bool.must[1].term[yy.fieldsMap[field].portField] = port;
+        obj.bool.must[1].term[yy.fieldsMap[exp].portField] = port;
       } else {
-        throw field + " doesn't support port";
+        throw exp + " doesn't support port";
       }
 
-      if (ip1 === undefined) {
+      if (ip === undefined) {
         obj = obj.bool.must[1];
       }
     }
 
     return obj;
+  }
+
+  // Special case of ip=
+  function allIp(ip, port) {
+    let infos = getIpInfoList(yy, port !== -1);
+
+    var ors = [];
+    for (let info of infos) {
+      obj = singleIp(info.exp, info.dbField, ip, port);
+      if (obj) {
+        ors.push(obj);
+      }
+    }
+
+    return {bool: {should: ors}};
   }
 
 
@@ -333,72 +391,72 @@ function parseIpPort(yy, field, ipPortStr) {
       return obj;
   }
 
-  // Support '10.10.10/16:4321'
+  // Support ':80' and '.80'
+  if ((ipPortStr[0] === ':' && ipPortStr[0] !== ':') ||
+      (ipPortStr[0] === '.')) {
+    if (dbField !== "ipall") {
+      return singleIp(field, dbField, undefined, +ipPortStr.slice(1));
+    } else {
+      return allIp(undefined, +ipPortStr.slice(1));
+    }
+  }
 
-  var ip1, ip2;
-  var colons = ipPortStr.split(':');
-  var slash = colons[0].split('/');
-  var dots = slash[0].split('.');
+  // Support ip4: '10.10.10.10' '10.10.10/16:80' '10.10.10:80' '10.10.10/16'
+  // Support ip6: '1::2' '1::2/16.80' '1::2.80' '1::2/16'
+  var ip;
   var port = -1;
-  if (colons[1]) {
-    port = parseInt(colons[1], 10);
-  }
+  var colons = ipPortStr.split(':');
 
-  if (dots.length === 4) {
-    ip1 = ip2 = (parseInt(dots[0], 10) << 24) | (parseInt(dots[1], 10) << 16) | (parseInt(dots[2], 10) << 8) | parseInt(dots[3], 10);
-  } else if (dots.length === 3) {
-    ip1 = (parseInt(dots[0], 10) << 24) | (parseInt(dots[1], 10) << 16) | (parseInt(dots[2], 10) << 8);
-    ip2 = (parseInt(dots[0], 10) << 24) | (parseInt(dots[1], 10) << 16) | (parseInt(dots[2], 10) << 8) | 255;
-  } else if (dots.length === 2) {
-    ip1 = (parseInt(dots[0], 10) << 24) | (parseInt(dots[1], 10) << 16);
-    ip2 = (parseInt(dots[0], 10) << 24) | (parseInt(dots[1], 10) << 16) | (255 << 8) | 255;
-  } else if (dots.length === 1 && dots[0].length > 0) {
-    ip1 = (parseInt(dots[0], 10) << 24);
-    ip2 = (parseInt(dots[0], 10) << 24) | (255 << 16) | (255 << 8) | 255;
-  }
-
-  // Can't shift by 32 bits in javascript, who knew!
-  if (slash[1] && slash[1] !== '32') {
-    if (ip1 === undefined) {
-      ip1 = ip2 = 0xffffffff;
+  // More then 1 colon is ip 6
+  if (colons.length > 2) {
+    // Everything after . is port
+    let dots = ipPortStr.split('.');
+    if (dots.length > 1 && dots[1] !== '') {
+      port = +dots[1];
     }
-    var s = parseInt(slash[1], 10);
-    ip1 = ip1 & (0xffffffff << (32 - s));
-    ip2 = ip2 | (0xffffffff >>> s);
+    // Everything before . is ip and slash
+    ip = dots[0];
+  } else {
+    // everything after : is port
+    if (colons.length > 1 && colons[1] !== '') {
+      port = +colons[1];
+    }
+
+    // Have to do extra because we allow shorthand for /8, /16, /24
+    let slash = colons[0].split('/');
+    let dots = slash[0].split('.');
+
+    switch(dots.length) {
+    case 4:
+      ip = `${dots[0]}.${dots[1]}.${dots[2]}.${dots[3]}`;
+      break;
+    case 3:
+      ip = `${dots[0]}.${dots[1]}.${dots[2]}.0`;
+      if (slash[1] === undefined) {slash[1] = '24';}
+      break;
+    case 2:
+      ip = `${dots[0]}.${dots[1]}.0.0`;
+      if (slash[1] === undefined) {slash[1] = '16';}
+      break;
+    case 1:
+      if (dots[0].length > 0) {
+        ip = `${dots[0]}.0.0.0`;
+        if (slash[1] === undefined) {slash[1] = '8';}
+      }
+      break;
+    }
+
+    // Add the slash back to the ip
+    if (slash[1] && slash[1] !== '32') {
+      ip = `${ip}/${slash[1]}`;
+    }
   }
-  
+
   if (dbField !== "ipall") {
-    return singleIp(dbField, ip1, ip2, port);
+    return singleIp(field, dbField, ip, port);
+  } else {
+    return allIp(ip, port);
   }
-
-  var ors = [];
-  var completed = {};
-  for (field in yy.fieldsMap) {
-    var info = yy.fieldsMap[field];
-
-    // If ip itself or not an ip field stop
-    if (field === "ip" || info.type !== "ip")
-      continue;
-
-    // Already completed
-    if (completed[info.dbField])
-      continue;
-    completed[info.dbField] = 1;
-
-    // If port specified then skip ips without ports
-    if (port !== -1 && !info.portField)
-      continue;
-
-    if (info.requiredRight && yy[info.requiredRight] !== true) {
-      continue;
-    }
-    obj = singleIp(info.dbField, ip1, ip2, port);
-    if (obj) {
-      ors.push(obj);
-    }
-  }
-
-  return {bool: {should: ors}};
 }
 
 function stripQuotes (str) {
@@ -485,7 +543,80 @@ function formatQuery(yy, field, op, value)
     throw "Invalid operator '" + op + "' for " + field;
   }
 
-  switch (info.type) {
+  /* look for value that starts with $ */
+  if (value[0] === '$') {
+    if (op !== "eq" && op !== "ne") {
+      throw 'Shortcuts only support == and !=';
+    }
+
+    value = value.substr(1); /* remove $ */
+    if (!yy.lookups || !yy.lookups[value]) {
+      throw value + ' - Shortcut not found';
+    }
+
+    var lookup = yy.lookups[value];
+
+    obj = { terms: {} };
+    obj.terms[info.dbField] = {
+      index : `${yy.prefix}lookups`,
+      id : lookup._id,
+      type: 'lookup'
+    };
+
+    var type = info.type2 || info.type;
+    var lookupType = yy.lookupTypeMap[type];
+
+    if (!lookupType) {
+      throw "Unsupported field type: " + type
+    }
+
+    if (!lookup._source[lookupType]) {
+      throw 'lookup must be of type ' + lookupType;
+    }
+
+    switch (type) {
+    case 'ip':
+      if (field === 'ip') {
+        let infos = getIpInfoList(yy, false);
+        var ors = [];
+        for (let info of infos) {
+          obj = formatQuery(yy, info.exp, op, '$' + value);
+          if (obj) {
+            ors.push(obj);
+          }
+        }
+        obj = {bool: {should: ors}};
+      } else {
+        obj.terms[info.dbField].path = 'ip';
+      }
+
+      if (op === "ne") {
+        return { bool: { must_not: obj } };
+      }
+      return obj;
+    case 'integer':
+      obj.terms[info.dbField].path = 'number';
+      if (op === "ne") {
+        return { bool: { must_not: obj } };
+      }
+      return obj;
+    case 'lotermfield':
+    case 'lotextfield':
+    case 'termfield':
+    case 'textfield':
+    case 'uptermfield':
+    case 'uptextfield':
+      obj.terms[info.dbField].path = 'string';
+      if (op === "ne") {
+        return { bool: { must_not: obj } };
+      }
+      return obj;
+    default:
+      throw "Unsupported field type: " + type;
+    }
+  }
+
+  switch (info.type2 || info.type) {
   case "ip":
     if (value[0] === "/")
       throw value + " - Regex not supported for ip queries";
@@ -493,11 +624,28 @@ function formatQuery(yy, field, op, value)
     if (value.indexOf("*") !== -1)
       throw value + " - Wildcard not supported for ip queries";
 
+    if (value === "ipv4") {
+      value = "0.0.0.0/0";
+    } if (value === "ipv6") {
+      value = "0.0.0.0/0";
+      if (op === "ne")
+        op = "eq";
+      else
+        op = "ne";
+    }
+
     if (op === "eq")
       return parseIpPort(yy, field, value);
     if (op === "ne")
       return {bool: {must_not: parseIpPort(yy, field, value)}};
-    throw "Invalid operator '" + op + "' for ip";
+
+    if (value[0] === "\[")
+      throw value + " - List queries not supported for gt/lt queries - " + value;
+
+    obj = {range: {}};
+    obj.range[info.dbField] = {};
+    obj.range[info.dbField][op] = value;
+    return obj;
   case "integer":
     if (value[0] === "/")
       throw value + " - Regex queries not supported for integer queries";
@@ -581,6 +729,25 @@ function formatQuery(yy, field, op, value)
     obj.range[info.dbField] = {};
     obj.range[info.dbField][op] = parseSeconds(stripQuotes(value));
     return obj;
+  case "date":
+    if (value[0] === "/")
+      throw value + " - Regex queries not supported for date queries";
+
+    if (op === "eq" || op === "ne") {
+      obj = termOrTermsDate(info.dbField, value);
+      if (op === "ne") {
+        obj = {bool: {must_not: obj}};
+      }
+      return obj;
+    }
+
+    if (value[0] === "\[")
+      throw value + " - List queries not supported for gt/lt queries - " + value;
+
+    obj = {range: {}};
+    obj.range[info.dbField] = {};
+    obj.range[info.dbField][op] = moment.unix(parseSeconds(stripQuotes(value))).format();
+    return obj;
   default:
     throw "Unknown field type: " + info.type;
   }
@@ -650,6 +817,10 @@ function stringQuery(yy, field, str) {
         should.regexp[rawField] = str.substring(1, str.length-1);
         obj.push(should);
       } else if (typeof str === "string" && str.indexOf("*") !== -1) {
+        if (str === "*") {
+          throw "Please use 'EXISTS!' instead of a '*' in expression";
+        }
+
         should = {wildcard: {}};
         should.wildcard[rawField] = str;
         obj.push(should);
@@ -667,8 +838,8 @@ function stringQuery(yy, field, str) {
           }
           terms.terms[dbField].push(str);
         } else {
-          should = {match: {}};
-          should.match[dbField] = {query: str, type: "phrase", operator: "and"}
+          should = {match_phrase: {}};
+          should.match_phrase[dbField] = str;
           obj.push(should);
         }
       }
@@ -691,12 +862,15 @@ function stringQuery(yy, field, str) {
     obj = {term: {}};
     obj.term[dbField] = str;
   } else if (typeof str === "string" && str.indexOf("*") !== -1) {
+    if (str === "*") {
+      throw "Please use 'EXISTS!' instead of a '*' in expression";
+    }
     dbField = field2Raw(yy, field);
     obj = {wildcard: {}};
     obj.wildcard[dbField] = str;
   } else if (info.type.match(/textfield/)) {
-    obj = {match: {}};
-    obj.match[dbField] = {query: str, type: "phrase", operator: "and"}
+    obj = {match_phrase: {}};
+    obj.match_phrase[dbField] = str;
   } else if (info.type.match(/termfield/)) {
     obj = {term: {}};
     obj.term[dbField] = str;
@@ -707,61 +881,52 @@ function stringQuery(yy, field, str) {
 
 if (!global.moloch) global.moloch = {};
 global.moloch.utf8ToHex = function (utf8) {
-    var hex = new Buffer(stripQuotes(utf8)).toString("hex").toLowerCase();
+    var hex = Buffer.from(stripQuotes(utf8)).toString("hex").toLowerCase();
     hex = hex.replace(/2a/g, '*');
     return hex;
 }
 
-global.moloch.ipv6ToHex = function (ip) {
-  ip = stripQuotes(ip);
-  if (ip[0] === "[") {
-    var closing = ip.indexOf(']');
-    ip =  ip.substring(1, closing);
-  }
-
-  if (ip.indexOf("*") !== -1 && ip.indexOf("::") !== -1) {
-    throw "Can't use :: in ipv6 and * at the same time";
-  }
-  var parts = ip.split(":");
-  for (var i = 0; i < parts.length; i++) {
-    if (parts[i].indexOf("*") !== -1) {
-      continue;
-    }
-    if (parts[i] === "") {
-      for (var j = 0; j <= 8 - parts.length; j++) {
-        parts[i] += "0000";
-      }
-    } else {
-      while (parts[i].length < 4) {
-        parts[i] = "0" + parts[i];
-      }
-    }
-  }
-  return parts.join("");
+global.moloch.dash2Colon = function (str) {
+  return str.replace(/-/g, ':');
 }
 
 var protocols = {
     icmp:   1,
     tcp:    6,
     udp:    17,
+    esp:    50,
     icmp6:  58,
-    icmpv6: 58
+    icmpv6: 58,
+    sctp:   132,
 };
 
 global.moloch.ipProtocolLookup = function (text) {
-    if (typeof text !== "string") {
-        for (var i = 0; i < text.length; i++) {
-            if (!protocols[text[i]] && isNaN(text[i]))
-                throw ("Unknown protocol string " + text);
-            text[i] = protocols[text[i]] || +text[i];
-        }
-        return text;
-    } else {
-        if (!protocols[text] && isNaN(text))
-            throw ("Unknown protocol string " + text);
-        return protocols[text] || +text;
+  if (typeof text !== "string") {
+    for (var i = 0; i < text.length; i++) {
+      if (!protocols[text[i]] && isNaN(text[i]))
+        throw ("Unknown protocol string " + text);
+      text[i] = protocols[text[i]] || +text[i];
     }
+    return text;
+  } else {
+    if (!protocols[text] && isNaN(text))
+      throw ("Unknown protocol string " + text);
+    return protocols[text] || +text;
+  }
 };
+
+// Remove the "http://", "https://", etc from http.uri queries
+global.moloch.removeProtocol = function (text) {
+  text = text.replace(/^[a-z]+:\/\//i, '');
+  return text;
+}
+
+// Remove the "http://", "https://" and after the first slash, etc from host queries
+global.moloch.removeProtocolAndURI = function (text) {
+  text = text.replace(/^[a-z]+:\/\//i, '');
+  text = text.replace(/\/.*/, '');
+  return text;
+}
 
 function ListToArray(text) {
   if (text[0] !== "[" || text[text.length -1] !== "]")
@@ -788,8 +953,14 @@ function termOrTermsInt(dbField, str) {
     });
   } else {
     str = stripQuotes(str);
-    if (str.match(/[^\d]+/))
+    let match;
+    if ((match = str.match(/(-?\d+)-(-?\d+)/))) {
+      obj = {range: {}};
+      obj.range[dbField] = {gte: match[1], lte: match[2]};
+      return obj;
+    } else if (str.match(/[^\d]+/)) {
       throw str + " is not a number";
+    }
     obj = {term: {}};
     obj.term[dbField] = str;
   }
@@ -808,6 +979,23 @@ function termOrTermsSeconds(dbField, str) {
     str = parseSeconds(stripQuotes(str));
     obj = {term: {}};
     obj.term[dbField] = str;
+  }
+  return obj;
+}
+
+function termOrTermsDate(dbField, str) {
+  var obj = {};
+  if (str[0] === "[" && str[str.length -1] === "]") {
+    obj = {bool: {should: []}};
+    ListToArray(str).forEach(function(str) {
+      let r = {range: {}};
+      r.range[dbField] = {gte: d, lte: d};
+      obj.bool.should.push(r);
+    });
+  } else {
+    let d = moment.unix(parseSeconds(stripQuotes(str))).format();
+    obj = {range: {}};
+    obj.range[dbField] = {gte: d, lte: d};
   }
   return obj;
 }
@@ -847,7 +1035,7 @@ function parseSeconds(str) {
     }
 
     if (snap) {
-      d.startOf(snap); 
+      d.startOf(snap);
       if (n = m[5].match(/^(w|week|weeks)(\d+)$/)) {
         d.day(n[2]);
       }
@@ -1014,7 +1202,7 @@ showPosition:function () {
     },
 
 // test the lexed token: return FALSE when not a match, otherwise return token
-test_match:function (match, indexed_rule) {
+test_match:function(match, indexed_rule) {
         var token,
             lines,
             backup;
@@ -1144,7 +1332,7 @@ next:function () {
     },
 
 // return next match that has a token
-lex:function lex() {
+lex:function lex () {
         var r = this.next();
         if (r) {
             return r;
@@ -1154,12 +1342,12 @@ lex:function lex() {
     },
 
 // activates a new lexer condition state (pushes the new lexer condition state onto the condition stack)
-begin:function begin(condition) {
+begin:function begin (condition) {
         this.conditionStack.push(condition);
     },
 
 // pop the previously active lexer condition state off the condition stack
-popState:function popState() {
+popState:function popState () {
         var n = this.conditionStack.length - 1;
         if (n > 0) {
             return this.conditionStack.pop();
@@ -1169,7 +1357,7 @@ popState:function popState() {
     },
 
 // produce the lexer rule set which is active for the currently active lexer condition state
-_currentRules:function _currentRules() {
+_currentRules:function _currentRules () {
         if (this.conditionStack.length && this.conditionStack[this.conditionStack.length - 1]) {
             return this.conditions[this.conditionStack[this.conditionStack.length - 1]].rules;
         } else {
@@ -1178,7 +1366,7 @@ _currentRules:function _currentRules() {
     },
 
 // return the currently active lexer condition state; when an index argument is provided it produces the N-th previous condition state, if available
-topState:function topState(n) {
+topState:function topState (n) {
         n = this.conditionStack.length - 1 - Math.abs(n || 0);
         if (n >= 0) {
             return this.conditionStack[n];
@@ -1188,7 +1376,7 @@ topState:function topState(n) {
     },
 
 // alias for begin(condition)
-pushState:function pushState(condition) {
+pushState:function pushState (condition) {
         this.begin(condition);
     },
 
@@ -1248,7 +1436,7 @@ case 22:console.log(yy_.yytext);
 break;
 }
 },
-rules: [/^(?:\s+)/,/^(?:"(?:\\?.)*?")/,/^(?:\/(?:\\?.)*?\/)/,/^(?:[-+a-zA-Z0-9_.@:*?\/]+)/,/^(?:\[[^\]\\]*(?:\\.[^\]\\]*)*\])/,/^(?:EXISTS!)/,/^(?:<=)/,/^(?:<)/,/^(?:>=)/,/^(?:>)/,/^(?:!=)/,/^(?:==)/,/^(?:=)/,/^(?:\|\|)/,/^(?:\|)/,/^(?:&&)/,/^(?:&)/,/^(?:\()/,/^(?:\))/,/^(?:!)/,/^(?:$)/,/^(?:.)/,/^(?:.)/],
+rules: [/^(?:\s+)/,/^(?:"(?:\\?.)*?")/,/^(?:\/(?:\\?.)*?\/)/,/^(?:[-+a-zA-Z0-9_.@:*?\/$]+)/,/^(?:\[[^\]\\]*(?:\\.[^\]\\]*)*\])/,/^(?:EXISTS!)/,/^(?:<=)/,/^(?:<)/,/^(?:>=)/,/^(?:>)/,/^(?:!=)/,/^(?:==)/,/^(?:=)/,/^(?:\|\|)/,/^(?:\|)/,/^(?:&&)/,/^(?:&)/,/^(?:\()/,/^(?:\))/,/^(?:!)/,/^(?:$)/,/^(?:.)/,/^(?:.)/],
 conditions: {"INITIAL":{"rules":[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22],"inclusive":true}}
 });
 return lexer;
@@ -1263,10 +1451,10 @@ return new Parser;
 
 
 if (typeof require !== 'undefined' && typeof exports !== 'undefined') {
-exports.parser = parser;
-exports.Parser = parser.Parser;
-exports.parse = function () { return parser.parse.apply(parser, arguments); };
-exports.main = function commonjsMain(args) {
+exports.parser = molochparser;
+exports.Parser = molochparser.Parser;
+exports.parse = function () { return molochparser.parse.apply(molochparser, arguments); };
+exports.main = function commonjsMain (args) {
     if (!args[1]) {
         console.log('Usage: '+args[0]+' FILE');
         process.exit(1);
