@@ -38,13 +38,12 @@ typedef int (* HASH_CMP_FUNC)(const void *key, const void *element);
 
 #define HASH_INIT(name, varname, hashfunc, cmpfunc) \
   do { \
-       int i; \
        (varname).size = sizeof((varname).buckets)/sizeof((varname).buckets[0]); \
        (varname).hash = hashfunc; \
        (varname).cmp = cmpfunc; \
        (varname).count = 0; \
-       for (i = 0; i < (varname).size; i++) { \
-           DLL_INIT(name, &((varname).buckets[i])); \
+       for (int _i = 0; _i < (varname).size; _i++) { \
+           DLL_INIT(name, &((varname).buckets[_i])); \
        } \
      } while (0)
 
@@ -61,14 +60,13 @@ typedef int (* HASH_CMP_FUNC)(const void *key, const void *element);
 
 #define HASHP_INIT(name, varname, sz, hashfunc, cmpfunc) \
   do { \
-       int i; \
        (varname).size = sz; \
        (varname).hash = hashfunc; \
        (varname).cmp = cmpfunc; \
        (varname).count = 0; \
        (varname).buckets = malloc(sz * sizeof((varname).buckets[0])); \
-       for (i = 0; i < (varname).size; i++) { \
-           DLL_INIT(name, &((varname).buckets[i])); \
+       for (int _i = 0; _i < (varname).size; _i++) { \
+           DLL_INIT(name, &((varname).buckets[_i])); \
        } \
      } while (0)
 
@@ -116,26 +114,22 @@ typedef int (* HASH_CMP_FUNC)(const void *key, const void *element);
 
 #define HASH_COUNT(name, varname) ((varname).count)
 
+#define HASH_BUCKET_COUNT(name, varname, h) ((varname).buckets[((uint32_t)h) % (varname).size].name##count)
+
 #define HASH_FORALL_POP_HEAD(name, varname, element, code) \
-  do { \
-      int  _##name##b; \
-      for ( _##name##b = 0;  _##name##b < (varname).size;  _##name##b++) {\
-          while((varname).buckets[_##name##b].name##count) { \
-              DLL_POP_HEAD(name, &((varname).buckets[_##name##b]), element); \
-              (varname).count--; \
-              code \
-          } \
+  for ( int _##name##b = 0;  _##name##b < (varname).size;  _##name##b++) {\
+      while((varname).buckets[_##name##b].name##count) { \
+          DLL_POP_HEAD(name, &((varname).buckets[_##name##b]), element); \
+          (varname).count--; \
+          code \
       } \
-  } while(0)
+  }
 
 #define HASH_FORALL(name, varname, element, code) \
-  do { \
-      int  _##name##b; \
-      for ( _##name##b = 0;  _##name##b < (varname).size;  _##name##b++) {\
-          for (element = (varname).buckets[_##name##b].name##next; element != (void*)&((varname).buckets[_##name##b]); element = element->name##next) { \
-              code \
-          } \
+  for ( int _##name##b = 0;  _##name##b < (varname).size;  _##name##b++) {\
+      for (element = (varname).buckets[_##name##b].name##next; element != (void*)&((varname).buckets[_##name##b]); element = element->name##next) { \
+          code \
       } \
-  } while(0)
+  }
 
 #endif
