@@ -19,6 +19,7 @@
 #include "patricia.h"
 #include <inttypes.h>
 #include <arpa/inet.h>
+#include <errno.h>
 
 //#define DEBUG_PACKET
 
@@ -27,27 +28,50 @@ extern MolochConfig_t        config;
 
 MolochPcapFileHdr_t          pcapFileHeader;
 
-uint64_t                     totalPackets = 0;
-uint64_t                     totalBytes = 0;
-uint64_t                     totalSessions = 0;
+uint64_t                     totalPackets;
+LOCAL uint64_t               totalBytes[MOLOCH_MAX_PACKET_THREADS];
 
 LOCAL uint32_t               initialDropped = 0;
-struct timeval               initialPacket;
+struct timeval               initialPacket; // Don't make LOCAL for now because of netflow plugin
 
 extern void                 *esServer;
 extern uint32_t              pluginsCbs;
 
+uint64_t                     writtenBytes;
+uint64_t                     unwrittenBytes;
+
 LOCAL int                    mac1Field;
 LOCAL int                    mac2Field;
+LOCAL int                    oui1Field;
+LOCAL int                    oui2Field;
 LOCAL int                    vlanField;
 LOCAL int                    greIpField;
+LOCAL int                    icmpTypeField;
+LOCAL int                    icmpCodeField;
 
 LOCAL uint64_t               droppedFrags;
 
 time_t                       lastPacketSecs[MOLOCH_MAX_PACKET_THREADS];
-int                          inProgress[MOLOCH_MAX_PACKET_THREADS];
+LOCAL int                    inProgress[MOLOCH_MAX_PACKET_THREADS];
 
-LOCAL patricia_tree_t       *ipTree = 0;
+LOCAL patricia_tree_t       *ipTree4 = 0;
+LOCAL patricia_tree_t       *ipTree6 = 0;
+
+LOCAL int                    maxTcpOutOfOrderPackets;
+
+extern MolochFieldOps_t      readerFieldOps[256];
+
+/******************************************************************************/
+
+#define MOLOCH_PACKET_SUCCESS          0
+#define MOLOCH_PACKET_IP_DROPPED       1
+#define MOLOCH_PACKET_OVERLOAD_DROPPED 2
+#define MOLOCH_PACKET_CORRUPT          3
+#define MOLOCH_PACKET_UNKNOWN          4
+#define MOLOCH_PACKET_IPPORT_DROPPED   5
+#define MOLOCH_PACKET_MAX              6
+
+LOCAL uint64_t               packetStats[MOLOCH_PACKET_MAX];
 
 /******************************************************************************/
 extern MolochSessionHead_t   tcpWriteQ[MOLOCH_MAX_PACKET_THREADS];
@@ -55,9 +79,13 @@ extern MolochSessionHead_t   tcpWriteQ[MOLOCH_MAX_PACKET_THREADS];
 LOCAL  MolochPacketHead_t    packetQ[MOLOCH_MAX_PACKET_THREADS];
 LOCAL  uint32_t              overloadDrops[MOLOCH_MAX_PACKET_THREADS];
 
-MOLOCH_LOCK_DEFINE(frags);
+LOCAL  MOLOCH_LOCK_DEFINE(frags);
 
-int moloch_packet_ip4(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len);
+LOCAL int moloch_packet_ip4(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len);
+LOCAL int moloch_packet_ip6(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len);
+LOCAL int moloch_packet_frame_relay(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len);
+LOCAL int moloch_packet_ppp(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len);
+LOCAL int moloch_packet_ether(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len);
 
 typedef struct molochfrags_t {
     struct molochfrags_t  *fragh_next, *fragh_prev;
@@ -83,8 +111,16 @@ typedef HASH_VAR(h_, MolochFragsHash_t, MolochFragsHead_t, 199337);
 MolochFragsHash_t          fragsHash;
 MolochFragsHead_t          fragsList;
 
+// These are in network byte order
+MolochDropHashGroup_t      packetDrop4;
+MolochDropHashGroup_t      packetDrop6;
+
+#ifndef IPPROTO_IPV4
+#define IPPROTO_IPV4            4
+#endif
+
 /******************************************************************************/
-void moloch_packet_free(MolochPacket_t *packet)
+LOCAL void moloch_packet_free(MolochPacket_t *packet)
 {
     if (packet->copied) {
         free(packet->pkt);
@@ -95,6 +131,16 @@ void moloch_packet_free(MolochPacket_t *packet)
 /******************************************************************************/
 void moloch_packet_tcp_free(MolochSession_t *session)
 {
+    if (session->tcpData.td_count == 1 && session->tcpFlagCnt[MOLOCH_TCPFLAG_PSH] == 1) {
+        MolochTcpData_t *ftd = DLL_PEEK_HEAD(td_, &session->tcpData);
+        const int which = ftd->packet->direction;
+        const uint8_t *data = ftd->packet->pkt + ftd->dataOffset;
+        const int len = ftd->len;
+
+        moloch_parsers_classify_tcp(session, data, len, which);
+        moloch_packet_process_data(session, data, len, which);
+    }
+
     MolochTcpData_t *td;
     while (DLL_POP_HEAD(td_, &session->tcpData, td)) {
         moloch_packet_free(td->packet);
@@ -103,13 +149,13 @@ void moloch_packet_tcp_free(MolochSession_t *session)
 }
 /******************************************************************************/
 // Idea from gopacket tcpassembly/assemply.go
-LOCAL int32_t moloch_packet_sequence_diff (uint32_t a, uint32_t b)
+LOCAL int64_t moloch_packet_sequence_diff (int64_t a, int64_t b)
 {
     if (a > 0xc0000000 && b < 0x40000000)
-        return (a + 0xffffffffLL - b);
+        return a + 0x100000000LL - b;
 
     if (b > 0xc0000000 && a < 0x40000000)
-        return (a - b - 0xffffffffLL);
+        return a - b - 0x100000000LL;
 
     return b - a;
 }
@@ -117,14 +163,11 @@ LOCAL int32_t moloch_packet_sequence_diff (uint32_t a, uint32_t b)
 void moloch_packet_process_data(MolochSession_t *session, const uint8_t *data, int len, int which)
 {
     int i;
-    int totConsumed = 0;
-    int consumed = 0;
 
     for (i = 0; i < session->parserNum; i++) {
         if (session->parserInfo[i].parserFunc) {
-            consumed = session->parserInfo[i].parserFunc(session, session->parserInfo[i].uw, data, len, which);
+            int consumed = session->parserInfo[i].parserFunc(session, session->parserInfo[i].uw, data, len, which);
             if (consumed) {
-                totConsumed += consumed;
                 session->consumed[which] += consumed;
             }
 
@@ -134,7 +177,7 @@ void moloch_packet_process_data(MolochSession_t *session, const uint8_t *data, i
     }
 }
 /******************************************************************************/
-void moloch_packet_tcp_finish(MolochSession_t *session)
+LOCAL void moloch_packet_tcp_finish(MolochSession_t *session)
 {
     MolochTcpData_t            *ftd;
     MolochTcpData_t            *next;
@@ -142,7 +185,7 @@ void moloch_packet_tcp_finish(MolochSession_t *session)
     MolochTcpDataHead_t * const tcpData = &session->tcpData;
 
 #ifdef DEBUG_PACKET
-    LOG("START");
+    LOG("START %u %u", session->tcpSeq[0], session->tcpSeq[1]);
     DLL_FOREACH(td_, tcpData, ftd) {
         LOG("dir: %d seq: %8u ack: %8u len: %4u", ftd->packet->direction, ftd->seq, ftd->ack, ftd->len);
     }
@@ -152,7 +195,18 @@ void moloch_packet_tcp_finish(MolochSession_t *session)
         const int which = ftd->packet->direction;
         const uint32_t tcpSeq = session->tcpSeq[which];
 
-        if (tcpSeq >= ftd->seq && tcpSeq < (ftd->seq + ftd->len)) {
+        /* The sequence number we are looking for is past the start of the packet */
+        if (tcpSeq >= ftd->seq) {
+
+            /* The sequence number we are looking for is past the end of the packet, free it */
+            if (tcpSeq >= ftd->seq + ftd->len) {
+                DLL_REMOVE(td_, tcpData, ftd);
+                moloch_packet_free(ftd->packet);
+                MOLOCH_TYPE_FREE(MolochTcpData_t, ftd);
+                continue;
+            }
+
+            /* This packet has the sequence number we are looking for */
             const int offset = tcpSeq - ftd->seq;
             const uint8_t *data = ftd->packet->pkt + ftd->dataOffset + offset;
             const int len = ftd->len - offset;
@@ -172,7 +226,7 @@ void moloch_packet_tcp_finish(MolochSession_t *session)
             session->databytes[which] += len;
             session->totalDatabytes[which] += len;
 
-            if (config.yara) {
+            if (config.yara && config.yaraEveryPacket && !session->stopYara) {
                 moloch_yara_execute(session, data, len, 0);
             }
 
@@ -186,11 +240,17 @@ void moloch_packet_tcp_finish(MolochSession_t *session)
 }
 
 /******************************************************************************/
-void moloch_packet_process_icmp(MolochSession_t * const UNUSED(session), MolochPacket_t * const UNUSED(packet))
+LOCAL void moloch_packet_process_icmp(MolochSession_t * const UNUSED(session), MolochPacket_t * const packet)
 {
+    const uint8_t *data = packet->pkt + packet->payloadOffset;
+
+    if (packet->payloadLen >= 2) {
+        moloch_field_int_add(icmpTypeField, session, data[0]);
+        moloch_field_int_add(icmpCodeField, session, data[1]);
+    }
 }
 /******************************************************************************/
-void moloch_packet_process_udp(MolochSession_t * const session, MolochPacket_t * const packet)
+LOCAL void moloch_packet_process_udp(MolochSession_t * const session, MolochPacket_t * const packet)
 {
     const uint8_t *data = packet->pkt + packet->payloadOffset + 8;
     int            len = packet->payloadLen - 8;
@@ -202,8 +262,11 @@ void moloch_packet_process_udp(MolochSession_t * const session, MolochPacket_t *
         session->firstBytesLen[packet->direction] = MIN(8, len);
         memcpy(session->firstBytes[packet->direction], data, session->firstBytesLen[packet->direction]);
 
-        if (!session->stopSPI)
-            moloch_parsers_classify_udp(session, data, len, packet->direction);
+        moloch_parsers_classify_udp(session, data, len, packet->direction);
+
+        if (config.yara && config.yaraEveryPacket && !session->stopYara) {
+            moloch_yara_execute(session, data, len, 0);
+        }
     }
 
     int i;
@@ -214,21 +277,23 @@ void moloch_packet_process_udp(MolochSession_t * const session, MolochPacket_t *
     }
 }
 /******************************************************************************/
-int moloch_packet_process_tcp(MolochSession_t * const session, MolochPacket_t * const packet)
+SUPPRESS_ALIGNMENT
+LOCAL int moloch_packet_process_tcp(MolochSession_t * const session, MolochPacket_t * const packet)
 {
-    if (session->stopSPI || session->stopTCP)
-        return 1;
-
     struct tcphdr       *tcphdr = (struct tcphdr *)(packet->pkt + packet->payloadOffset);
 
 
     int            len = packet->payloadLen - 4*tcphdr->th_off;
 
+    const uint32_t seq = ntohl(tcphdr->th_seq);
+
 #ifdef DEBUG_PACKET
-    LOG("poffset: %d plen: %d len: %d", packet->payloadOffset, packet->payloadLen, len);
+    LOG("poffset: %d plen: %d len: %d seq: %u ack: %u", packet->payloadOffset, packet->payloadLen, len, seq, ntohl(tcphdr->th_ack));
 #endif
 
-    const uint32_t seq = ntohl(tcphdr->th_seq);
+    if (tcphdr->th_win == 0 && (tcphdr->th_flags & TH_RST) == 0) {
+        session->tcpFlagCnt[MOLOCH_TCPFLAG_SRC_ZERO + packet->direction]++;
+    }
 
     if (len < 0)
         return 1;
@@ -249,6 +314,11 @@ int moloch_packet_process_tcp(MolochSession_t * const session, MolochPacket_t * 
             }
         } else {
             session->tcpFlagCnt[MOLOCH_TCPFLAG_SYN]++;
+            if (session->synTime == 0) {
+                session->synTime = (packet->ts.tv_sec - session->firstPacket.tv_sec) * 1000000 +
+                                   (packet->ts.tv_usec - session->firstPacket.tv_usec) + 1;
+                session->ackTime = 0;
+            }
         }
 
         session->haveTcpSession = 1;
@@ -275,12 +345,18 @@ int moloch_packet_process_tcp(MolochSession_t * const session, MolochPacket_t * 
 
     if ((tcphdr->th_flags & (TH_FIN | TH_RST | TH_PUSH | TH_SYN | TH_ACK)) == TH_ACK) {
         session->tcpFlagCnt[MOLOCH_TCPFLAG_ACK]++;
+        if (session->ackTime == 0) {
+            session->ackTime = (packet->ts.tv_sec - session->firstPacket.tv_sec) * 1000000 +
+                               (packet->ts.tv_usec - session->firstPacket.tv_usec) + 1;
+        }
     }
 
     if (tcphdr->th_flags & TH_PUSH) {
         session->tcpFlagCnt[MOLOCH_TCPFLAG_PSH]++;
     }
 
+    if (session->stopTCP)
+        return 1;
 
     // If we've seen SYN but no SYN_ACK and no tcpSeq set, then just assume we've missed the syn-ack
     if (session->haveTcpSession && session->tcpFlagCnt[MOLOCH_TCPFLAG_SYN_ACK] == 0 && session->tcpSeq[packet->direction] == 0) {
@@ -290,7 +366,7 @@ int moloch_packet_process_tcp(MolochSession_t * const session, MolochPacket_t * 
 
     MolochTcpDataHead_t * const tcpData = &session->tcpData;
 
-    if (DLL_COUNT(td_, tcpData) > 256) {
+    if (DLL_COUNT(td_, tcpData) > maxTcpOutOfOrderPackets) {
         moloch_packet_tcp_free(session);
         moloch_session_add_tag(session, "incomplete-tcp");
         session->stopTCP = 1;
@@ -312,9 +388,12 @@ int moloch_packet_process_tcp(MolochSession_t * const session, MolochPacket_t * 
     }
 
     if (tcphdr->th_flags & TH_ACK) {
-        if (session->haveTcpSession && (session->ackedUnseenSegment & (1 << packet->direction)) == 0 &&
-            (moloch_packet_sequence_diff(session->tcpSeq[(packet->direction+1)%2], ntohl(tcphdr->th_ack)) > 1)) {
-                static char *tags[2] = {"acked-unseen-segment-src", "acked-unseen-segment-dst"};
+        if (session->haveTcpSession &&  // Seen a SYN
+            (session->ackedUnseenSegment & (1 << packet->direction)) == 0 &&  // Haven't already tagged
+            session->tcpSeq[(packet->direction+1)%2] != 0 &&                  // The syn-ack isn't what is missing
+            (moloch_packet_sequence_diff(session->tcpSeq[(packet->direction+1)%2], ntohl(tcphdr->th_ack)) > 1)) { // more then one byte missing
+
+                static const char *tags[2] = {"acked-unseen-segment-src", "acked-unseen-segment-dst"};
                 moloch_session_add_tag(session, tags[packet->direction]);
                 session->ackedUnseenSegment |= (1 << packet->direction);
         }
@@ -325,8 +404,8 @@ int moloch_packet_process_tcp(MolochSession_t * const session, MolochPacket_t * 
         return 1;
 
     // This packet is before what we are processing
-    int32_t diff = moloch_packet_sequence_diff(session->tcpSeq[packet->direction], seq + len);
-    if (diff <= 0)
+    int64_t diff = moloch_packet_sequence_diff(session->tcpSeq[packet->direction], seq + len);
+    if (session->haveTcpSession && diff <= 0)
         return 1;
 
     MolochTcpData_t *ftd, *td = MOLOCH_TYPE_ALLOC(MolochTcpData_t);
@@ -339,7 +418,7 @@ int moloch_packet_process_tcp(MolochSession_t * const session, MolochPacket_t * 
     td->dataOffset = packet->payloadOffset + 4*tcphdr->th_off;
 
 #ifdef DEBUG_PACKET
-    LOG("dir: %d seq: %u ack: %u len: %d diff0: %d", packet->direction, seq, ack, len, diff);
+    LOG("dir: %d seq: %u ack: %u len: %d diff0: %lld", packet->direction, seq, ack, len, diff);
 #endif
 
     if (DLL_COUNT(td_, tcpData) == 0) {
@@ -385,7 +464,7 @@ int moloch_packet_process_tcp(MolochSession_t * const session, MolochPacket_t * 
         }
 
         if (session->haveTcpSession && (session->outOfOrder & (1 << packet->direction)) == 0) {
-            static char *tags[2] = {"out-of-order-src", "out-of-order-dst"};
+            static const char *tags[2] = {"out-of-order-src", "out-of-order-dst"};
             moloch_session_add_tag(session, tags[packet->direction]);
             session->outOfOrder |= (1 << packet->direction);
         }
@@ -423,230 +502,260 @@ void moloch_packet_flush()
     }
 }
 /******************************************************************************/
-LOCAL void *moloch_packet_thread(void *threadp)
+SUPPRESS_ALIGNMENT
+LOCAL void moloch_packet_process(MolochPacket_t *packet, int thread)
 {
-    MolochPacket_t  *packet;
-    int thread = (long)threadp;
-
-    while (1) {
-        MOLOCH_LOCK(packetQ[thread].lock);
-        inProgress[thread] = 0;
-        if (DLL_COUNT(packet_, &packetQ[thread]) == 0) {
-            struct timeval tv;
-            struct timespec ts;
-            gettimeofday(&tv, NULL);
-            ts.tv_sec = tv.tv_sec + 1;
-            ts.tv_nsec = 0;
-            MOLOCH_COND_TIMEDWAIT(packetQ[thread].lock, ts);
-        }
-        inProgress[thread] = 1;
-        DLL_POP_HEAD(packet_, &packetQ[thread], packet);
-        MOLOCH_UNLOCK(packetQ[thread].lock);
-
-        moloch_session_process_commands(thread);
-
-        if (!packet)
-            continue;
 #ifdef DEBUG_PACKET
-        LOG("Processing %p %d", packet, packet->pktlen);
+    LOG("Processing %p %d", packet, packet->pktlen);
 #endif
 
-        lastPacketSecs[thread] = packet->ts.tv_sec;
+    lastPacketSecs[thread] = packet->ts.tv_sec;
 
-        MolochSession_t     *session;
-        struct ip           *ip4 = (struct ip*)(packet->pkt + packet->ipOffset);
-        struct ip6_hdr      *ip6 = (struct ip6_hdr*)(packet->pkt + packet->ipOffset);
-        struct tcphdr       *tcphdr = 0;
-        struct udphdr       *udphdr = 0;
-        char                 sessionId[MOLOCH_SESSIONID_LEN];
+    MolochSession_t     *session;
+    struct ip           *ip4 = (struct ip*)(packet->pkt + packet->ipOffset);
+    struct ip6_hdr      *ip6 = (struct ip6_hdr*)(packet->pkt + packet->ipOffset);
+    struct tcphdr       *tcphdr = 0;
+    struct udphdr       *udphdr = 0;
+    char                 sessionId[MOLOCH_SESSIONID_LEN];
 
-        switch (packet->protocol) {
-        case IPPROTO_TCP:
-            tcphdr = (struct tcphdr *)(packet->pkt + packet->payloadOffset);
+    switch (packet->protocol) {
+    case IPPROTO_TCP:
+        tcphdr = (struct tcphdr *)(packet->pkt + packet->payloadOffset);
 
-            if (packet->v6) {
-                moloch_session_id6(sessionId, ip6->ip6_src.s6_addr, tcphdr->th_sport,
-                                   ip6->ip6_dst.s6_addr, tcphdr->th_dport);
-            } else {
-                moloch_session_id(sessionId, ip4->ip_src.s_addr, tcphdr->th_sport,
-                                  ip4->ip_dst.s_addr, tcphdr->th_dport);
-            }
-            break;
-        case IPPROTO_UDP:
-            udphdr = (struct udphdr *)(packet->pkt + packet->payloadOffset);
-            if (packet->v6) {
-                moloch_session_id6(sessionId, ip6->ip6_src.s6_addr, udphdr->uh_sport,
-                                   ip6->ip6_dst.s6_addr, udphdr->uh_dport);
-            } else {
-                moloch_session_id(sessionId, ip4->ip_src.s_addr, udphdr->uh_sport,
-                                  ip4->ip_dst.s_addr, udphdr->uh_dport);
-            }
-            break;
-            break;
-        case IPPROTO_ICMP:
-            if (packet->v6) {
-                moloch_session_id6(sessionId, ip6->ip6_src.s6_addr, 0,
-                                   ip6->ip6_dst.s6_addr, 0);
-            } else {
-                moloch_session_id(sessionId, ip4->ip_src.s_addr, 0,
-                                  ip4->ip_dst.s_addr, 0);
-            }
-            break;
-        case IPPROTO_ICMPV6:
+        if (packet->v6) {
+            moloch_session_id6(sessionId, ip6->ip6_src.s6_addr, tcphdr->th_sport,
+                               ip6->ip6_dst.s6_addr, tcphdr->th_dport);
+        } else {
+            moloch_session_id(sessionId, ip4->ip_src.s_addr, tcphdr->th_sport,
+                              ip4->ip_dst.s_addr, tcphdr->th_dport);
+        }
+        break;
+    case IPPROTO_UDP:
+        udphdr = (struct udphdr *)(packet->pkt + packet->payloadOffset);
+        if (packet->v6) {
+            moloch_session_id6(sessionId, ip6->ip6_src.s6_addr, udphdr->uh_sport,
+                               ip6->ip6_dst.s6_addr, udphdr->uh_dport);
+        } else {
+            moloch_session_id(sessionId, ip4->ip_src.s_addr, udphdr->uh_sport,
+                              ip4->ip_dst.s_addr, udphdr->uh_dport);
+        }
+        break;
+    case IPPROTO_ESP:
+    case IPPROTO_ICMP:
+        if (packet->v6) {
             moloch_session_id6(sessionId, ip6->ip6_src.s6_addr, 0,
                                ip6->ip6_dst.s6_addr, 0);
-            break;
-        }
-
-        int isNew;
-        session = moloch_session_find_or_create(packet->ses, packet->hash, sessionId, &isNew); // Returns locked session
-
-        if (isNew) {
-            session->saveTime = packet->ts.tv_sec + config.tcpSaveTimeout;
-            session->firstPacket = packet->ts;
-
-            session->protocol = packet->protocol;
-            if (ip4->ip_v == 4) {
-                ((uint32_t *)session->addr1.s6_addr)[2] = htonl(0xffff);
-                ((uint32_t *)session->addr1.s6_addr)[3] = ip4->ip_src.s_addr;
-                ((uint32_t *)session->addr2.s6_addr)[2] = htonl(0xffff);
-                ((uint32_t *)session->addr2.s6_addr)[3] = ip4->ip_dst.s_addr;
-                session->ip_tos = ip4->ip_tos;
-            } else {
-                session->addr1 = ip6->ip6_src;
-                session->addr2 = ip6->ip6_dst;
-                session->ip_tos = 0;
-            }
-            session->thread = thread;
-
-            moloch_parsers_initial_tag(session);
-
-            switch (session->protocol) {
-            case IPPROTO_TCP:
-               /* If antiSynDrop option is set to true, capture will assume that
-                * if the syn-ack ip4 was captured first then the syn probably got dropped.*/
-                if ((tcphdr->th_flags & TH_SYN) && (tcphdr->th_flags & TH_ACK) && (config.antiSynDrop)) {
-                    struct in6_addr tmp;
-                    tmp = session->addr1;
-                    session->addr1 = session->addr2;
-                    session->addr2 = tmp;
-                    session->port1 = ntohs(tcphdr->th_dport);
-                    session->port2 = ntohs(tcphdr->th_sport);
-                } else {
-                    session->port1 = ntohs(tcphdr->th_sport);
-                    session->port2 = ntohs(tcphdr->th_dport);
-                }
-                if (moloch_http_is_moloch(session->h_hash, sessionId)) {
-                    if (config.debug) {
-                        char buf[1000];
-                        LOG("Ignoring connection %s", moloch_session_id_string(session->sessionId, buf));
-                    }
-                    session->stopSPI = 1;
-                    session->stopSaving = 1;
-                }
-                break;
-            case IPPROTO_UDP:
-                session->port1 = ntohs(udphdr->uh_sport);
-                session->port2 = ntohs(udphdr->uh_dport);
-                break;
-            case IPPROTO_ICMP:
-                break;
-            }
-
-            if (pluginsCbs & MOLOCH_PLUGIN_NEW)
-                moloch_plugins_cb_new(session);
-        }
-
-        int dir;
-        if (ip4->ip_v == 4) {
-            dir = (MOLOCH_V6_TO_V4(session->addr1) == ip4->ip_src.s_addr &&
-                   MOLOCH_V6_TO_V4(session->addr2) == ip4->ip_dst.s_addr);
         } else {
-            dir = (memcmp(session->addr1.s6_addr, ip6->ip6_src.s6_addr, 16) == 0 &&
-                   memcmp(session->addr2.s6_addr, ip6->ip6_dst.s6_addr, 16) == 0);
+            moloch_session_id(sessionId, ip4->ip_src.s_addr, 0,
+                              ip4->ip_dst.s_addr, 0);
         }
+        break;
+    case IPPROTO_SCTP:
+        udphdr = (struct udphdr *)(packet->pkt + packet->payloadOffset); /* Not really udp, but port in same location */
+        if (packet->v6) {
+            moloch_session_id6(sessionId, ip6->ip6_src.s6_addr, udphdr->uh_sport,
+                               ip6->ip6_dst.s6_addr, udphdr->uh_dport);
+        } else {
+            moloch_session_id(sessionId, ip4->ip_src.s_addr, udphdr->uh_sport,
+                              ip4->ip_dst.s_addr, udphdr->uh_dport);
+        }
+        break;
+    case IPPROTO_ICMPV6:
+        moloch_session_id6(sessionId, ip6->ip6_src.s6_addr, 0,
+                           ip6->ip6_dst.s6_addr, 0);
+        break;
+    }
 
-        packet->direction = 0;
+    int isNew;
+    session = moloch_session_find_or_create(packet->ses, packet->hash, sessionId, &isNew); // Returns locked session
+
+    if (isNew) {
+        session->saveTime = packet->ts.tv_sec + config.tcpSaveTimeout;
+        session->firstPacket = packet->ts;
+
+        session->protocol = packet->protocol;
+        if (ip4->ip_v == 4) {
+            ((uint32_t *)session->addr1.s6_addr)[2] = htonl(0xffff);
+            ((uint32_t *)session->addr1.s6_addr)[3] = ip4->ip_src.s_addr;
+            ((uint32_t *)session->addr2.s6_addr)[2] = htonl(0xffff);
+            ((uint32_t *)session->addr2.s6_addr)[3] = ip4->ip_dst.s_addr;
+            session->ip_tos = ip4->ip_tos;
+        } else {
+            session->addr1 = ip6->ip6_src;
+            session->addr2 = ip6->ip6_dst;
+            session->ip_tos = 0;
+        }
+        session->thread = thread;
+
+        moloch_parsers_initial_tag(session);
+
+        if (readerFieldOps[packet->readerPos].num)
+            moloch_field_ops_run(session, &readerFieldOps[packet->readerPos]);
+
         switch (session->protocol) {
-        case IPPROTO_UDP:
-            udphdr = (struct udphdr *)(packet->pkt + packet->payloadOffset);
-            packet->direction = (dir &&
-                                 session->port1 == ntohs(udphdr->uh_sport) &&
-                                 session->port2 == ntohs(udphdr->uh_dport))?0:1;
-            session->databytes[packet->direction] += (packet->pktlen - 8);
-            break;
         case IPPROTO_TCP:
-            tcphdr = (struct tcphdr *)(packet->pkt + packet->payloadOffset);
-            packet->direction = (dir &&
-                                 session->port1 == ntohs(tcphdr->th_sport) &&
-                                 session->port2 == ntohs(tcphdr->th_dport))?0:1;
-            session->tcp_flags |= tcphdr->th_flags;
+           /* If antiSynDrop option is set to true, capture will assume that
+            * if the syn-ack ip4 was captured first then the syn probably got dropped.*/
+            if ((tcphdr->th_flags & TH_SYN) && (tcphdr->th_flags & TH_ACK) && (config.antiSynDrop)) {
+                struct in6_addr tmp;
+                tmp = session->addr1;
+                session->addr1 = session->addr2;
+                session->addr2 = tmp;
+                session->port1 = ntohs(tcphdr->th_dport);
+                session->port2 = ntohs(tcphdr->th_sport);
+            } else {
+                session->port1 = ntohs(tcphdr->th_sport);
+                session->port2 = ntohs(tcphdr->th_dport);
+            }
+            if (moloch_http_is_moloch(session->h_hash, sessionId)) {
+                if (config.debug) {
+                    char buf[1000];
+                    LOG("Ignoring connection %s", moloch_session_id_string(session->sessionId, buf));
+                }
+                session->stopSPI = 1;
+                moloch_packet_free(packet);
+                return;
+            }
+            break;
+        case IPPROTO_UDP:
+            session->port1 = ntohs(udphdr->uh_sport);
+            session->port2 = ntohs(udphdr->uh_dport);
+            break;
+        case IPPROTO_SCTP:
+            session->port1 = ntohs(udphdr->uh_sport);
+            session->port2 = ntohs(udphdr->uh_dport);
+            break;
+        case IPPROTO_ESP:
+            session->stopSaving = 1;
             break;
         case IPPROTO_ICMP:
-            packet->direction = (dir)?0:1;
             break;
         }
 
-        /* Check if the stop saving bpf filters match */
-        if (session->packets[packet->direction] == 0 && session->stopSaving == 0) {
-            moloch_rules_run_session_setup(session, packet);
-        }
+        if (pluginsCbs & MOLOCH_PLUGIN_NEW)
+            moloch_plugins_cb_new(session);
+    } else if (session->stopSPI) {
+        moloch_packet_free(packet);
+        return;
+    }
 
-        session->packets[packet->direction]++;
-        session->bytes[packet->direction] += packet->pktlen;
-        session->lastPacket = packet->ts;
+    int dir;
+    if (ip4->ip_v == 4) {
+        dir = (MOLOCH_V6_TO_V4(session->addr1) == ip4->ip_src.s_addr &&
+               MOLOCH_V6_TO_V4(session->addr2) == ip4->ip_dst.s_addr);
+    } else {
+        dir = (memcmp(session->addr1.s6_addr, ip6->ip6_src.s6_addr, 16) == 0 &&
+               memcmp(session->addr2.s6_addr, ip6->ip6_dst.s6_addr, 16) == 0);
+    }
 
-        uint32_t packets = session->packets[0] + session->packets[1];
+    packet->direction = 0;
+    switch (session->protocol) {
+    case IPPROTO_UDP:
+        udphdr = (struct udphdr *)(packet->pkt + packet->payloadOffset);
+        packet->direction = (dir &&
+                             session->port1 == ntohs(udphdr->uh_sport) &&
+                             session->port2 == ntohs(udphdr->uh_dport))?0:1;
+        session->databytes[packet->direction] += (packet->pktlen -packet->payloadOffset - 8);
+        break;
+    case IPPROTO_SCTP:
+        udphdr = (struct udphdr *)(packet->pkt + packet->payloadOffset);
+        packet->direction = (dir &&
+                             session->port1 == ntohs(udphdr->uh_sport) &&
+                             session->port2 == ntohs(udphdr->uh_dport))?0:1;
+        session->databytes[packet->direction] += (packet->pktlen - 8);
+        break;
+    case IPPROTO_TCP:
+        tcphdr = (struct tcphdr *)(packet->pkt + packet->payloadOffset);
+        packet->direction = (dir &&
+                             session->port1 == ntohs(tcphdr->th_sport) &&
+                             session->port2 == ntohs(tcphdr->th_dport))?0:1;
+        session->tcp_flags |= tcphdr->th_flags;
+        break;
+    case IPPROTO_ICMP:
+    case IPPROTO_ICMPV6:
+        packet->direction = (dir)?0:1;
+        session->databytes[packet->direction] += (packet->pktlen -packet->payloadOffset);
+        break;
+    case IPPROTO_ESP:
+        packet->direction = (dir)?0:1;
+        break;
+    }
 
-        if (session->stopSaving == 0 || packets < session->stopSaving) {
-            moloch_writer_write(session, packet);
+    if (isNew) {
+        moloch_rules_session_create(session);
+    }
 
+    /* Check if the stop saving bpf filters match */
+    if (session->packets[packet->direction] == 0 && session->stopSaving == 0xffff) {
+        moloch_rules_run_session_setup(session, packet);
+    }
+
+    session->packets[packet->direction]++;
+    session->bytes[packet->direction] += packet->pktlen;
+    session->lastPacket = packet->ts;
+
+    uint32_t packets = session->packets[0] + session->packets[1];
+
+    if (packets <= session->stopSaving) {
+        moloch_writer_write(session, packet);
+
+        // If writerFilePos is 0, then the writer couldn't save the packet
+        if (packet->writerFilePos == 0) {
+            if (!session->diskOverload) {
+                moloch_session_add_tag(session, "pcap-disk-overload");
+                session->diskOverload = 1;
+            }
+            MOLOCH_THREAD_INCR_NUM(unwrittenBytes, packet->pktlen);
+        } else {
+
+            MOLOCH_THREAD_INCR_NUM(writtenBytes, packet->pktlen);
+
+            // If the last fileNum used in the session isn't the same as the
+            // lastest packets fileNum then we need to add to the filePos and
+            // fileNum arrays.
             int16_t len;
             if (session->lastFileNum != packet->writerFileNum) {
                 session->lastFileNum = packet->writerFileNum;
                 g_array_append_val(session->fileNumArray, packet->writerFileNum);
                 int64_t pos = -1LL * packet->writerFileNum;
                 g_array_append_val(session->filePosArray, pos);
-                len = 0;
-                g_array_append_val(session->fileLenArray, len);
+
+                if (config.enablePacketLen) {
+                    len = 0;
+                    g_array_append_val(session->fileLenArray, len);
+                }
             }
 
             g_array_append_val(session->filePosArray, packet->writerFilePos);
-            len = 16 + packet->pktlen;
-            g_array_append_val(session->fileLenArray, len);
 
-            if (packets >= config.maxPackets || session->midSave) {
-                moloch_session_mid_save(session, packet->ts.tv_sec);
+            if (config.enablePacketLen) {
+                len = 16 + packet->pktlen;
+                g_array_append_val(session->fileLenArray, len);
             }
         }
 
-        if (pcapFileHeader.linktype == 1 && session->firstBytesLen[packet->direction] < 8 && session->packets[packet->direction] < 10) {
-            const uint8_t *pcapData = packet->pkt;
-            char str1[20];
-            char str2[20];
-            snprintf(str1, sizeof(str1), "%02x:%02x:%02x:%02x:%02x:%02x",
-                    pcapData[0],
-                    pcapData[1],
-                    pcapData[2],
-                    pcapData[3],
-                    pcapData[4],
-                    pcapData[5]);
+        if (packets >= config.maxPackets || session->midSave) {
+            moloch_session_mid_save(session, packet->ts.tv_sec);
+        }
+    } else {
+        // If we hit stopSaving for this session and try and save 1 more packet then
+        // add truncated-pcap tag to the session
+        if (packets - 1 == session->stopSaving) {
+            moloch_session_add_tag(session, "truncated-pcap");
+        }
+        MOLOCH_THREAD_INCR_NUM(unwrittenBytes, packet->pktlen);
+    }
 
+    if (session->firstBytesLen[packet->direction] < 8 && session->packets[packet->direction] < 10) {
+        const uint8_t *pcapData = packet->pkt;
 
-            snprintf(str2, sizeof(str2), "%02x:%02x:%02x:%02x:%02x:%02x",
-                    pcapData[6],
-                    pcapData[7],
-                    pcapData[8],
-                    pcapData[9],
-                    pcapData[10],
-                    pcapData[11]);
-
+        if (pcapFileHeader.linktype == 1) {
             if (packet->direction == 1) {
-                moloch_field_string_add(mac1Field, session, str1, 17, TRUE);
-                moloch_field_string_add(mac2Field, session, str2, 17, TRUE);
+                moloch_field_macoui_add(session, mac1Field, oui1Field, pcapData+0);
+                moloch_field_macoui_add(session, mac2Field, oui2Field, pcapData+6);
             } else {
-                moloch_field_string_add(mac1Field, session, str2, 17, TRUE);
-                moloch_field_string_add(mac2Field, session, str1, 17, TRUE);
+                moloch_field_macoui_add(session, mac1Field, oui1Field, pcapData+6);
+                moloch_field_macoui_add(session, mac2Field, oui2Field, pcapData+0);
             }
 
             int n = 12;
@@ -656,68 +765,185 @@ LOCAL void *moloch_packet_thread(void *threadp)
                 n += 4;
             }
 
-            switch(packet->vpnType) {
-            case MOLOCH_PACKET_VPNTYPE_GRE:
-                ip4 = (struct ip*)(packet->pkt + packet->vpnIpOffset);
-                moloch_field_int_add(greIpField, session, ip4->ip_src.s_addr);
-                moloch_field_int_add(greIpField, session, ip4->ip_dst.s_addr);
-                moloch_session_add_protocol(session, "gre");
-                break;
-            case MOLOCH_PACKET_VPNTYPE_PPPOE:
-                moloch_session_add_protocol(session, "pppoe");
-                break;
+        }
+
+        if (packet->vlan)
+            moloch_field_int_add(vlanField, session, packet->vlan);
+
+        if (packet->tunnel & MOLOCH_PACKET_TUNNEL_GRE) {
+            ip4 = (struct ip*)(packet->pkt + packet->vpnIpOffset);
+            moloch_field_ip4_add(greIpField, session, ip4->ip_src.s_addr);
+            moloch_field_ip4_add(greIpField, session, ip4->ip_dst.s_addr);
+            moloch_session_add_protocol(session, "gre");
+        }
+
+        if (packet->tunnel & MOLOCH_PACKET_TUNNEL_PPPOE) {
+            moloch_session_add_protocol(session, "pppoe");
+        }
+
+        if (packet->tunnel & MOLOCH_PACKET_TUNNEL_PPP) {
+            moloch_session_add_protocol(session, "ppp");
+        }
+
+        if (packet->tunnel & MOLOCH_PACKET_TUNNEL_MPLS) {
+            moloch_session_add_protocol(session, "mpls");
+        }
+
+        if (packet->tunnel & MOLOCH_PACKET_TUNNEL_GTP) {
+            moloch_session_add_protocol(session, "gtp");
+        }
+
+        if (packet->tunnel & MOLOCH_PACKET_TUNNEL_VXLAN) {
+            moloch_session_add_protocol(session, "vxlan");
+        }
+    }
+
+
+    int freePacket = 1;
+    switch(packet->ses) {
+    case SESSION_ICMP:
+        moloch_packet_process_icmp(session, packet);
+        break;
+    case SESSION_UDP:
+        moloch_packet_process_udp(session, packet);
+        break;
+    case SESSION_TCP:
+        freePacket = moloch_packet_process_tcp(session, packet);
+        moloch_packet_tcp_finish(session);
+        break;
+    }
+
+    if (freePacket) {
+        moloch_packet_free(packet);
+    }
+}
+/******************************************************************************/
+#ifndef FUZZLOCH
+LOCAL void *moloch_packet_thread(void *threadp)
+{
+    int thread = (long)threadp;
+    const uint32_t maxPackets75 = config.maxPackets*0.75;
+    uint32_t skipCount = 0;
+
+    while (1) {
+        MolochPacket_t  *packet;
+
+        MOLOCH_LOCK(packetQ[thread].lock);
+        inProgress[thread] = 0;
+        if (DLL_COUNT(packet_, &packetQ[thread]) == 0) {
+            struct timespec ts;
+            clock_gettime(CLOCK_REALTIME_COARSE, &ts);
+            ts.tv_sec++;
+            MOLOCH_COND_TIMEDWAIT(packetQ[thread].lock, ts);
+
+            /* If we are in live capture mode and we haven't received any packets for 10 seconds we set current time to 10
+             * seconds in the past so moloch_session_process_commands will clean things up.  10 seconds is arbitrary but
+             * we want to make sure we don't set the time ahead of any packets that are currently being read off the wire
+             */
+            if (!config.pcapReadOffline && DLL_COUNT(packet_, &packetQ[thread]) == 0 && ts.tv_sec - 10 > lastPacketSecs[thread]) {
+                lastPacketSecs[thread] = ts.tv_sec - 10;
             }
-
         }
+        inProgress[thread] = 1;
+        DLL_POP_HEAD(packet_, &packetQ[thread], packet);
+        MOLOCH_UNLOCK(packetQ[thread].lock);
 
-
-        int freePacket = 1;
-        switch(packet->ses) {
-        case SESSION_ICMP:
-            moloch_packet_process_icmp(session, packet);
-            break;
-        case SESSION_UDP:
-            moloch_packet_process_udp(session, packet);
-            break;
-        case SESSION_TCP:
-            freePacket = moloch_packet_process_tcp(session, packet);
-            moloch_packet_tcp_finish(session);
-            break;
+        // Only process commands if the packetQ is less then 75% full or every 8 packets
+        if (likely(DLL_COUNT(packet_, &packetQ[thread]) < maxPackets75) || (skipCount & 0x7) == 0) {
+            moloch_session_process_commands(thread);
+            if (!packet)
+                continue;
+        } else {
+            skipCount++;
         }
-
-        if (freePacket) {
-            moloch_packet_free(packet);
-        }
+        moloch_packet_process(packet, thread);
     }
 
     return NULL;
 }
+#endif
+/******************************************************************************/
+static FILE *unknownPacketFile[3];
+LOCAL void moloch_packet_save_unknown_packet(int type, MolochPacket_t * const packet)
+{
+    static MOLOCH_LOCK_DEFINE(lock);
+
+    struct moloch_pcap_sf_pkthdr hdr;
+    hdr.ts.tv_sec  = packet->ts.tv_sec;
+    hdr.ts.tv_usec = packet->ts.tv_usec;
+    hdr.caplen     = packet->pktlen;
+    hdr.pktlen     = packet->pktlen;
+
+    MOLOCH_LOCK(lock);
+    if (!unknownPacketFile[type]) {
+        char               str[PATH_MAX];
+        static const char *names[] = {"unknown.ether", "unknown.ip", "corrupt"};
+
+        snprintf(str, sizeof(str), "%s/%s.%d.pcap", config.pcapDir[0], names[type], getpid());
+        unknownPacketFile[type] = fopen(str, "w");
+
+	// TODO-- should we also add logic to pick right pcapDir when there are multiple?
+        if (unknownPacketFile[type] == NULL) {
+          LOGEXIT("Unable to open pcap file %s to store unknown type %s.  Error %s", str, names[type], strerror (errno));
+          MOLOCH_UNLOCK(lock);
+          return;
+        }
+
+        fwrite(&pcapFileHeader, 24, 1, unknownPacketFile[type]);
+    }
+
+    fwrite(&hdr, 16, 1, unknownPacketFile[type]);
+    fwrite(packet->pkt, packet->pktlen, 1, unknownPacketFile[type]);
+    MOLOCH_UNLOCK(lock);
+}
 
 /******************************************************************************/
-int moloch_packet_gre4(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+LOCAL int moloch_packet_erspan(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+{
+    if (unlikely(len) < 8 || unlikely(!data))
+        return MOLOCH_PACKET_CORRUPT;
+
+    if ((*data >> 4) == 1)
+        return moloch_packet_ether(batch, packet, data + 8, len - 8);
+
+
+    if (config.logUnknownProtocols)
+        LOG("Unknown ERSPAN protocol %d", *data >> 4);
+    if (BIT_ISSET(0x88be, config.etherSavePcap))
+        moloch_packet_save_unknown_packet(0, packet);
+    return MOLOCH_PACKET_UNKNOWN;
+}
+/******************************************************************************/
+LOCAL int moloch_packet_gre4(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
 {
     BSB bsb;
-
-    if (len < 4)
-        return 1;
+    if (unlikely(len) < 4 || unlikely(!data))
+        return MOLOCH_PACKET_CORRUPT;
 
     BSB_INIT(bsb, data, len);
+
     uint16_t flags_version = 0;
     BSB_IMPORT_u16(bsb, flags_version);
     uint16_t type = 0;
     BSB_IMPORT_u16(bsb, type);
 
-    if (type != 0x0800) {
+    switch (type) {
+    case 0x0800:
+    case 0x86dd:
+    case 0x6559:
+    case 0x880b:
+    case 0x88be:
+        break;
+    default:
         if (config.logUnknownProtocols)
             LOG("Unknown GRE protocol 0x%04x(%d)", type, type);
-        return 1;
+        if (BIT_ISSET(type, config.etherSavePcap))
+            moloch_packet_save_unknown_packet(0, packet);
+        return MOLOCH_PACKET_UNKNOWN;
     }
 
-    uint16_t offset = 0;
-
     if (flags_version & (0x8000 | 0x4000)) {
-        BSB_IMPORT_skip(bsb, 2);
-        BSB_IMPORT_u16(bsb, offset);
+        BSB_IMPORT_skip(bsb, 4); // skip len and offset
     }
 
     // key
@@ -742,10 +968,30 @@ int moloch_packet_gre4(MolochPacketBatch_t * batch, MolochPacket_t * const packe
         }
     }
 
-    if (BSB_IS_ERROR(bsb))
-        return 1;
+    // ack number
+    if (flags_version & 0x0080) {
+        BSB_IMPORT_skip(bsb, 4);
+    }
 
-    return moloch_packet_ip4(batch, packet, BSB_WORK_PTR(bsb), BSB_REMAINING(bsb));
+    if (BSB_IS_ERROR(bsb))
+        return MOLOCH_PACKET_CORRUPT;
+
+    switch (type) {
+    case 0x0800:
+        return moloch_packet_ip4(batch, packet, BSB_WORK_PTR(bsb), BSB_REMAINING(bsb));
+    case 0x86dd:
+        return moloch_packet_ip6(batch, packet, BSB_WORK_PTR(bsb), BSB_REMAINING(bsb));
+    case 0x6559:
+        return moloch_packet_frame_relay(batch, packet, BSB_WORK_PTR(bsb), BSB_REMAINING(bsb));
+    case 0x880b:
+        return moloch_packet_ppp(batch, packet, BSB_WORK_PTR(bsb), BSB_REMAINING(bsb));
+    case 0x88be:
+        return moloch_packet_erspan(batch, packet, BSB_WORK_PTR(bsb), BSB_REMAINING(bsb));
+    default:
+        if (BIT_ISSET(type, config.etherSavePcap))
+            moloch_packet_save_unknown_packet(0, packet);
+        return MOLOCH_PACKET_UNKNOWN;
+    }
 }
 /******************************************************************************/
 void moloch_packet_frags_free(MolochFrags_t * const frags)
@@ -760,7 +1006,8 @@ void moloch_packet_frags_free(MolochFrags_t * const frags)
     MOLOCH_TYPE_FREE(MolochFrags_t, frags);
 }
 /******************************************************************************/
-gboolean moloch_packet_frags_process(MolochPacket_t * const packet)
+SUPPRESS_ALIGNMENT
+LOCAL gboolean moloch_packet_frags_process(MolochPacket_t * const packet)
 {
     MolochPacket_t * fpacket;
     MolochFrags_t   *frags;
@@ -821,12 +1068,11 @@ gboolean moloch_packet_frags_process(MolochPacket_t * const packet)
 
     int off = 0;
     struct ip *fip4;
-    uint16_t fip_off;
 
     int payloadLen = 0;
     DLL_FOREACH(packet_, &frags->packets, fpacket) {
         fip4 = (struct ip*)(fpacket->pkt + fpacket->ipOffset);
-        fip_off = ntohs(fip4->ip_off) & IP_OFFMASK;
+        uint16_t fip_off = ntohs(fip4->ip_off) & IP_OFFMASK;
         if (fip_off != off)
             break;
         off += fpacket->payloadLen/8;
@@ -858,7 +1104,7 @@ gboolean moloch_packet_frags_process(MolochPacket_t * const packet)
 
     // Copy payload
     DLL_FOREACH(packet_, &frags->packets, fpacket) {
-        struct ip *fip4 = (struct ip*)(fpacket->pkt + fpacket->ipOffset);
+        fip4 = (struct ip*)(fpacket->pkt + fpacket->ipOffset);
         uint16_t fip_off = ntohs(fip4->ip_off) & IP_OFFMASK;
 
         if (packet->payloadOffset+(fip_off*8) + fpacket->payloadLen <= packet->pktlen)
@@ -879,15 +1125,17 @@ gboolean moloch_packet_frags_process(MolochPacket_t * const packet)
     return TRUE;
 }
 /******************************************************************************/
-void moloch_packet_frags4(MolochPacketBatch_t *batch, MolochPacket_t * const packet)
+LOCAL void moloch_packet_frags4(MolochPacketBatch_t *batch, MolochPacket_t * const packet)
 {
     MolochFrags_t *frags;
 
     // ALW - Should change frags_process to make the copy when needed
-    uint8_t *pkt = malloc(packet->pktlen);
-    memcpy(pkt, packet->pkt, packet->pktlen);
-    packet->pkt = pkt;
-    packet->copied = 1;
+    if (!packet->copied) {
+        uint8_t *pkt = malloc(packet->pktlen);
+        memcpy(pkt, packet->pkt, packet->pktlen);
+        packet->pkt = pkt;
+        packet->copied = 1;
+    }
 
     MOLOCH_LOCK(frags);
     // Remove expired entries
@@ -913,58 +1161,74 @@ int moloch_packet_frags_outstanding()
     return 0;
 }
 /******************************************************************************/
-int moloch_packet_ip(MolochPacketBatch_t *batch, MolochPacket_t * const packet, const char * const sessionId)
+LOCAL void moloch_packet_log(int ses)
 {
-    totalBytes += packet->pktlen;
+    MolochReaderStats_t stats;
+    if (moloch_reader_stats(&stats)) {
+        stats.dropped = 0;
+        stats.total = totalPackets;
+    }
 
+    LOG("packets: %" PRIu64 " current sessions: %u/%u oldest: %d - recv: %" PRIu64 " drop: %" PRIu64 " (%0.2f) queue: %d disk: %d packet: %d close: %d ns: %d frags: %d/%d pstats: %" PRIu64 "/%" PRIu64 "/%" PRIu64 "/%" PRIu64 "/%" PRIu64 "/%" PRIu64,
+      totalPackets,
+      moloch_session_watch_count(ses),
+      moloch_session_monitoring(),
+      moloch_session_idle_seconds(ses),
+      stats.total,
+      stats.dropped - initialDropped,
+      (stats.total?(stats.dropped - initialDropped)*(double)100.0/stats.total:0),
+      moloch_http_queue_length(esServer),
+      moloch_writer_queue_length(),
+      moloch_packet_outstanding(),
+      moloch_session_close_outstanding(),
+      moloch_session_need_save_outstanding(),
+      moloch_packet_frags_outstanding(),
+      moloch_packet_frags_size(),
+      packetStats[MOLOCH_PACKET_SUCCESS],
+      packetStats[MOLOCH_PACKET_IP_DROPPED],
+      packetStats[MOLOCH_PACKET_OVERLOAD_DROPPED],
+      packetStats[MOLOCH_PACKET_CORRUPT],
+      packetStats[MOLOCH_PACKET_UNKNOWN],
+      packetStats[MOLOCH_PACKET_IPPORT_DROPPED]
+      );
+
+      if (config.debug > 0) {
+          moloch_rules_stats();
+      }
+}
+/******************************************************************************/
+LOCAL int moloch_packet_ip(MolochPacketBatch_t *batch, MolochPacket_t * const packet, const char * const sessionId)
+{
     if (totalPackets == 0) {
         MolochReaderStats_t stats;
         if (!moloch_reader_stats(&stats)) {
             initialDropped = stats.dropped;
         }
         initialPacket = packet->ts;
-        LOG("Initial Packet = %ld", initialPacket.tv_sec);
-        LOG("%" PRIu64 " Initial Dropped = %d", totalPackets, initialDropped);
+        if (!config.pcapReadOffline)
+            LOG("Initial Packet = %ld Initial Dropped = %u", initialPacket.tv_sec, initialDropped);
     }
 
-    if ((++totalPackets) % config.logEveryXPackets == 0) {
-        MolochReaderStats_t stats;
-        if (moloch_reader_stats(&stats)) {
-            stats.dropped = 0;
-            stats.total = totalPackets;
-        }
-
-        LOG("packets: %" PRIu64 " current sessions: %u/%u oldest: %d - recv: %" PRIu64 " drop: %" PRIu64 " (%0.2f) queue: %d disk: %d packet: %d close: %d ns: %d frags: %d/%d",
-          totalPackets,
-          moloch_session_watch_count(packet->ses),
-          moloch_session_monitoring(),
-          moloch_session_idle_seconds(packet->ses),
-          stats.total,
-          stats.dropped - initialDropped,
-          (stats.dropped - initialDropped)*(double)100.0/stats.total,
-          moloch_http_queue_length(esServer),
-          moloch_writer_queue_length(),
-          moloch_packet_outstanding(),
-          moloch_session_close_outstanding(),
-          moloch_session_need_save_outstanding(),
-          moloch_packet_frags_outstanding(),
-          moloch_packet_frags_size()
-          );
+    MOLOCH_THREAD_INCR(totalPackets);
+    if (totalPackets % config.logEveryXPackets == 0) {
+        moloch_packet_log(packet->ses);
     }
 
     packet->hash = moloch_session_hash(sessionId);
     uint32_t thread = packet->hash % config.packetThreads;
 
+    totalBytes[thread] += packet->pktlen;
+
     if (DLL_COUNT(packet_, &packetQ[thread]) >= config.maxPacketsInQueue) {
         MOLOCH_LOCK(packetQ[thread].lock);
         overloadDrops[thread]++;
         if ((overloadDrops[thread] % 10000) == 1) {
-            LOG("WARNING - Packet Q %d is overflowing, total dropped %u, increase packetThreads or maxPacketsInQueue", thread, overloadDrops[thread]);
+            LOG("WARNING - Packet Q %u is overflowing, total dropped so far %u.  See https://molo.ch/faq#why-am-i-dropping-packets and modify %s", thread, overloadDrops[thread], config.configFile);
         }
         packet->pkt = 0;
         MOLOCH_COND_SIGNAL(packetQ[thread].lock);
         MOLOCH_UNLOCK(packetQ[thread].lock);
-        return 1;
+        return MOLOCH_PACKET_OVERLOAD_DROPPED;
     }
 
     if (!packet->copied) {
@@ -974,12 +1238,75 @@ int moloch_packet_ip(MolochPacketBatch_t *batch, MolochPacket_t * const packet, 
         packet->copied = 1;
     }
 
+#ifdef FUZZLOCH
+    moloch_session_process_commands(thread);
+    moloch_packet_process(packet, thread);
+#else
     DLL_PUSH_TAIL(packet_, &batch->packetQ[thread], packet);
+#endif
     batch->count++;
-    return 0;
+    return MOLOCH_PACKET_SUCCESS;
 }
 /******************************************************************************/
-int moloch_packet_ip4(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+LOCAL int moloch_packet_ip_gtp(MolochPacketBatch_t *batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+{
+    if (len < 12) {
+        return MOLOCH_PACKET_CORRUPT;
+    }
+    BSB bsb;
+    BSB_INIT(bsb, data, len);
+
+    uint8_t  flags = 0;
+    uint8_t  next = 0;
+
+
+    BSB_IMPORT_u08(bsb, flags);
+    BSB_IMPORT_skip(bsb, 1); // mtype
+    BSB_IMPORT_skip(bsb, 2); // mlen
+    BSB_IMPORT_skip(bsb, 4); // teid
+    if (flags & 0x7) {
+        BSB_IMPORT_skip(bsb, 3);
+        BSB_IMPORT_u08(bsb, next);
+    }
+
+    while (next != 0 && !BSB_IS_ERROR(bsb)) {
+        uint8_t extlen = 0;
+        BSB_IMPORT_u08(bsb, extlen);
+        if (extlen == 0) {
+            return MOLOCH_PACKET_CORRUPT;
+        }
+        BSB_IMPORT_skip(bsb, extlen*4-2);
+        BSB_IMPORT_u08(bsb, next);
+    }
+
+    if (BSB_IS_ERROR(bsb)) {
+        return MOLOCH_PACKET_CORRUPT;
+    }
+
+    packet->tunnel |= MOLOCH_PACKET_TUNNEL_GTP;
+
+    // Should check for v4 vs v6 here
+    BSB_IMPORT_u08(bsb, flags);
+    BSB_IMPORT_rewind(bsb, 1);
+
+    if ((flags & 0xf0) == 0x60)
+        return moloch_packet_ip6(batch, packet, BSB_WORK_PTR(bsb), BSB_REMAINING(bsb));
+    return moloch_packet_ip4(batch, packet, BSB_WORK_PTR(bsb), BSB_REMAINING(bsb));
+}
+/******************************************************************************/
+LOCAL int moloch_packet_ip4_vxlan(MolochPacketBatch_t *batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+{
+    if (len < 8) {
+        return MOLOCH_PACKET_CORRUPT;
+    }
+
+    packet->tunnel |= MOLOCH_PACKET_TUNNEL_VXLAN;
+
+    return moloch_packet_ether(batch, packet, data+8, len-8);
+}
+/******************************************************************************/
+SUPPRESS_ALIGNMENT
+LOCAL int moloch_packet_ip4(MolochPacketBatch_t *batch, MolochPacket_t * const packet, const uint8_t *data, int len)
 {
     struct ip           *ip4 = (struct ip*)data;
     struct tcphdr       *tcphdr = 0;
@@ -988,43 +1315,45 @@ int moloch_packet_ip4(MolochPacketBatch_t * batch, MolochPacket_t * const packet
 
     if (len < (int)sizeof(struct ip)) {
 #ifdef DEBUG_PACKET
-        LOG("too small for header %p %d", packet, len);
+        LOG("BAD PACKET: too small for header %p %d", packet, len);
 #endif
-        return 1;
+        return MOLOCH_PACKET_CORRUPT;
+    }
+
+    if (ip4->ip_v != 4) {
+#ifdef DEBUG_PACKET
+        LOG("BAD PACKET: ip4->ip_v4 %d != 4", ip4->ip_v);
+#endif
+        return MOLOCH_PACKET_CORRUPT;
     }
 
     int ip_len = ntohs(ip4->ip_len);
     if (len < ip_len) {
 #ifdef DEBUG_PACKET
-        LOG("incomplete %p %d", packet, len);
+        LOG("BAD PACKET: incomplete %p %d %d", packet, len, ip_len);
 #endif
-        return 1;
+        return MOLOCH_PACKET_CORRUPT;
     }
 
     int ip_hdr_len = 4 * ip4->ip_hl;
-    if (len < ip_hdr_len) {
+    if (ip_hdr_len < 4 * 5 || len < ip_hdr_len || ip_len < ip_hdr_len) {
 #ifdef DEBUG_PACKET
-        LOG("too small for header and options %p %d", packet, len);
+        LOG("BAD PACKET: too small for header and options %p %d %d", packet, len, ip_hdr_len);
 #endif
-        return 1;
+        return MOLOCH_PACKET_CORRUPT;
     }
-    if (ipTree) {
-        prefix_t prefix;
+    if (ipTree4) {
         patricia_node_t *node;
 
-        prefix.family = AF_INET;
-        prefix.bitlen = 32;
+        if ((node = patricia_search_best3 (ipTree4, (u_char*)&ip4->ip_src, 32)) && node->data == NULL)
+            return MOLOCH_PACKET_IP_DROPPED;
 
-        prefix.add.sin= ip4->ip_src;
-        if ((node = patricia_search_best2 (ipTree, &prefix, 1)) && node->data == NULL)
-            return 1;
-
-        prefix.add.sin= ip4->ip_dst;
-        if ((node = patricia_search_best2 (ipTree, &prefix, 1)) && node->data == NULL)
-            return 1;
+        if ((node = patricia_search_best3 (ipTree4, (u_char*)&ip4->ip_dst, 32)) && node->data == NULL)
+            return MOLOCH_PACKET_IP_DROPPED;
     }
 
     packet->ipOffset = (uint8_t*)data - packet->pkt;
+    packet->v6 = 0;
     packet->payloadOffset = packet->ipOffset + ip_hdr_len;
     packet->payloadLen = ip_len - ip_hdr_len;
 
@@ -1035,36 +1364,91 @@ int moloch_packet_ip4(MolochPacketBatch_t * batch, MolochPacket_t * const packet
 
     if ((ip_flags & IP_MF) || ip_off > 0) {
         moloch_packet_frags4(batch, packet);
-        return 0;
+        return MOLOCH_PACKET_SUCCESS;
     }
 
     switch (ip4->ip_p) {
+    case IPPROTO_IPV4:
+        return moloch_packet_ip4(batch, packet, data + ip_hdr_len, len - ip_hdr_len);
+        break;
     case IPPROTO_TCP:
         if (len < ip_hdr_len + (int)sizeof(struct tcphdr)) {
 #ifdef DEBUG_PACKET
-            LOG("too small for tcp hdr %p %d", packet, len);
+            LOG("BAD PACKET: too small for tcp hdr %p %d", packet, len);
 #endif
-            return 1;
+            return MOLOCH_PACKET_CORRUPT;
         }
 
         tcphdr = (struct tcphdr *)((char*)ip4 + ip_hdr_len);
+
+        if (packetDrop4.drops[tcphdr->th_sport] &&
+            moloch_drophash_should_drop(&packetDrop4, tcphdr->th_sport, &ip4->ip_src.s_addr, packet->ts.tv_sec)) {
+
+            return MOLOCH_PACKET_IPPORT_DROPPED;
+        }
+
+        if (packetDrop4.drops[tcphdr->th_dport] &&
+            moloch_drophash_should_drop(&packetDrop4, tcphdr->th_dport, &ip4->ip_dst.s_addr, packet->ts.tv_sec)) {
+
+            return MOLOCH_PACKET_IPPORT_DROPPED;
+        }
+
         moloch_session_id(sessionId, ip4->ip_src.s_addr, tcphdr->th_sport,
                           ip4->ip_dst.s_addr, tcphdr->th_dport);
         packet->ses = SESSION_TCP;
+
         break;
     case IPPROTO_UDP:
         if (len < ip_hdr_len + (int)sizeof(struct udphdr)) {
 #ifdef DEBUG_PACKET
-        LOG("too small for udp header %p %d", packet, len);
+        LOG("BAD PACKET: too small for udp header %p %d", packet, len);
 #endif
-            return 1;
+            return MOLOCH_PACKET_CORRUPT;
         }
 
         udphdr = (struct udphdr *)((char*)ip4 + ip_hdr_len);
 
+        // See if this is really GTP
+        if (udphdr->uh_dport == 0x6808 && len > ip_hdr_len + (int)sizeof(struct udphdr) + 12) {
+            int rem = len - ip_hdr_len - sizeof(struct udphdr *);
+            uint8_t *buf = (uint8_t *)ip4 + ip_hdr_len + sizeof(struct udphdr *);
+            if ((buf[0] & 0xf0) == 0x30 && buf[1] == 0xff && (buf[2] << 8 | buf[3]) == rem - 8) {
+                return moloch_packet_ip_gtp(batch, packet, buf, rem);
+            }
+        }
+
+        // See if this is really VXLAN
+        if (udphdr->uh_dport == 0xb512 && len > ip_hdr_len + (int)sizeof(struct udphdr) + 8) {
+            int rem = len - ip_hdr_len - sizeof(struct udphdr *);
+            uint8_t *buf = (uint8_t *)ip4 + ip_hdr_len + sizeof(struct udphdr *);
+            if ((buf[0] & 0x77) == 0 && (buf[1] & 0xb7) == 0) {
+                return moloch_packet_ip4_vxlan(batch, packet, buf, rem);
+            }
+        }
+
         moloch_session_id(sessionId, ip4->ip_src.s_addr, udphdr->uh_sport,
                           ip4->ip_dst.s_addr, udphdr->uh_dport);
         packet->ses = SESSION_UDP;
+        break;
+    case IPPROTO_SCTP:
+        if (len < ip_hdr_len + 12) {
+#ifdef DEBUG_PACKET
+            LOG("BAD PACKET: too small for sctp hdr %p %d", packet, len);
+#endif
+            return MOLOCH_PACKET_CORRUPT;
+        }
+        udphdr = (struct udphdr *)((char*)ip4 + ip_hdr_len); /* Not really udp, but port in same location */
+
+        moloch_session_id(sessionId, ip4->ip_src.s_addr, udphdr->uh_sport,
+                          ip4->ip_dst.s_addr, udphdr->uh_dport);
+        packet->ses = SESSION_SCTP;
+        break;
+    case IPPROTO_ESP:
+        if (!config.trackESP)
+            return MOLOCH_PACKET_UNKNOWN;
+        moloch_session_id(sessionId, ip4->ip_src.s_addr, 0,
+                          ip4->ip_dst.s_addr, 0);
+        packet->ses = SESSION_ESP;
         break;
     case IPPROTO_ICMP:
         moloch_session_id(sessionId, ip4->ip_src.s_addr, 0,
@@ -1072,20 +1456,25 @@ int moloch_packet_ip4(MolochPacketBatch_t * batch, MolochPacket_t * const packet
         packet->ses = SESSION_ICMP;
         break;
     case IPPROTO_GRE:
-        packet->vpnType = MOLOCH_PACKET_VPNTYPE_GRE;
+        packet->tunnel |= MOLOCH_PACKET_TUNNEL_GRE;
         packet->vpnIpOffset = packet->ipOffset; // ipOffset will get reset
         return moloch_packet_gre4(batch, packet, data + ip_hdr_len, len - ip_hdr_len);
+    case IPPROTO_IPV6:
+        return moloch_packet_ip6(batch, packet, data + ip_hdr_len, len - ip_hdr_len);
     default:
         if (config.logUnknownProtocols)
             LOG("Unknown protocol %d", ip4->ip_p);
-        return 1;
+        if (BIT_ISSET(ip4->ip_p, config.ipSavePcap))
+            moloch_packet_save_unknown_packet(1, packet);
+        return MOLOCH_PACKET_UNKNOWN;
     }
     packet->protocol = ip4->ip_p;
 
     return moloch_packet_ip(batch, packet, sessionId);
 }
 /******************************************************************************/
-int moloch_packet_ip6(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+SUPPRESS_ALIGNMENT
+LOCAL int moloch_packet_ip6(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
 {
     struct ip6_hdr      *ip6 = (struct ip6_hdr *)data;
     struct tcphdr       *tcphdr = 0;
@@ -1093,12 +1482,22 @@ int moloch_packet_ip6(MolochPacketBatch_t * batch, MolochPacket_t * const packet
     char                 sessionId[MOLOCH_SESSIONID_LEN];
 
     if (len < (int)sizeof(struct ip6_hdr)) {
-        return 1;
+        return MOLOCH_PACKET_CORRUPT;
     }
 
     int ip_len = ntohs(ip6->ip6_plen);
     if (len < ip_len) {
-        return 1;
+        return MOLOCH_PACKET_CORRUPT;
+    }
+
+    if (ipTree6) {
+        patricia_node_t *node;
+
+        if ((node = patricia_search_best3 (ipTree6, (u_char*)&ip6->ip6_src, 128)) && node->data == NULL)
+            return MOLOCH_PACKET_IP_DROPPED;
+
+        if ((node = patricia_search_best3 (ipTree6, (u_char*)&ip6->ip6_dst, 128)) && node->data == NULL)
+            return MOLOCH_PACKET_IP_DROPPED;
     }
 
     int ip_hdr_len = sizeof(struct ip6_hdr);
@@ -1107,6 +1506,9 @@ int moloch_packet_ip6(MolochPacketBatch_t * batch, MolochPacket_t * const packet
     packet->v6 = 1;
 
 
+#ifdef DEBUG_PACKET
+    LOG("Got ip6 header %p %d", packet, packet->pktlen);
+#endif
     int nxt = ip6->ip6_nxt;
     int done = 0;
     do {
@@ -1114,18 +1516,39 @@ int moloch_packet_ip6(MolochPacketBatch_t * batch, MolochPacket_t * const packet
         case IPPROTO_HOPOPTS:
         case IPPROTO_DSTOPTS:
         case IPPROTO_ROUTING:
+            if (len < ip_hdr_len + 2) {
+#ifdef DEBUG_PACKET
+                LOG("ERROR - %d < %d + 2", len, ip_hdr_len);
+#endif
+                return MOLOCH_PACKET_CORRUPT;
+            }
             nxt = data[ip_hdr_len];
             ip_hdr_len += ((data[ip_hdr_len+1] + 1) << 3);
             break;
         case IPPROTO_FRAGMENT:
+#ifdef DEBUG_PACKET
             LOG("ERROR - Don't support ip6 fragements yet!");
-            return 1;
+#endif
+            return MOLOCH_PACKET_UNKNOWN;
         case IPPROTO_TCP:
             if (len < ip_hdr_len + (int)sizeof(struct tcphdr)) {
-                return 1;
+                return MOLOCH_PACKET_CORRUPT;
             }
 
             tcphdr = (struct tcphdr *)(data + ip_hdr_len);
+
+
+            if (packetDrop6.drops[tcphdr->th_sport] &&
+                moloch_drophash_should_drop(&packetDrop6, tcphdr->th_sport, &ip6->ip6_src, packet->ts.tv_sec)) {
+
+                return MOLOCH_PACKET_IPPORT_DROPPED;
+            }
+
+            if (packetDrop6.drops[tcphdr->th_dport] &&
+                moloch_drophash_should_drop(&packetDrop6, tcphdr->th_dport, &ip6->ip6_dst, packet->ts.tv_sec)) {
+
+                return MOLOCH_PACKET_IPPORT_DROPPED;
+            }
 
             moloch_session_id6(sessionId, ip6->ip6_src.s6_addr, tcphdr->th_sport,
                                ip6->ip6_dst.s6_addr, tcphdr->th_dport);
@@ -1134,7 +1557,7 @@ int moloch_packet_ip6(MolochPacketBatch_t * batch, MolochPacket_t * const packet
             break;
         case IPPROTO_UDP:
             if (len < ip_hdr_len + (int)sizeof(struct udphdr)) {
-                return 1;
+                return MOLOCH_PACKET_CORRUPT;
             }
 
             udphdr = (struct udphdr *)(data + ip_hdr_len);
@@ -1142,7 +1565,36 @@ int moloch_packet_ip6(MolochPacketBatch_t * batch, MolochPacket_t * const packet
             moloch_session_id6(sessionId, ip6->ip6_src.s6_addr, udphdr->uh_sport,
                                ip6->ip6_dst.s6_addr, udphdr->uh_dport);
 
+            // See if this is really GTP
+            if (udphdr->uh_dport == 0x6808 && len > ip_hdr_len + (int)sizeof(struct udphdr) + 12) {
+                int rem = len - ip_hdr_len - sizeof(struct udphdr *);
+                const uint8_t *buf = (uint8_t *)udphdr + sizeof(struct udphdr *);
+                if ((buf[0] & 0xf0) == 0x30 && buf[1] == 0xff && (buf[2] << 8 | buf[3]) == rem - 8) {
+                    return moloch_packet_ip_gtp(batch, packet, buf, rem);
+                }
+            }
+
             packet->ses = SESSION_UDP;
+            done = 1;
+            break;
+        case IPPROTO_SCTP:
+            if (len < ip_hdr_len + 12) {
+                return MOLOCH_PACKET_CORRUPT;
+            }
+
+            udphdr = (struct udphdr *)(data + ip_hdr_len); /* Not really udp, but port in same location */
+
+            moloch_session_id6(sessionId, ip6->ip6_src.s6_addr, udphdr->uh_sport,
+                               ip6->ip6_dst.s6_addr, udphdr->uh_dport);
+            packet->ses = SESSION_SCTP;
+            done = 1;
+            break;
+        case IPPROTO_ESP:
+            if (!config.trackESP)
+                return MOLOCH_PACKET_UNKNOWN;
+            moloch_session_id6(sessionId, ip6->ip6_src.s6_addr, 0,
+                               ip6->ip6_dst.s6_addr, 0);
+            packet->ses = SESSION_ESP;
             done = 1;
             break;
         case IPPROTO_ICMP:
@@ -1157,48 +1609,156 @@ int moloch_packet_ip6(MolochPacketBatch_t * batch, MolochPacket_t * const packet
             packet->ses = SESSION_ICMP;
             done = 1;
             break;
+        case IPPROTO_IPV4:
+            return moloch_packet_ip4(batch, packet, data + ip_hdr_len, len - ip_hdr_len);
+        case IPPROTO_IPV6:
+            return moloch_packet_ip6(batch, packet, data + ip_hdr_len, len - ip_hdr_len);
         default:
-            LOG("Unknown protocol %d", ip6->ip6_nxt);
-            return 1;
+            if (config.logUnknownProtocols)
+                LOG("Unknown protocol %d %d", ip6->ip6_nxt, nxt);
+            if (BIT_ISSET(nxt, config.ipSavePcap))
+                moloch_packet_save_unknown_packet(1, packet);
+            return MOLOCH_PACKET_UNKNOWN;
         }
         if (ip_hdr_len > len) {
+#ifdef DEBUG_PACKET
             LOG ("ERROR - Corrupt packet ip_hdr_len = %d nxt = %d len = %d", ip_hdr_len, nxt, len);
-            return 1;
+#endif
+            return MOLOCH_PACKET_CORRUPT;
         }
     } while (!done);
 
     packet->protocol = nxt;
     packet->payloadOffset = packet->ipOffset + ip_hdr_len;
-    packet->payloadLen = ip_len - ip_hdr_len + sizeof(struct ip6_hdr);
+
+    if (ip_len + (int)sizeof(struct ip6_hdr) < ip_hdr_len) {
+#ifdef DEBUG_PACKET
+        LOG ("ERROR - %d + %ld < %d", ip_len, (long)sizeof(struct ip6_hdr), ip_hdr_len);
+#endif
+        return MOLOCH_PACKET_CORRUPT;
+    }
+    packet->payloadLen = ip_len + sizeof(struct ip6_hdr) - ip_hdr_len;
+
+    if (packet->pktlen < packet->payloadOffset + packet->payloadLen) {
+#ifdef DEBUG_PACKET
+        LOG ("ERROR - %d < %d + %d", packet->pktlen, packet->payloadOffset, packet->payloadLen);
+#endif
+        return MOLOCH_PACKET_CORRUPT;
+    }
+
     return moloch_packet_ip(batch, packet, sessionId);
 }
 /******************************************************************************/
-int moloch_packet_pppoe(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+LOCAL int moloch_packet_frame_relay(MolochPacketBatch_t *batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+{
+    if (len < 4)
+        return MOLOCH_PACKET_CORRUPT;
+
+    if (data[2] == 0x03 || data[3] == 0xcc)
+        return moloch_packet_ip4(batch, packet, data+4, len-4);
+    if (data[2] == 0x08 || data[3] == 0x00)
+        return moloch_packet_ip4(batch, packet, data+4, len-4);
+    if (data[2] == 0x86 || data[3] == 0xdd)
+        return moloch_packet_ip6(batch, packet, data+4, len-4);
+
+    return MOLOCH_PACKET_UNKNOWN;
+}
+/******************************************************************************/
+LOCAL int moloch_packet_pppoe(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
 {
     if (len < 8 || data[0] != 0x11 || data[1] != 0) {
-        return 1;
+#ifdef DEBUG_PACKET
+        LOG("BAD PACKET: Len or bytes %d %d %d", len, data[0], data[1]);
+#endif
+        return MOLOCH_PACKET_CORRUPT;
     }
 
     uint16_t plen = data[4] << 8 | data[5];
     uint16_t type = data[6] << 8 | data[7];
     if (plen != len-6)
-        return 1;
+        return MOLOCH_PACKET_CORRUPT;
 
-    packet->vpnType = MOLOCH_PACKET_VPNTYPE_PPPOE;
+    packet->tunnel |= MOLOCH_PACKET_TUNNEL_PPPOE;
     switch (type) {
     case 0x21:
         return moloch_packet_ip4(batch, packet, data + 8, plen-2);
     case 0x57:
         return moloch_packet_ip6(batch, packet, data + 8, plen-2);
     default:
-        return 1;
+#ifdef DEBUG_PACKET
+        LOG("BAD PACKET: Unknown pppoe type %d", type);
+#endif
+        return MOLOCH_PACKET_UNKNOWN;
     }
 }
 /******************************************************************************/
-int moloch_packet_ether(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+LOCAL int moloch_packet_ppp(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+{
+    if (len < 4 || data[2] != 0x00) {
+#ifdef DEBUG_PACKET
+        LOG("BAD PACKET: Len or bytes %d %d %d", len, data[2], data[3]);
+#endif
+        return MOLOCH_PACKET_CORRUPT;
+    }
+
+    packet->tunnel |= MOLOCH_PACKET_TUNNEL_PPP;
+    switch (data[3]) {
+    case 0x21:
+        return moloch_packet_ip4(batch, packet, data + 4, len-4);
+    case 0x57:
+        return moloch_packet_ip6(batch, packet, data + 4, len-4);
+    default:
+#ifdef DEBUG_PACKET
+        LOG("BAD PACKET: Unknown ppp type %d", data[3]);
+#endif
+        if (BIT_ISSET(0x880b, config.etherSavePcap))
+            moloch_packet_save_unknown_packet(0, packet);
+        return MOLOCH_PACKET_UNKNOWN;
+    }
+}
+/******************************************************************************/
+LOCAL int moloch_packet_mpls(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+{
+    while (1) {
+        if (len < 4 + (int)sizeof(struct ip)) {
+#ifdef DEBUG_PACKET
+            LOG("BAD PACKET: Len %d", len);
+#endif
+            return MOLOCH_PACKET_CORRUPT;
+        }
+
+        int S = data[2] & 0x1;
+
+        data += 4;
+        len -= 4;
+
+        if (S) {
+            packet->tunnel |= MOLOCH_PACKET_TUNNEL_MPLS;
+            switch (data[0] >> 4) {
+            case 4:
+                return moloch_packet_ip4(batch, packet, data, len);
+            case 6:
+                return moloch_packet_ip6(batch, packet, data, len);
+            default:
+#ifdef DEBUG_PACKET
+                LOG("BAD PACKET: Unknown mpls type %d", data[0] >> 4);
+#endif
+                if (BIT_ISSET(0x8847, config.etherSavePcap))
+                    moloch_packet_save_unknown_packet(0, packet);
+                return MOLOCH_PACKET_UNKNOWN;
+            }
+        }
+    }
+    return MOLOCH_PACKET_CORRUPT;
+}
+/******************************************************************************/
+LOCAL int moloch_packet_ether(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
 {
     if (len < 14) {
-        return 1;
+#ifdef DEBUG_PACKET
+        LOG("BAD PACKET: Too short %d", len);
+#endif
+        return MOLOCH_PACKET_CORRUPT;
     }
     int n = 12;
     while (n+2 < len) {
@@ -1211,22 +1771,70 @@ int moloch_packet_ether(MolochPacketBatch_t * batch, MolochPacket_t * const pack
             return moloch_packet_ip6(batch, packet, data+n, len - n);
         case 0x8864:
             return moloch_packet_pppoe(batch, packet, data+n, len - n);
+        case 0x8847:
+            return moloch_packet_mpls(batch, packet, data+n, len - n);
         case 0x8100:
             n += 2;
             break;
         default:
-            return 1;
+#ifdef DEBUG_PACKET
+            LOG("BAD PACKET: Unknown ethertype %x", ethertype);
+#endif
+            if (BIT_ISSET(ethertype, config.etherSavePcap))
+                moloch_packet_save_unknown_packet(0, packet);
+            return MOLOCH_PACKET_UNKNOWN;
         } // switch
     }
-    return 1;
+#ifdef DEBUG_PACKET
+    LOG("BAD PACKET: bad len %d < %d", n+2, len);
+#endif
+    return MOLOCH_PACKET_CORRUPT;
 }
 /******************************************************************************/
-int moloch_packet_nflog(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+LOCAL int moloch_packet_sll(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+{
+    if (len < 16) {
+#ifdef DEBUG_PACKET
+        LOG("BAD PACKET: Too short %d", len);
+#endif
+        return MOLOCH_PACKET_CORRUPT;
+    }
+
+    int ethertype = data[14] << 8 | data[15];
+    switch (ethertype) {
+    case 0x0800:
+        return moloch_packet_ip4(batch, packet, data+16, len - 16);
+    case 0x86dd:
+        return moloch_packet_ip6(batch, packet, data+16, len - 16);
+    case 0x8864:
+        return moloch_packet_pppoe(batch, packet, data+16, len - 16);
+    case 0x8847:
+        return moloch_packet_mpls(batch, packet, data+16, len - 16);
+    case 0x8100:
+        if ((data[20] & 0xf0) == 0x60)
+            return moloch_packet_ip6(batch, packet, data+20, len - 20);
+        else
+            return moloch_packet_ip4(batch, packet, data+20, len - 20);
+    default:
+#ifdef DEBUG_PACKET
+        LOG("BAD PACKET: Unknown ethertype %x", ethertype);
+#endif
+        if (BIT_ISSET(ethertype, config.etherSavePcap))
+            moloch_packet_save_unknown_packet(0, packet);
+        return MOLOCH_PACKET_UNKNOWN;
+    } // switch
+    return MOLOCH_PACKET_CORRUPT;
+}
+/******************************************************************************/
+LOCAL int moloch_packet_nflog(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
 {
     if (len < 14 ||
         (data[0] != AF_INET && data[0] != AF_INET6) ||
         data[1] != 0) {
-        return 1;
+#ifdef DEBUG_PACKET
+        LOG("BAD PACKET: Wrong type %d", data[0]);
+#endif
+        return MOLOCH_PACKET_CORRUPT;
     }
     int n = 4;
     while (n+4 < len) {
@@ -1234,7 +1842,10 @@ int moloch_packet_nflog(MolochPacketBatch_t * batch, MolochPacket_t * const pack
 
         // Make sure length is at least header and not bigger then remaining packet
         if (length < 4 || length > len - n) {
-            return 1;
+#ifdef DEBUG_PACKET
+            LOG("BAD PACKET: Wrong len %d", length);
+#endif
+            return MOLOCH_PACKET_CORRUPT;
         }
 
         if (data[n+3] == 0 && data[n+2] == 9) {
@@ -1247,7 +1858,40 @@ int moloch_packet_nflog(MolochPacketBatch_t * batch, MolochPacket_t * const pack
             n += ((length + 3) & 0xfffffc);
         }
     }
-    return 1;
+#ifdef DEBUG_PACKET
+    LOG("BAD PACKET: Not sure");
+#endif
+    return MOLOCH_PACKET_CORRUPT;
+}
+/******************************************************************************/
+LOCAL int moloch_packet_radiotap(MolochPacketBatch_t * batch, MolochPacket_t * const packet, const uint8_t *data, int len)
+{
+    if (data[0] != 0 || len < 36)
+        return MOLOCH_PACKET_UNKNOWN;
+
+    int hl = packet->pkt[2];
+    if (hl + 24 + 8 >= len)
+        return MOLOCH_PACKET_UNKNOWN;
+
+    if (data[hl] != 8)
+        return MOLOCH_PACKET_UNKNOWN;
+
+    hl += 24 + 3;
+
+    if (data[hl] != 0 || data[hl+1] != 0 || data[hl+2] != 0)
+        return MOLOCH_PACKET_UNKNOWN;
+
+    hl += 3;
+
+    if (data[hl] == 0x08 && data[hl+1] == 0x00) {
+        hl += 2;
+        return moloch_packet_ip4(batch, packet, data+hl, len - hl);
+    }
+    if (data[hl] == 0x86 && data[hl+1] == 0xdd) {
+        hl += 2;
+        return moloch_packet_ip6(batch, packet, data+hl, len - hl);
+    }
+    return MOLOCH_PACKET_UNKNOWN;
 }
 /******************************************************************************/
 void moloch_packet_batch_init(MolochPacketBatch_t *batch)
@@ -1280,30 +1924,60 @@ void moloch_packet_batch(MolochPacketBatch_t * batch, MolochPacket_t * const pac
     int rc;
 
 #ifdef DEBUG_PACKET
-    LOG("enter %p %d %d", packet, pcapFileHeader.linktype, packet->pktlen);
+    LOG("enter %p %u %d", packet, pcapFileHeader.linktype, packet->pktlen);
 #endif
 
     switch(pcapFileHeader.linktype) {
     case 0: // NULL
-        if (packet->pktlen > 4)
-            rc = moloch_packet_ip4(batch, packet, packet->pkt+4, packet->pktlen-4);
-        else
-            rc = 1;
+        if (packet->pktlen > 4) {
+            if (packet->pkt[0] == 30)
+                rc = moloch_packet_ip6(batch, packet, packet->pkt+4, packet->pktlen-4);
+            else
+                rc = moloch_packet_ip4(batch, packet, packet->pkt+4, packet->pktlen-4);
+        } else {
+#ifdef DEBUG_PACKET
+            LOG("BAD PACKET: Too short %d", packet->pktlen);
+#endif
+            rc = MOLOCH_PACKET_CORRUPT;
+        }
         break;
     case 1: // Ether
         rc = moloch_packet_ether(batch, packet, packet->pkt, packet->pktlen);
         break;
     case 12: // LOOP
     case 101: // RAW
+        rc = moloch_packet_ip4(batch, packet, packet->pkt, packet->pktlen);
+        break;
+    case 107: // Frame Relay
+        rc = moloch_packet_frame_relay(batch, packet, packet->pkt, packet->pktlen);
+        break;
     case 113: // SLL
+        if (packet->pkt[0] == 0 && packet->pkt[1] <= 4)
+            rc = moloch_packet_sll(batch, packet, packet->pkt, packet->pktlen);
+        else
+            rc = moloch_packet_ip4(batch, packet, packet->pkt, packet->pktlen);
+        break;
+    case 127: // radiotap
+        rc = moloch_packet_radiotap(batch, packet, packet->pkt, packet->pktlen);
+        break;
+    case 228: //RAW IPv4
         rc = moloch_packet_ip4(batch, packet, packet->pkt, packet->pktlen);
         break;
     case 239: // NFLOG
         rc = moloch_packet_nflog(batch, packet, packet->pkt, packet->pktlen);
         break;
     default:
-        LOGEXIT("ERROR - Unsupported pcap link type %d", pcapFileHeader.linktype);
+        if (config.ignoreErrors)
+            rc = MOLOCH_PACKET_CORRUPT;
+        else
+            LOGEXIT("ERROR - Unsupported pcap link type %u", pcapFileHeader.linktype);
     }
+    if (rc == MOLOCH_PACKET_CORRUPT && config.corruptSavePcap) {
+        moloch_packet_save_unknown_packet(2, packet);
+    }
+
+    MOLOCH_THREAD_INCR(packetStats[rc]);
+
     if (rc) {
         moloch_packet_free(packet);
     }
@@ -1321,7 +1995,8 @@ int moloch_packet_outstanding()
     return count;
 }
 /******************************************************************************/
-uint32_t moloch_packet_frag_hash(const void *key)
+SUPPRESS_UNSIGNED_INTEGER_OVERFLOW
+LOCAL uint32_t moloch_packet_frag_hash(const void *key)
 {
     int i;
     uint32_t n = 0;
@@ -1331,11 +2006,22 @@ uint32_t moloch_packet_frag_hash(const void *key)
     return n;
 }
 /******************************************************************************/
-int moloch_packet_frag_cmp(const void *keyv, const void *elementv)
+LOCAL int moloch_packet_frag_cmp(const void *keyv, const void *elementv)
 {
     MolochFrags_t *element = (MolochFrags_t *)elementv;
 
     return memcmp(keyv, element->key, 10) == 0;
+}
+/******************************************************************************/
+LOCAL gboolean moloch_packet_save_drophash(gpointer UNUSED(user_data))
+{
+    if (packetDrop4.changed)
+        moloch_drophash_save(&packetDrop4);
+
+    if (packetDrop6.changed)
+        moloch_drophash_save(&packetDrop6);
+
+    return TRUE;
 }
 /******************************************************************************/
 void moloch_packet_init()
@@ -1347,94 +2033,138 @@ void moloch_packet_init()
     pcapFileHeader.thiszone = 0;
     pcapFileHeader.sigfigs = 0;
 
+    char filename[PATH_MAX];
+    snprintf(filename, sizeof(filename), "/tmp/%s.tcp.drops.4", config.nodeName);
+    moloch_drophash_init(&packetDrop4, filename, TRUE);
+
+    snprintf(filename, sizeof(filename), "/tmp/%s.tcp.drops.6", config.nodeName);
+    moloch_drophash_init(&packetDrop6, filename, FALSE);
+    g_timeout_add_seconds(10, moloch_packet_save_drophash, 0);
+
     mac1Field = moloch_field_define("general", "lotermfield",
-        "mac.src", "Src MAC", "mac1-term",
+        "mac.src", "Src MAC", "srcMac",
         "Source ethernet mac addresses set for session",
-        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_COUNT | MOLOCH_FIELD_FLAG_LINKED_SESSIONS,
-        NULL);
+        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT | MOLOCH_FIELD_FLAG_LINKED_SESSIONS,
+        "transform", "dash2Colon",
+        (char *)NULL);
 
     mac2Field = moloch_field_define("general", "lotermfield",
-        "mac.dst", "Dst MAC", "mac2-term",
+        "mac.dst", "Dst MAC", "dstMac",
         "Destination ethernet mac addresses set for session",
-        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_COUNT | MOLOCH_FIELD_FLAG_LINKED_SESSIONS,
-        NULL);
+        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT | MOLOCH_FIELD_FLAG_LINKED_SESSIONS,
+        "transform", "dash2Colon",
+        (char *)NULL);
 
     moloch_field_define("general", "lotermfield",
         "mac", "Src or Dst MAC", "macall",
         "Shorthand for mac.src or mac.dst",
         0,  MOLOCH_FIELD_FLAG_FAKE,
         "regex", "^mac\\\\.(?:(?!\\\\.cnt$).)*$",
-        NULL);
+        "transform", "dash2Colon",
+        (char *)NULL);
+
+    oui1Field = moloch_field_define("general", "termfield",
+        "oui.src", "Src OUI", "srcOui",
+        "Source ethernet oui set for session",
+        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT | MOLOCH_FIELD_FLAG_LINKED_SESSIONS,
+        (char *)NULL);
+
+    oui2Field = moloch_field_define("general", "termfield",
+        "oui.dst", "Dst OUI", "dstOui",
+        "Destination ethernet oui set for session",
+        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT | MOLOCH_FIELD_FLAG_LINKED_SESSIONS,
+        (char *)NULL);
+
 
     vlanField = moloch_field_define("general", "integer",
         "vlan", "VLan", "vlan",
         "vlan value",
-        MOLOCH_FIELD_TYPE_INT_GHASH,  MOLOCH_FIELD_FLAG_COUNT | MOLOCH_FIELD_FLAG_LINKED_SESSIONS,
-        NULL);
+        MOLOCH_FIELD_TYPE_INT_GHASH,  MOLOCH_FIELD_FLAG_CNT | MOLOCH_FIELD_FLAG_LINKED_SESSIONS,
+        (char *)NULL);
 
     greIpField = moloch_field_define("general", "ip",
-        "gre.ip", "GRE IP", "greip",
+        "gre.ip", "GRE IP", "greIp",
         "GRE ip addresses for session",
-        MOLOCH_FIELD_TYPE_IP_GHASH,  MOLOCH_FIELD_FLAG_COUNT | MOLOCH_FIELD_FLAG_LINKED_SESSIONS,
-        NULL);
-
-    moloch_field_define("general", "lotermfield",
-        "tipv6.src", "IPv6 Src", "tipv61-term",
-        "Temporary IPv6 Source",
-        0,  MOLOCH_FIELD_FLAG_FAKE,
-        "portField", "p1",
-        "transform", "ipv6ToHex",
-        NULL);
-
-    moloch_field_define("general", "lotermfield",
-        "tipv6.dst", "IPv6 Dst", "tipv62-term",
-        "Temporary IPv6 Destination",
-        0,  MOLOCH_FIELD_FLAG_FAKE,
-        "portField", "p2",
-        "transform", "ipv6ToHex",
-        NULL);
+        MOLOCH_FIELD_TYPE_IP_GHASH,  MOLOCH_FIELD_FLAG_CNT | MOLOCH_FIELD_FLAG_LINKED_SESSIONS,
+        (char *)NULL);
 
     moloch_field_define("general", "integer",
         "tcpflags.syn", "TCP Flag SYN", "tcpflags.syn",
         "Count of packets with SYN and no ACK flag set",
         0,  MOLOCH_FIELD_FLAG_FAKE,
-        NULL);
+        (char *)NULL);
 
     moloch_field_define("general", "integer",
         "tcpflags.syn-ack", "TCP Flag SYN-ACK", "tcpflags.syn-ack",
         "Count of packets with SYN and ACK flag set",
         0,  MOLOCH_FIELD_FLAG_FAKE,
-        NULL);
+        (char *)NULL);
 
     moloch_field_define("general", "integer",
         "tcpflags.ack", "TCP Flag ACK", "tcpflags.ack",
         "Count of packets with only the ACK flag set",
         0,  MOLOCH_FIELD_FLAG_FAKE,
-        NULL);
+        (char *)NULL);
 
     moloch_field_define("general", "integer",
         "tcpflags.psh", "TCP Flag PSH", "tcpflags.psh",
         "Count of packets with PSH flag set",
         0,  MOLOCH_FIELD_FLAG_FAKE,
-        NULL);
+        (char *)NULL);
 
     moloch_field_define("general", "integer",
         "tcpflags.fin", "TCP Flag FIN", "tcpflags.fin",
         "Count of packets with FIN flag set",
         0,  MOLOCH_FIELD_FLAG_FAKE,
-        NULL);
+        (char *)NULL);
 
     moloch_field_define("general", "integer",
         "tcpflags.rst", "TCP Flag RST", "tcpflags.rst",
         "Count of packets with RST flag set",
         0,  MOLOCH_FIELD_FLAG_FAKE,
-        NULL);
+        (char *)NULL);
 
     moloch_field_define("general", "integer",
         "tcpflags.urg", "TCP Flag URG", "tcpflags.urg",
         "Count of packets with URG flag set",
         0,  MOLOCH_FIELD_FLAG_FAKE,
-        NULL);
+        (char *)NULL);
+
+    icmpTypeField = moloch_field_define("general", "integer",
+        "icmp.type", "ICMP Type", "icmp.type",
+        "ICMP type field values",
+        MOLOCH_FIELD_TYPE_INT_GHASH, 0,
+        (char *)NULL);
+
+    icmpCodeField = moloch_field_define("general", "integer",
+        "icmp.code", "ICMP Code", "icmp.code",
+        "ICMP code field values",
+        MOLOCH_FIELD_TYPE_INT_GHASH, 0,
+        (char *)NULL);
+
+    moloch_field_define("general", "integer",
+        "packets.src", "Src Packets", "srcPackets",
+        "Total number of packets sent by source in a session",
+        0,  MOLOCH_FIELD_FLAG_FAKE,
+        (char *)NULL);
+
+    moloch_field_define("general", "integer",
+        "packets.dst", "Dst Packets", "dstPackets",
+        "Total number of packets sent by destination in a session",
+        0,  MOLOCH_FIELD_FLAG_FAKE,
+        (char *)NULL);
+
+    moloch_field_define("general", "integer",
+        "initRTT", "Initial RTT", "initRTT",
+        "Initial round trip time, difference between SYN and ACK timestamp divided by 2 in ms",
+        0,  MOLOCH_FIELD_FLAG_FAKE,
+        (char *)NULL);
+
+    moloch_field_define("general", "termfield",
+        "communityId", "Community Id", "communityId",
+        "Community id flow hash",
+        0,  MOLOCH_FIELD_FLAG_FAKE,
+        (char *)NULL);
 
     int t;
     for (t = 0; t < config.packetThreads; t++) {
@@ -1443,7 +2173,9 @@ void moloch_packet_init()
         MOLOCH_LOCK_INIT(packetQ[t].lock);
         MOLOCH_COND_INIT(packetQ[t].lock);
         snprintf(name, sizeof(name), "moloch-pkt%d", t);
-        g_thread_new(name, &moloch_packet_thread, (gpointer)(long)t);
+#ifndef FUZZLOCH
+        g_thread_unref(g_thread_new(name, &moloch_packet_thread, (gpointer)(long)t));
+#endif
     }
 
     HASH_INIT(fragh_, fragsHash, moloch_packet_frag_hash, moloch_packet_frag_cmp);
@@ -1451,6 +2183,7 @@ void moloch_packet_init()
 
     moloch_add_can_quit(moloch_packet_outstanding, "packet outstanding");
     moloch_add_can_quit(moloch_packet_frags_outstanding, "packet frags outstanding");
+    maxTcpOutOfOrderPackets = moloch_config_int(NULL, "maxTcpOutOfOrderPackets", 256, 64, 10000);
 }
 /******************************************************************************/
 uint64_t moloch_packet_dropped_packets()
@@ -1479,13 +2212,30 @@ uint64_t moloch_packet_dropped_overload()
     return count;
 }
 /******************************************************************************/
+uint64_t moloch_packet_total_bytes()
+{
+    uint64_t count = 0;
+
+    int t;
+
+    for (t = 0; t < config.packetThreads; t++) {
+        count += totalBytes[t];
+    }
+    return count;
+}
+/******************************************************************************/
 void moloch_packet_add_packet_ip(char *ipstr, int mode)
 {
     patricia_node_t *node;
-    if (!ipTree) {
-        ipTree = New_Patricia(128);
+    if (strchr(ipstr, '.') != 0) {
+        if (!ipTree4)
+            ipTree4 = New_Patricia(32);
+        node = make_and_lookup(ipTree4, ipstr);
+    } else {
+        if (!ipTree6)
+            ipTree6 = New_Patricia(128);
+        node = make_and_lookup(ipTree6, ipstr);
     }
-    node = make_and_lookup(ipTree, ipstr);
     node->data = (void *)(long)mode;
 }
 /******************************************************************************/
@@ -1496,10 +2246,45 @@ void moloch_packet_set_linksnap(int linktype, int snaplen)
     moloch_rules_recompile();
 }
 /******************************************************************************/
+void moloch_packet_drophash_add(MolochSession_t *session, int which, int min)
+{
+    if (session->ses != SESSION_TCP)
+        return;
+
+    // packetDrop is kept in network byte order
+    const int port = (which == 0)?htons(session->port1):htons(session->port2);
+
+    if (MOLOCH_SESSION_v6(session)) {
+        if (which == 0) {
+            moloch_drophash_add(&packetDrop6, port, (void*)&session->addr1, session->lastPacket.tv_sec, min*60);
+        } else {
+            moloch_drophash_add(&packetDrop6, port, (void*)&session->addr2, session->lastPacket.tv_sec, min*60);
+        }
+    } else {
+        if (which == 0) {
+            moloch_drophash_add(&packetDrop4, port, &((uint32_t *)session->addr1.s6_addr)[3], session->lastPacket.tv_sec, min*60);
+        } else {
+            moloch_drophash_add(&packetDrop4, port, &((uint32_t *)session->addr2.s6_addr)[3], session->lastPacket.tv_sec, min*60);
+        }
+    }
+}
+/******************************************************************************/
 void moloch_packet_exit()
 {
-    if (ipTree) {
-        Destroy_Patricia(ipTree, NULL);
-        ipTree = 0;
+    if (ipTree4) {
+        Destroy_Patricia(ipTree4, NULL);
+        ipTree4 = 0;
     }
+
+    if (ipTree6) {
+        Destroy_Patricia(ipTree6, NULL);
+        ipTree6 = 0;
+    }
+    moloch_packet_log(SESSION_TCP);
+    if (unknownPacketFile[0])
+        fclose(unknownPacketFile[0]);
+    if (unknownPacketFile[1])
+        fclose(unknownPacketFile[1]);
+    if (unknownPacketFile[2])
+        fclose(unknownPacketFile[2]);
 }
