@@ -28,12 +28,13 @@ typedef struct {
     GString         *authString;
 
     GString         *valueString[2];
+    gboolean        isValueStringLowerCase[2];
 
     char             header[2][40];
     short            pos[2];
     http_parser      parsers[2];
 
-    GChecksum       *checksum[2];
+    GChecksum       *checksum[4];
     const char      *magicString[2];
 
     uint16_t         wParsers:2;
@@ -42,37 +43,44 @@ typedef struct {
     uint16_t         inBody:2;
     uint16_t         urlWhich:1;
     uint16_t         which:1;
+    uint16_t         isConnect:1;
+    uint16_t         reclassify:2;
 } HTTPInfo_t;
 
 extern MolochConfig_t        config;
-static http_parser_settings  parserSettings;
+LOCAL  http_parser_settings  parserSettings;
 extern uint32_t              pluginsCbs;
-static MolochStringHashStd_t httpReqHeaders;
-static MolochStringHashStd_t httpResHeaders;
+LOCAL  MolochStringHashStd_t httpReqHeaders;
+LOCAL  MolochStringHashStd_t httpResHeaders;
 
-static int cookieKeyField;
-static int cookieValueField;
-static int hostField;
-static int userField;
-static int atField;
-static int urlsField;
-static int xffField;
-static int uaField;
-static int tagsReqField;
-static int tagsResField;
-static int md5Field;
-static int verReqField;
-static int verResField;
-static int pathField;
-static int keyField;
-static int valueField;
-static int magicField;
-static int statuscodeField;
-static int methodField;
+LOCAL  int cookieKeyField;
+LOCAL  int cookieValueField;
+LOCAL  int hostField;
+LOCAL  int userField;
+LOCAL  int atField;
+LOCAL  int urlsField;
+LOCAL  int xffField;
+LOCAL  int uaField;
+LOCAL  int tagsReqField;
+LOCAL  int tagsResField;
+LOCAL  int md5Field;
+LOCAL  int sha256Field;
+LOCAL  int verReqField;
+LOCAL  int verResField;
+LOCAL  int pathField;
+LOCAL  int keyField;
+LOCAL  int valueField;
+LOCAL  int magicField;
+LOCAL  int statuscodeField;
+LOCAL  int methodField;
+LOCAL  int reqBodyField;
+LOCAL  int headerReqField;
+LOCAL  int headerReqValue;
+LOCAL  int headerResField;
+LOCAL  int headerResValue;
 
 /******************************************************************************/
-int
-moloch_hp_cb_on_message_begin (http_parser *parser)
+LOCAL int moloch_hp_cb_on_message_begin (http_parser *parser)
 {
     HTTPInfo_t            *http = parser->data;
     MolochSession_t       *session = http->session;
@@ -86,6 +94,9 @@ moloch_hp_cb_on_message_begin (http_parser *parser)
     http->inValue  &= ~(1 << http->which);
     http->inBody   &= ~(1 << http->which);
     g_checksum_reset(http->checksum[http->which]);
+    if (config.supportSha256) {
+        g_checksum_reset(http->checksum[http->which+2]);
+    }
 
     if (pluginsCbs & MOLOCH_PLUGIN_HP_OMB)
         moloch_plugins_cb_hp_omb(session, parser);
@@ -93,8 +104,7 @@ moloch_hp_cb_on_message_begin (http_parser *parser)
     return 0;
 }
 /******************************************************************************/
-int
-moloch_hp_cb_on_url (http_parser *parser, const char *at, size_t length)
+LOCAL int moloch_hp_cb_on_url (http_parser *parser, const char *at, size_t length)
 {
     HTTPInfo_t            *http = parser->data;
 
@@ -112,8 +122,7 @@ moloch_hp_cb_on_url (http_parser *parser, const char *at, size_t length)
 }
 
 /******************************************************************************/
-int
-moloch_hp_cb_on_body (http_parser *parser, const char *at, size_t length)
+LOCAL int moloch_hp_cb_on_body (http_parser *parser, const char *at, size_t length)
 {
     HTTPInfo_t            *http = parser->data;
     MolochSession_t       *session = http->session;
@@ -132,9 +141,20 @@ moloch_hp_cb_on_body (http_parser *parser, const char *at, size_t length)
 
         http->magicString[http->which] = moloch_parsers_magic(session, magicField, at, length);
         http->inBody |= (1 << http->which);
+
+        /* Put small requests in a field. */
+        if (http->which == http->urlWhich && length <= config.maxReqBody && length > 0) {
+            if (!config.reqBodyOnlyUtf8 || g_utf8_validate(at, length, NULL) == TRUE) {
+                moloch_field_string_add(reqBodyField, session, at, length, TRUE);
+            }
+        }
+
     }
 
     g_checksum_update(http->checksum[http->which], (guchar *)at, length);
+    if (config.supportSha256) {
+        g_checksum_update(http->checksum[http->which+2], (guchar *)at, length);
+    }
 
     if (pluginsCbs & MOLOCH_PLUGIN_HP_OB)
         moloch_plugins_cb_hp_ob(session, parser, at, length);
@@ -143,8 +163,7 @@ moloch_hp_cb_on_body (http_parser *parser, const char *at, size_t length)
 }
 
 /******************************************************************************/
-void
-moloch_http_parse_authorization(MolochSession_t *session, char *str)
+LOCAL void moloch_http_parse_authorization(MolochSession_t *session, char *str)
 {
     gsize olen;
 
@@ -155,17 +174,14 @@ moloch_http_parse_authorization(MolochSession_t *session, char *str)
     if (!space)
         return;
 
-    char *lower = g_ascii_strdown(str, space-str);
-    if (!moloch_field_string_add(atField, session, lower, space-str, FALSE)) {
-        g_free(lower);
-    }
+    moloch_field_string_add_lower(atField, session, str, space-str);
 
     if (strncasecmp("basic", str, 5) == 0) {
         str += 5;
         while (isspace(*str)) str++;
 
         // Yahoo reused Basic
-        if (memcmp("token=", str, 6) != 0) {
+        if (strlen(str) > 6 && memcmp("token=", str, 6) != 0) {
             g_base64_decode_inplace(str, &olen);
             char *colon = strchr(str, ':');
             if (colon)
@@ -197,8 +213,7 @@ moloch_http_parse_authorization(MolochSession_t *session, char *str)
     }
 }
 /******************************************************************************/
-int
-moloch_hp_cb_on_message_complete (http_parser *parser)
+LOCAL int moloch_hp_cb_on_message_complete (http_parser *parser)
 {
     HTTPInfo_t            *http = parser->data;
     MolochSession_t       *session = http->session;
@@ -213,24 +228,28 @@ moloch_hp_cb_on_message_complete (http_parser *parser)
     if (http->inBody & (1 << http->which)) {
         const char *md5 = g_checksum_get_string(http->checksum[http->which]);
         moloch_field_string_uw_add(md5Field, session, (char*)md5, 32, (gpointer)http->magicString[http->which], TRUE);
+        if (config.supportSha256) {
+            const char *sha256 = g_checksum_get_string(http->checksum[http->which+2]);
+            moloch_field_string_uw_add(sha256Field, session, (char*)sha256, 64, (gpointer)http->magicString[http->which], TRUE);
+        }
     }
 
     return 0;
 }
 
 /******************************************************************************/
-void
-http_add_value(MolochSession_t *session, HTTPInfo_t *http)
+/******************************************************************************/
+LOCAL void http_add_value(MolochSession_t *session, HTTPInfo_t *http)
 {
-    int                    pos  = http->pos[http->which];
-    char                  *s    = http->valueString[http->which]->str;
-    int                    l    = http->valueString[http->which]->len;
+    int                     pos  = http->pos[http->which];
+    char                    *s   = http->valueString[http->which]->str;
+    int                     l    = http->valueString[http->which]->len;
+    gboolean                c    = http->isValueStringLowerCase[http->which];
 
     while (isspace(*s)) {
         s++;
         l--;
     }
-
 
     switch (config.fields[pos]->type) {
     case MOLOCH_FIELD_TYPE_INT:
@@ -242,28 +261,28 @@ http_add_value(MolochSession_t *session, HTTPInfo_t *http)
     case MOLOCH_FIELD_TYPE_STR:
     case MOLOCH_FIELD_TYPE_STR_ARRAY:
     case MOLOCH_FIELD_TYPE_STR_HASH:
-        moloch_field_string_add(pos, session, s, l, TRUE);
+    case MOLOCH_FIELD_TYPE_STR_GHASH:
+        if (c == TRUE)
+            moloch_field_string_add_lower(pos, session, s, l);
+        else
+            moloch_field_string_add(pos, session, s, l, TRUE);
         break;
-    case MOLOCH_FIELD_TYPE_IP_HASH:
     case MOLOCH_FIELD_TYPE_IP_GHASH:
     {
         int i;
         gchar **parts = g_strsplit(http->valueString[http->which]->str, ",", 0);
 
         for (i = 0; parts[i]; i++) {
-            gchar *ip = parts[i];
-            while (*ip == ' ')
-                ip++;
+            moloch_field_ip_add_str(pos, session, parts[i]);
 
-            in_addr_t ia = inet_addr(ip);
+            /* Add back maybe
             if (ia == 0 || ia == 0xffffffff) {
                 moloch_session_add_tag(session, "http:bad-xff");
                 if (config.debug)
                     LOG("INFO - Didn't understand ip: %s %s %d", http->valueString[http->which]->str, ip, ia);
                 continue;
             }
-
-            moloch_field_int_add(pos, session, ia);
+            */
         }
 
         g_strfreev(parts);
@@ -271,13 +290,12 @@ http_add_value(MolochSession_t *session, HTTPInfo_t *http)
     }
     } /* SWITCH */
 
-
     g_string_truncate(http->valueString[http->which], 0);
     http->pos[http->which] = 0;
+    http->isValueStringLowerCase[http->which] = FALSE;
 }
 /******************************************************************************/
-int
-moloch_hp_cb_on_header_field (http_parser *parser, const char *at, size_t length)
+LOCAL int moloch_hp_cb_on_header_field (http_parser *parser, const char *at, size_t length)
 {
     HTTPInfo_t            *http = parser->data;
     MolochSession_t       *session = http->session;
@@ -315,13 +333,11 @@ moloch_hp_cb_on_header_field (http_parser *parser, const char *at, size_t length
 }
 
 /******************************************************************************/
-int
-moloch_hp_cb_on_header_value (http_parser *parser, const char *at, size_t length)
+LOCAL int moloch_hp_cb_on_header_value (http_parser *parser, const char *at, size_t length)
 {
     HTTPInfo_t            *http = parser->data;
     MolochSession_t       *session = http->session;
-    char                   header[200];
-    MolochString_t        *hstring = 0;
+    MolochString_t        *hstring;
 
 #ifdef HTTPDEBUG
     LOG("HTTPDEBUG: which: %d value: %.*s", http->which, (int)length, at);
@@ -340,12 +356,24 @@ moloch_hp_cb_on_header_value (http_parser *parser, const char *at, size_t length
 
         http->pos[http->which] = (long)(hstring?hstring->uw:0);
 
-        snprintf(header, sizeof(header), "http:header:%s", lower);
-        g_free(lower);
+        if (http->pos[http->which] == 0) { // Header was not defined
+            if ((http->which == 0) && config.parseHTTPHeaderRequestAll) { // Header in request
+                moloch_field_string_add(headerReqField, session, lower, -1, TRUE);
+                http->pos[http->which] = (long) headerReqValue;
+                http->isValueStringLowerCase[http->which] = TRUE;
+            }
+            else if ((http->which == 1) && config.parseHTTPHeaderResponseAll) { // Header in response
+                moloch_field_string_add(headerResField, session, lower, -1, TRUE);
+                http->pos[http->which] = (long) headerResValue;
+                http->isValueStringLowerCase[http->which] = TRUE;
+            }
+        }
+
         if (http->which == http->urlWhich)
-            moloch_session_add_tag_type(session, tagsReqField, header);
+            moloch_field_string_add(tagsReqField, session, lower, -1, TRUE);
         else
-            moloch_session_add_tag_type(session, tagsResField, header);
+            moloch_field_string_add(tagsResField, session, lower, -1, TRUE);
+        g_free(lower);
     }
 
     moloch_plugins_cb_hp_ohv(session, parser, at, length);
@@ -354,8 +382,9 @@ moloch_hp_cb_on_header_value (http_parser *parser, const char *at, size_t length
     if (parser->method) {
         if (strcasecmp("host", http->header[http->which]) == 0) {
             if (!http->hostString)
-                http->hostString = g_string_new_len("//", 2);
-            g_string_append_len(http->hostString, at, length);
+                http->hostString = g_string_new_len(at, length);
+            else
+                g_string_append_len(http->hostString, at, length);
         } else if (strcasecmp("cookie", http->header[http->which]) == 0) {
             if (!http->cookieString)
                 http->cookieString = g_string_new_len(at, length);
@@ -379,17 +408,20 @@ moloch_hp_cb_on_header_value (http_parser *parser, const char *at, size_t length
     return 0;
 }
 /******************************************************************************/
-int
-moloch_hp_cb_on_headers_complete (http_parser *parser)
+LOCAL int moloch_hp_cb_on_headers_complete (http_parser *parser)
 {
     HTTPInfo_t            *http = parser->data;
     MolochSession_t       *session = http->session;
     char                   version[20];
 
-
 #ifdef HTTPDEBUG
     LOG("HTTPDEBUG: which: %d code: %d method: %d", http->which, parser->status_code, parser->method);
 #endif
+
+    if (parser->method == HTTP_CONNECT) {
+        http->isConnect = 1;
+        http->reclassify = 3;
+    }
 
     int len = snprintf(version, sizeof(version), "%d.%d", parser->http_major, parser->http_minor);
 
@@ -452,11 +484,11 @@ moloch_hp_cb_on_headers_complete (http_parser *parser)
 
     gboolean truncated = FALSE;
     if (http->urlString && http->hostString) {
-        char *colon = strchr(http->hostString->str+2, ':');
+        char *colon = strchr(http->hostString->str, ':');
         if (colon) {
-            moloch_field_string_add(hostField, session, http->hostString->str+2, colon - http->hostString->str-2, TRUE);
+            moloch_field_string_add(hostField, session, http->hostString->str, colon - http->hostString->str, TRUE);
         } else {
-            moloch_field_string_add(hostField, session, http->hostString->str+2, http->hostString->len-2, TRUE);
+            moloch_field_string_add(hostField, session, http->hostString->str, http->hostString->len, TRUE);
         }
 
         char *question = strchr(http->urlString->str, '?');
@@ -504,7 +536,7 @@ moloch_hp_cb_on_headers_complete (http_parser *parser)
         }
 
         if (http->urlString->str[0] != '/') {
-            char *result = strstr(http->urlString->str, http->hostString->str+2);
+            char *result = strstr(http->urlString->str, http->hostString->str);
 
             /* If the host header is in the first 8 bytes of url then just use the url */
             if (result && result - http->urlString->str <= 8) {
@@ -512,7 +544,8 @@ moloch_hp_cb_on_headers_complete (http_parser *parser)
                     truncated = TRUE;
                     g_string_truncate(http->urlString, MAX_URL_LENGTH);
                 }
-                moloch_field_string_add(urlsField, session, http->urlString->str, http->urlString->len, FALSE);
+                if (!moloch_field_string_add(urlsField, session, http->urlString->str, http->urlString->len, FALSE))
+                    g_free(http->urlString->str);
                 g_string_free(http->urlString, FALSE);
                 g_string_free(http->hostString, TRUE);
             } else {
@@ -524,7 +557,9 @@ moloch_hp_cb_on_headers_complete (http_parser *parser)
                     truncated = TRUE;
                     g_string_truncate(http->hostString, MAX_URL_LENGTH);
                 }
-                moloch_field_string_add(urlsField, session, http->hostString->str, http->hostString->len, FALSE);
+                if (!moloch_field_string_add(urlsField, session, http->hostString->str, http->hostString->len, FALSE))
+                    g_free(http->hostString->str);
+
                 g_string_free(http->urlString, TRUE);
                 g_string_free(http->hostString, FALSE);
             }
@@ -536,7 +571,8 @@ moloch_hp_cb_on_headers_complete (http_parser *parser)
                 truncated = TRUE;
                 g_string_truncate(http->hostString, MAX_URL_LENGTH);
             }
-            moloch_field_string_add(urlsField, session, http->hostString->str, http->hostString->len, FALSE);
+            if (!moloch_field_string_add(urlsField, session, http->hostString->str, http->hostString->len, FALSE))
+                g_free(http->hostString->str);
             g_string_free(http->urlString, TRUE);
             g_string_free(http->hostString, FALSE);
         }
@@ -549,16 +585,17 @@ moloch_hp_cb_on_headers_complete (http_parser *parser)
             truncated = TRUE;
             g_string_truncate(http->urlString, MAX_URL_LENGTH);
         }
-        moloch_field_string_add(urlsField, session, http->urlString->str, http->urlString->len, FALSE);
+        if (!moloch_field_string_add(urlsField, session, http->urlString->str, http->urlString->len, FALSE))
+                g_free(http->urlString->str);
         g_string_free(http->urlString, FALSE);
 
         http->urlString = NULL;
     } else if (http->hostString) {
-        char *colon = strchr(http->hostString->str+2, ':');
+        char *colon = strchr(http->hostString->str, ':');
         if (colon) {
-            moloch_field_string_add(hostField, session, http->hostString->str+2, colon - http->hostString->str-2, TRUE);
+            moloch_field_string_add(hostField, session, http->hostString->str, colon - http->hostString->str, TRUE);
         } else {
-            moloch_field_string_add(hostField, session, http->hostString->str+2, http->hostString->len-2, TRUE);
+            moloch_field_string_add(hostField, session, http->hostString->str, http->hostString->len, TRUE);
         }
 
         g_string_free(http->hostString, TRUE);
@@ -578,7 +615,7 @@ moloch_hp_cb_on_headers_complete (http_parser *parser)
 
 /*############################## SHARED ##############################*/
 /******************************************************************************/
-int http_parse(MolochSession_t *session, void *uw, const unsigned char *data, int remaining, int which)
+LOCAL int http_parse(MolochSession_t *session, void *uw, const unsigned char *data, int remaining, int which)
 {
     HTTPInfo_t            *http          = uw;
 
@@ -586,6 +623,19 @@ int http_parse(MolochSession_t *session, void *uw, const unsigned char *data, in
 #ifdef HTTPDEBUG
     LOG("HTTPDEBUG: enter %d - %d %.*s", http->which, remaining, remaining, data);
 #endif
+
+    if (http->isConnect) {
+        // Check if either side needs to be classified
+        if (http->reclassify & (1 << which)) {
+            http->reclassify &= ~(1 << which);
+            moloch_parsers_classify_tcp(session, data, remaining, which);
+        }
+        // Both sides have been reclassified, remove http parser
+        if (!http->reclassify) {
+            moloch_parsers_unregister(session, uw);
+        }
+        return 0;
+    }
 
     if ((http->wParsers & (1 << http->which)) == 0) {
         return 0;
@@ -598,7 +648,7 @@ int http_parse(MolochSession_t *session, void *uw, const unsigned char *data, in
 #endif
         if (len <= 0) {
             http->wParsers &= ~(1 << http->which);
-            if (http->wParsers) {
+            if (!http->wParsers) {
                 moloch_parsers_unregister(session, uw);
             }
             break;
@@ -629,7 +679,7 @@ void http_save(MolochSession_t UNUSED(*session), void *uw, int final)
 
 }
 /******************************************************************************/
-void http_free(MolochSession_t UNUSED(*session), void *uw)
+LOCAL void http_free(MolochSession_t UNUSED(*session), void *uw)
 {
     HTTPInfo_t            *http          = uw;
 
@@ -645,14 +695,17 @@ void http_free(MolochSession_t UNUSED(*session), void *uw)
         g_string_free(http->valueString[0], TRUE);
     if (http->valueString[1])
         g_string_free(http->valueString[1], TRUE);
-
     g_checksum_free(http->checksum[0]);
     g_checksum_free(http->checksum[1]);
+    if (config.supportSha256) {
+        g_checksum_free(http->checksum[2]);
+        g_checksum_free(http->checksum[3]);
+    }
 
     MOLOCH_TYPE_FREE(HTTPInfo_t, http);
 }
 /******************************************************************************/
-void http_classify(MolochSession_t *session, const unsigned char *UNUSED(data), int UNUSED(len), int UNUSED(which), void *UNUSED(uw))
+LOCAL void http_classify(MolochSession_t *session, const unsigned char *UNUSED(data), int UNUSED(len), int UNUSED(which), void *UNUSED(uw))
 {
     if (moloch_session_has_protocol(session, "http"))
         return;
@@ -663,6 +716,10 @@ void http_classify(MolochSession_t *session, const unsigned char *UNUSED(data), 
 
     http->checksum[0] = g_checksum_new(G_CHECKSUM_MD5);
     http->checksum[1] = g_checksum_new(G_CHECKSUM_MD5);
+    if (config.supportSha256) {
+        http->checksum[2] = g_checksum_new(G_CHECKSUM_SHA256);
+        http->checksum[3] = g_checksum_new(G_CHECKSUM_SHA256);
+    }
 
     http_parser_init(&http->parsers[0], HTTP_BOTH);
     http_parser_init(&http->parsers[1], HTTP_BOTH);
@@ -686,140 +743,203 @@ static const char *method_strings[] =
     };
 
     hostField = moloch_field_define("http", "lotermfield",
-        "host.http", "Hostname", "ho",
+        "host.http", "Hostname", "http.host",
         "HTTP host header field",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
         "aliases", "[\"http.host\"]",
         "category", "host",
-        NULL);
+        (char *)NULL);
 
-    urlsField = moloch_field_define("http", "textfield",
-        "http.uri", "URI", "us",
+    moloch_field_define("http", "lotextfield",
+        "host.http.tokens", "Hostname Tokens", "http.hostTokens",
+        "HTTP host Tokens header field",
+        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_FAKE,
+        "aliases", "[\"http.host.tokens\"]",
+        (char *)NULL);
+
+    urlsField = moloch_field_define("http", "termfield",
+        "http.uri", "URI", "http.uri",
         "URIs for request",
-        MOLOCH_FIELD_TYPE_STR_ARRAY, MOLOCH_FIELD_FLAG_CNT,
-        "rawField", "rawus",
+        MOLOCH_FIELD_TYPE_STR_HASH, MOLOCH_FIELD_FLAG_CNT,
         "category", "[\"url\",\"host\"]",
-        NULL);
+        (char *)NULL);
+
+    moloch_field_define("http", "lotextfield",
+        "http.uri.tokens", "URI Tokens", "http.uriTokens",
+        "URIs Tokens for request",
+        MOLOCH_FIELD_TYPE_STR_HASH, MOLOCH_FIELD_FLAG_FAKE,
+        (char *)NULL);
 
     xffField = moloch_field_define("http", "ip",
-        "ip.xff", "XFF IP", "xff",
+        "ip.xff", "XFF IP", "http.xffIp",
         "X-Forwarded-For Header",
-        MOLOCH_FIELD_TYPE_IP_GHASH, MOLOCH_FIELD_FLAG_SCNT | MOLOCH_FIELD_FLAG_IPPRE,
+        MOLOCH_FIELD_TYPE_IP_GHASH, MOLOCH_FIELD_FLAG_CNT | MOLOCH_FIELD_FLAG_IPPRE,
         "category", "ip",
-        NULL);
+        (char *)NULL);
 
-    uaField = moloch_field_define("http", "textfield",
-        "http.user-agent", "Useragent", "ua",
+    uaField = moloch_field_define("http", "termfield",
+        "http.user-agent", "Useragent", "http.useragent",
         "User-Agent Header",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
-        "rawField", "rawua",
-        NULL);
+        (char *)NULL);
+
+    moloch_field_define("http", "lotextfield",
+        "http.user-agent.tokens", "Useragent Tokens", "http.useragentTokens",
+        "User-Agent Header Tokens",
+        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_FAKE,
+        (char *)NULL);
 
     tagsReqField = moloch_field_define("http", "lotermfield",
-        "http.hasheader.src", "Has Src Header", "hh1",
+        "http.hasheader.src", "Has Src Header", "http.requestHeader",
         "Request has header present",
-        MOLOCH_FIELD_TYPE_INT_GHASH,  MOLOCH_FIELD_FLAG_CNT,
-        NULL);
+        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
+        (char *)NULL);
 
     tagsResField = moloch_field_define("http", "lotermfield",
-        "http.hasheader.dst", "Has Dst Header", "hh2",
+        "http.hasheader.dst", "Has Dst Header", "http.responseHeader",
         "Response has header present",
-        MOLOCH_FIELD_TYPE_INT_GHASH,  MOLOCH_FIELD_FLAG_CNT,
-        NULL);
+        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
+        (char *)NULL);
+
+    headerReqField = moloch_field_define("http", "lotermfield",
+        "http.header.request.field", "Request Header Fields", "http.requestHeaderField",
+        "Contains Request header fields",
+        MOLOCH_FIELD_TYPE_STR_ARRAY, MOLOCH_FIELD_FLAG_NODB,
+        (char *)NULL);
+
+    headerReqValue = moloch_field_define("http","lotermfield",
+        "http.hasheader.src.value", "Request Header Values", "http.requestHeaderValue",
+        "Contains request header values",
+        MOLOCH_FIELD_TYPE_STR_ARRAY, MOLOCH_FIELD_FLAG_CNT,
+        (char *)NULL);
+
+    headerResField = moloch_field_define("http","lotermfield",
+        "http.header.response.field","Response Header fields", "http.responseHeaderField",
+        "Contains response header fields",
+        MOLOCH_FIELD_TYPE_STR_ARRAY, MOLOCH_FIELD_FLAG_NODB,
+        (char *)NULL);
+
+    headerResValue = moloch_field_define("http","lotermfield",
+        "http.hasheader.dst.value", "Response Header Values", "http.responseHeaderValue",
+        "Contains response header values",
+        MOLOCH_FIELD_TYPE_STR_ARRAY, MOLOCH_FIELD_FLAG_CNT,
+        (char *)NULL);
 
     moloch_field_define("http", "lotermfield",
         "http.hasheader", "Has Src or Dst Header", "hhall",
         "Shorthand for http.hasheader.src or http.hasheader.dst",
         0,  MOLOCH_FIELD_FLAG_FAKE,
-        "regex", "^http.hasheader\\\\.(?:(?!\\\\.cnt$).)*$",
-        NULL);
+        "regex", "^http\\\\.hasheader\\\\.(?:(?!(cnt|value)$).)*$",
+        (char *)NULL);
+
+    moloch_field_define("http", "lotermfield",
+        "http.hasheader.value", "Has Value in Src or Dst Header", "hhvalueall",
+        "Shorthand for http.hasheader.src.value or http.hasheader.dst.value",
+        0,  MOLOCH_FIELD_FLAG_FAKE,
+        "regex", "^http\\\\.hasheader\\\\.(src|dst)\\\\.value$",
+        (char *)NULL);
 
     md5Field = moloch_field_define("http", "lotermfield",
-        "http.md5", "Body MD5", "hmd5",
+        "http.md5", "Body MD5", "http.md5",
         "MD5 of http body response",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
         "category", "md5",
-        NULL);
+        (char *)NULL);
+
+    if (config.supportSha256) {
+        sha256Field = moloch_field_define("http", "lotermfield",
+            "http.sha256", "Body SHA256", "http.sha256",
+            "SHA256 of http body response",
+            MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
+            "category", "sha256",
+            (char *)NULL);
+    }
 
     moloch_field_define("http", "termfield",
         "http.version", "Version", "httpversion",
         "HTTP version number",
         0, MOLOCH_FIELD_FLAG_FAKE,
         "regex", "^http.version.[a-z]+$",
-        NULL);
+        (char *)NULL);
 
     verReqField = moloch_field_define("http", "termfield",
-        "http.version.src", "Src Version", "hsver",
+        "http.version.src", "Src Version", "http.clientVersion",
         "Request HTTP version number",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
-        NULL);
+        (char *)NULL);
 
     verResField = moloch_field_define("http", "termfield",
-        "http.version.dst", "Dst Version", "hdver",
+        "http.version.dst", "Dst Version", "http.serverVersion",
         "Response HTTP version number",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
-        NULL);
+        (char *)NULL);
 
     pathField = moloch_field_define("http", "termfield",
-        "http.uri.path", "URI Path", "hpath",
+        "http.uri.path", "URI Path", "http.path",
         "Path portion of URI",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
-        NULL);
+        (char *)NULL);
 
     keyField = moloch_field_define("http", "termfield",
-        "http.uri.key", "QS Keys", "hkey",
+        "http.uri.key", "QS Keys", "http.key",
         "Keys from query string of URI",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
-        NULL);
+        (char *)NULL);
 
     valueField = moloch_field_define("http", "termfield",
-        "http.uri.value", "QS Values", "hval",
+        "http.uri.value", "QS Values", "http.value",
         "Values from query string of URI",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
-        NULL);
+        (char *)NULL);
 
     cookieKeyField = moloch_field_define("http", "termfield",
-        "http.cookie.key", "Cookie Keys", "hckey-term",
+        "http.cookie.key", "Cookie Keys", "http.cookieKey",
         "The keys to cookies sent up in requests",
-        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_COUNT,
-        NULL);
+        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
+        (char *)NULL);
 
     cookieValueField = moloch_field_define("http", "termfield",
-        "http.cookie.value", "Cookie Values", "hcval-term",
+        "http.cookie.value", "Cookie Values", "http.cookieValue",
         "The values to cookies sent up in requests",
-        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_COUNT,
-        NULL);
+        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
+        (char *)NULL);
 
     methodField = moloch_field_define("http", "termfield",
-        "http.method", "Request Method", "http.method-term",
+        "http.method", "Request Method", "http.method",
         "HTTP Request Method",
-        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_COUNT,
-        NULL);
+        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
+        (char *)NULL);
 
     magicField = moloch_field_define("http", "termfield",
-        "http.bodymagic", "Body Magic", "http.bodymagic-term",
+        "http.bodymagic", "Body Magic", "http.bodyMagic",
         "The content type of body determined by libfile/magic",
-        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_COUNT,
-        NULL);
+        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
+        (char *)NULL);
 
     userField = moloch_field_define("http", "termfield",
-        "http.user", "User", "huser-term",
+        "http.user", "User", "http.user",
         "HTTP Auth User",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
         "category", "user",
-        NULL);
+        (char *)NULL);
 
     atField = moloch_field_define("http", "lotermfield",
-        "http.authtype", "Auth Type", "hat-term",
+        "http.authtype", "Auth Type", "http.authType",
         "HTTP Auth Type",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
-        NULL);
+        (char *)NULL);
 
     statuscodeField = moloch_field_define("http", "integer",
         "http.statuscode", "Status Code", "http.statuscode",
         "Response HTTP numeric status code",
-        MOLOCH_FIELD_TYPE_INT_GHASH,  MOLOCH_FIELD_FLAG_COUNT,
-        NULL);
+        MOLOCH_FIELD_TYPE_INT_GHASH,  MOLOCH_FIELD_FLAG_CNT,
+        (char *)NULL);
+
+    reqBodyField = moloch_field_define("http", "termfield",
+        "http.reqbody", "Request Body", "http.requestBody",
+        "HTTP Request Body",
+        MOLOCH_FIELD_TYPE_STR_HASH, 0,
+        (char *)NULL);
 
     HASH_INIT(s_, httpReqHeaders, moloch_string_hash, moloch_string_cmp);
     HASH_INIT(s_, httpResHeaders, moloch_string_hash, moloch_string_cmp);
@@ -827,8 +947,8 @@ static const char *method_strings[] =
     moloch_config_add_header(&httpReqHeaders, "x-forwarded-for", xffField);
     moloch_config_add_header(&httpReqHeaders, "user-agent", uaField);
     moloch_config_add_header(&httpReqHeaders, "host", hostField);
-    moloch_config_load_header("headers-http-request", "http", "Request header ", "http.", "hdrs.hreq-", &httpReqHeaders, 0);
-    moloch_config_load_header("headers-http-response", "http", "Response header ", "http.", "hdrs.hres-", &httpResHeaders, 0);
+    moloch_config_load_header("headers-http-request", "http", "Request header ", "http.", "http.request-", &httpReqHeaders, 0);
+    moloch_config_load_header("headers-http-response", "http", "Response header ", "http.", "http.response-", &httpResHeaders, 0);
 
     int i;
     for (i = 0; method_strings[i]; i++) {
@@ -846,4 +966,3 @@ static const char *method_strings[] =
     parserSettings.on_header_field = moloch_hp_cb_on_header_field;
     parserSettings.on_header_value = moloch_hp_cb_on_header_value;
 }
-
