@@ -14,34 +14,33 @@
  */
 #include "moloch.h"
 
-static int userField;
-static int hostField;
-static int serviceField;
+LOCAL  int userField;
+LOCAL  int hostField;
+LOCAL  int serviceField;
 extern MolochConfig_t        config;
 
 // Lots of info from https://www.pythian.com/blog/repost-oracle-protocol/
 
 /******************************************************************************/
-char *oracle_get_item(const char *data, char *needle, int needle_len, int *len)
+LOCAL char *oracle_get_item(const unsigned char *data, char *needle, int needle_len, int *len)
 {
-    const char *start = data + data[27];
-    char *item, *paren;
+    const unsigned char *start = data + data[27];
 
-    item = g_strstr_len((char *)start, data[25], (gchar *)needle);
+    unsigned char *item = (unsigned char *)g_strstr_len((char *)start, data[25], (gchar *)needle);
     if (item) {
-        paren = g_strstr_len(item, data[25] - (item - start), ")");
+        unsigned char *paren = (unsigned char *)g_strstr_len((char *)item, data[25] - (item - start), ")");
         if (paren) {
             *len = (paren-item)-needle_len;
             if (*len == 0)
                 return NULL;
 
-            return g_ascii_strdown(item+needle_len, *len);
+            return g_ascii_strdown((char *)item+needle_len, *len);
         }
     }
     return NULL;
 }
 /******************************************************************************/
-void oracle_classify(MolochSession_t *session, const unsigned char *data, int len, int which, void *UNUSED(uw))
+LOCAL void oracle_classify(MolochSession_t *session, const unsigned char *data, int len, int which, void *UNUSED(uw))
 {
     if (which != 0 || len <= 27 || len != data[1] || (data[25] + data[27] != len))
         return;
@@ -49,17 +48,17 @@ void oracle_classify(MolochSession_t *session, const unsigned char *data, int le
     char *buf;  // can't be more then 1 byte big
     int  blen;
 
-    buf = oracle_get_item((const char *)data, "HOST=", 5, &blen);
+    buf = oracle_get_item(data, "HOST=", 5, &blen); // Already lowercases
     if (buf && !moloch_field_string_add(hostField, session, buf, blen, FALSE)) {
         g_free(buf);
     }
 
-    buf = oracle_get_item((const char *)data, "USER=", 5, &blen);
+    buf = oracle_get_item(data, "USER=", 5, &blen); // Already lowercases
     if (buf && !moloch_field_string_add(userField, session, buf, blen, FALSE)) {
         g_free(buf);
     }
 
-    buf = oracle_get_item((const char *)data, "SERVICE_NAME=", 13, &blen);
+    buf = oracle_get_item(data, "SERVICE_NAME=", 13, &blen); // Already lowercases
     if (buf && !moloch_field_string_add(serviceField, session, buf, blen, FALSE)) {
         g_free(buf);
     }
@@ -72,22 +71,31 @@ void moloch_parser_init()
     moloch_parsers_classifier_register_tcp("oracle", NULL, 2, (unsigned char*)"\x00\x00\x01\x00\x00\x00", 6, oracle_classify);
 
     userField = moloch_field_define("oracle", "lotermfield",
-        "oracle.user", "User", "oracle.user-term",
+        "oracle.user", "User", "oracle.user",
         "Oracle User",
         MOLOCH_FIELD_TYPE_STR,  MOLOCH_FIELD_FLAG_LINKED_SESSIONS,
         "category", "user",
-        NULL);
+        (char *)NULL);
 
     hostField = moloch_field_define("oracle", "lotermfield",
-        "oracle.host", "Host", "oracle.host-term",
+        "oracle.host", "Host", "oracle.host",
         "Oracle Host",
         MOLOCH_FIELD_TYPE_STR,  MOLOCH_FIELD_FLAG_LINKED_SESSIONS,
-        NULL);
+        "category", "host",
+        "aliases", "[\"host.oracle\"]",
+        (char *)NULL);
+
+    moloch_field_define("oracle", "lotextfield",
+        "oracle.host.tokens", "Hostname Tokens", "oracle.hostTokens",
+        "Oracle Hostname Tokens",
+        MOLOCH_FIELD_TYPE_STR,  MOLOCH_FIELD_FLAG_FAKE,
+        "aliases", "[\"host.oracle.tokens\"]",
+        (char *)NULL);
 
     serviceField = moloch_field_define("oracle", "lotermfield",
-        "oracle.service", "Service", "oracle.service-term",
+        "oracle.service", "Service", "oracle.service",
         "Oracle Service",
         MOLOCH_FIELD_TYPE_STR,  MOLOCH_FIELD_FLAG_LINKED_SESSIONS,
-        NULL);
+        (char *)NULL);
 }
 
