@@ -16,21 +16,19 @@
 #include <arpa/inet.h>
 
 extern MolochConfig_t        config;
-static int userField;
-static int macField;
-static int endpointIpField;
-static int framedIpField;
+LOCAL  int userField;
+LOCAL  int macField;
+LOCAL  int endpointIpField;
+LOCAL  int framedIpField;
 
 /******************************************************************************/
-int radius_udp_parser(MolochSession_t *session, void *UNUSED(uw), const unsigned char *data, int len, int UNUSED(which))
+LOCAL int radius_udp_parser(MolochSession_t *session, void *UNUSED(uw), const unsigned char *data, int len, int UNUSED(which))
 {
     BSB bsb;
 
     BSB_INIT(bsb, data, len);
-    unsigned char code;
 
-    BSB_IMPORT_u08(bsb, code);
-    BSB_IMPORT_skip(bsb, 19);
+    BSB_IMPORT_skip(bsb, 20);
 
     unsigned char type = 0, length = 0;
     unsigned char *value;
@@ -41,6 +39,8 @@ int radius_udp_parser(MolochSession_t *session, void *UNUSED(uw), const unsigned
     while (BSB_REMAINING(bsb) > 2) {
         BSB_IMPORT_u08(bsb, type);
         BSB_IMPORT_u08(bsb, length);
+        if (length < 3)
+            break;
         length -= 2; // length includes the type/length
         BSB_IMPORT_ptr(bsb, value, length);
         if (BSB_IS_ERROR(bsb)) {
@@ -54,8 +54,10 @@ int radius_udp_parser(MolochSession_t *session, void *UNUSED(uw), const unsigned
             LOG("NAS-IP-Address: %d %d %u.%u.%u.%u", type, length, value[0], value[1], value[2], value[3]);
             break;*/
         case 8:
+            if (length != 4)
+                return 0;
             memcpy(&in.s_addr, value, 4);
-            moloch_field_int_add(framedIpField, session, in.s_addr);
+            moloch_field_ip4_add(framedIpField, session, in.s_addr);
             break;
         case 31:
             if (length == 12) {
@@ -76,8 +78,7 @@ int radius_udp_parser(MolochSession_t *session, void *UNUSED(uw), const unsigned
         case 66:
             memcpy(str, value, length);
             str[length] = 0;
-            inet_aton(str, &in);
-            moloch_field_int_add(endpointIpField, session, in.s_addr);
+            moloch_field_ip_add_str(endpointIpField, session, str);
             break;
 
 /*        default:
@@ -87,9 +88,9 @@ int radius_udp_parser(MolochSession_t *session, void *UNUSED(uw), const unsigned
     return 0;
 }
 /******************************************************************************/
-void radius_udp_classify(MolochSession_t *session, const unsigned char *UNUSED(data), int len, int UNUSED(which), void *UNUSED(uw))
+LOCAL void radius_udp_classify(MolochSession_t *session, const unsigned char *UNUSED(data), int len, int UNUSED(which), void *UNUSED(uw))
 {
-    if (len != ((data[2] << 8) | data[3])) {
+    if (len < 4 || len != ((data[2] << 8) | data[3])) {
         return;
     }
 
@@ -105,29 +106,29 @@ void radius_udp_classify(MolochSession_t *session, const unsigned char *UNUSED(d
 void moloch_parser_init()
 {
     userField = moloch_field_define("radius", "termfield",
-        "radius.user", "User", "radius.user-term",
+        "radius.user", "User", "radius.user",
         "RADIUS user",
-        MOLOCH_FIELD_TYPE_STR_HASH,     0, 
+        MOLOCH_FIELD_TYPE_STR_GHASH,  0,
         "category", "user",
-        NULL);
+        (char *)NULL);
 
     macField = moloch_field_define("radius", "lotermfield",
-        "radius.mac", "MAC", "radius.mac-term",
+        "radius.mac", "MAC", "radius.mac",
         "Radius Mac",
-        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_COUNT,
-        NULL); 
+        MOLOCH_FIELD_TYPE_STR_GHASH,  MOLOCH_FIELD_FLAG_CNT,
+        (char *)NULL);
 
     endpointIpField = moloch_field_define("radius", "ip",
-        "radius.endpoint-ip", "Endpoint IP", "radius.eip",
+        "radius.endpoint-ip", "Endpoint IP", "radius.endpointIp",
         "Radius endpoint ip addresses for session",
-        MOLOCH_FIELD_TYPE_IP_GHASH,  MOLOCH_FIELD_FLAG_COUNT,
-        NULL);
+        MOLOCH_FIELD_TYPE_IP_GHASH,  MOLOCH_FIELD_FLAG_CNT,
+        (char *)NULL);
 
     framedIpField = moloch_field_define("radius", "ip",
-        "radius.framed-ip", "Framed IP", "radius.fip",
+        "radius.framed-ip", "Framed IP", "radius.framedIp",
         "Radius framed ip addresses for session",
-        MOLOCH_FIELD_TYPE_IP_GHASH,  MOLOCH_FIELD_FLAG_COUNT,
-        NULL);
+        MOLOCH_FIELD_TYPE_IP_GHASH,  MOLOCH_FIELD_FLAG_CNT,
+        (char *)NULL);
 
 
     moloch_parsers_classifier_register_udp("radius", NULL, 0, (const unsigned char *)"\x01", 1, radius_udp_classify);
