@@ -18,15 +18,15 @@
 
 extern MolochConfig_t   config;
 
-static int domainField;
-static int userField;
-static int hostField;
-static int osField;
-static int verField;
-static int fnField;
-static int shareField;
+LOCAL  int domainField;
+LOCAL  int userField;
+LOCAL  int hostField;
+LOCAL  int osField;
+LOCAL  int verField;
+LOCAL  int fnField;
+LOCAL  int shareField;
 
-#define MAX_SMB_BUFFER 4096
+#define MAX_SMB_BUFFER 8192
 typedef struct {
     char               buf[2][MAX_SMB_BUFFER];
     uint32_t           remlen[2];
@@ -57,15 +57,19 @@ typedef struct {
 #define SMB2_FLAGS_SERVER_TO_REDIR 0x00000001
 
 /******************************************************************************/
-void smb_add_string(MolochSession_t *session, int field, char *buf, int len, int useunicode)
+LOCAL void smb_add_string(MolochSession_t *session, int field, char *buf, int len, int useunicode)
 {
+    if (len == 0)
+        return;
+
     GError *error = 0;
     gsize bread, bwritten;
 
     if (useunicode) {
         char *out = g_convert(buf, len, "utf-8", "ucs-2le", &bread, &bwritten, &error);
         if (error) {
-            LOG("ERROR %s", error->message);
+            if (config.debug)
+                LOG("ERROR %s", error->message);
             g_error_free(error);
         } else {
             if (!moloch_field_string_add(field, session, out, -1, FALSE)) {
@@ -78,7 +82,7 @@ void smb_add_string(MolochSession_t *session, int field, char *buf, int len, int
 }
 /******************************************************************************/
 // 2.2.13 AUTHENTICATE_MESSAGE from  http://download.microsoft.com/download/9/5/E/95EF66AF-9026-4BB0-A41D-A4F81802D92C/[MS-NLMP].pdf
-void smb_security_blob(MolochSession_t *session, unsigned char *data, int len)
+LOCAL void smb_security_blob(MolochSession_t *session, unsigned char *data, int len)
 {
     BSB bsb;
 
@@ -104,7 +108,7 @@ void smb_security_blob(MolochSession_t *session, unsigned char *data, int len)
     BSB_INIT(bsb, value, alen);
     value = moloch_parsers_asn_get_tlv(&bsb, &apc, &atag, &alen);
 
-    if (atag != 4 || memcmp("NTLMSSP", value, 7) != 0)
+    if (atag != 4 || alen < 7 || memcmp("NTLMSSP", value, 7) != 0)
         return;
 
     /* Woot, have the part we need to decode */
@@ -144,20 +148,21 @@ void smb_security_blob(MolochSession_t *session, unsigned char *data, int len)
     }
 }
 /******************************************************************************/
-void smb1_str_null_split(char *buf, int len, char **out, int max)
+LOCAL void smb1_str_null_split(char *buf, int len, char **out, int max)
 {
-    memset(out, 0, max*sizeof(char *));
-    out[0] = buf;
+    memset(out, 0, max * sizeof(char *));
     int i, p;
-    for (i = 0, p=1; i < len-1 && p < max; i++) {
+    int start = 0;
+    for (i = 0, p = 0; i < len && p < max; i++) {
         if (buf[i] == 0) {
-            out[p] = buf + i + 1;
+            out[p] = buf + start;
+            start = i + 1;
             p++;
         }
     }
 }
 /******************************************************************************/
-void smb1_parse_osverdomain(MolochSession_t *session, char *buf, int len, int useunicode)
+LOCAL void smb1_parse_osverdomain(MolochSession_t *session, char *buf, int len, int useunicode)
 {
     char        *out;
     gsize        bread, bwritten;
@@ -171,7 +176,8 @@ void smb1_parse_osverdomain(MolochSession_t *session, char *buf, int len, int us
     }
 
     if (error) {
-        LOG("ERROR %s", error->message);
+        if (config.debug)
+            LOG("ERROR %s", error->message);
         g_error_free(error);
         return;
     }
@@ -191,7 +197,7 @@ void smb1_parse_osverdomain(MolochSession_t *session, char *buf, int len, int us
     }
 }
 /******************************************************************************/
-void smb1_parse_userdomainosver(MolochSession_t *session, char *buf, int len, int useunicode)
+LOCAL void smb1_parse_userdomainosver(MolochSession_t *session, char *buf, int len, int useunicode)
 {
     char        *out;
     gsize        bread, bwritten;
@@ -205,7 +211,8 @@ void smb1_parse_userdomainosver(MolochSession_t *session, char *buf, int len, in
     }
 
     if (error) {
-        LOG("ERROR %s", error->message);
+        if (config.debug)
+            LOG("ERROR %s", error->message);
         g_error_free(error);
         return;
     }
@@ -227,19 +234,16 @@ void smb1_parse_userdomainosver(MolochSession_t *session, char *buf, int len, in
     }
 }
 /******************************************************************************/
-int smb1_parse(MolochSession_t *session, SMBInfo_t *smb, BSB *bsb, char *state, uint32_t *remlen, int which)
+LOCAL int smb1_parse(MolochSession_t *session, SMBInfo_t *smb, BSB *bsb, char *state, uint32_t *remlen, int which)
 {
     unsigned char *start = BSB_WORK_PTR(*bsb);
-    unsigned char cmd    = 0;
-
-    int          offset;
 
     switch (*state) {
     case SMB_SMBHEADER: {
-        unsigned char flags = 0;
+        uint8_t  cmd    = 0;
+        uint8_t  flags = 0;
         if (BSB_REMAINING(*bsb) < 32) {
             return 1;
-            break;
         }
         BSB_IMPORT_skip(*bsb, 4);
         BSB_IMPORT_u08(*bsb, cmd);
@@ -271,7 +275,7 @@ int smb1_parse(MolochSession_t *session, SMBInfo_t *smb, BSB *bsb, char *state, 
             *state = SMB_SKIP;
         }
 #ifdef SMBDEBUG
-        LOG("%d cmd: %x flags2: %x newstate: %d remlen: %d", which, cmd, smb->flags2[which], *state, *remlen);
+        LOG("%d cmd: %x flags2: %x newstate: %d remlen: %u", which, cmd, smb->flags2[which], *state, *remlen);
 #endif
         break;
     }
@@ -294,6 +298,8 @@ int smb1_parse(MolochSession_t *session, SMBInfo_t *smb, BSB *bsb, char *state, 
         int wordcount = 0;
         BSB_IMPORT_u08(*bsb, wordcount);
         BSB_IMPORT_skip(*bsb, wordcount*2+3);
+        if (BSB_IS_ERROR(*bsb))
+            return 1;
         smb_add_string(session, fnField, (char*)BSB_WORK_PTR(*bsb), BSB_REMAINING(*bsb), smb->flags2[which] & SMB1_FLAGS2_UNICODE);
         *state = SMB_SKIP;
         break;
@@ -308,6 +314,10 @@ int smb1_parse(MolochSession_t *session, SMBInfo_t *smb, BSB *bsb, char *state, 
         BSB_IMPORT_skip(*bsb, 2 + passlength);
 
         int offset = ((BSB_WORK_PTR(*bsb) - start) % 2 == 0)?2:1;
+
+        if (BSB_IS_ERROR(*bsb) || offset > BSB_REMAINING(*bsb)) {
+            return 1;
+        }
         smb_add_string(session, shareField, (char*)BSB_WORK_PTR(*bsb)+offset, BSB_REMAINING(*bsb)-offset, smb->flags2[which] & SMB1_FLAGS2_UNICODE);
         *state = SMB_SKIP;
         break;
@@ -315,6 +325,7 @@ int smb1_parse(MolochSession_t *session, SMBInfo_t *smb, BSB *bsb, char *state, 
 
     case SMB1_SETUP_ANDX: { // http://msdn.microsoft.com/en-us/library/ee441849.aspx
         if (BSB_REMAINING(*bsb) < *remlen) {
+            BSB_SET_ERROR(*bsb);
             return 1;
         }
         int wordcount = 0;
@@ -328,10 +339,14 @@ int smb1_parse(MolochSession_t *session, SMBInfo_t *smb, BSB *bsb, char *state, 
 
             BSB_IMPORT_skip(*bsb, 10);
 
+            if (securitylen > BSB_REMAINING(*bsb)) {
+                BSB_SET_ERROR(*bsb);
+                return 1;
+            }
             smb_security_blob(session, BSB_WORK_PTR(*bsb), securitylen);
             BSB_IMPORT_skip(*bsb, securitylen);
 
-            offset = ((BSB_WORK_PTR(*bsb) - start) % 2 == 0)?0:1;
+            int offset = ((BSB_WORK_PTR(*bsb) - start) % 2 == 0)?0:1;
             BSB_IMPORT_skip(*bsb, offset);
 
             if (!BSB_IS_ERROR(*bsb)) {
@@ -347,7 +362,7 @@ int smb1_parse(MolochSession_t *session, SMBInfo_t *smb, BSB *bsb, char *state, 
 
             BSB_IMPORT_skip(*bsb, 10 + ansipw + upw);
 
-            offset = ((BSB_WORK_PTR(*bsb) - start) % 2 == 0)?0:1;
+            int offset = ((BSB_WORK_PTR(*bsb) - start) % 2 == 0)?0:1;
             BSB_IMPORT_skip(*bsb, offset);
 
             if (!BSB_IS_ERROR(*bsb)) {
@@ -364,7 +379,7 @@ int smb1_parse(MolochSession_t *session, SMBInfo_t *smb, BSB *bsb, char *state, 
     return 0;
 }
 /******************************************************************************/
-int smb2_parse(MolochSession_t *session, SMBInfo_t *UNUSED(smb), BSB *bsb, char *state, uint32_t *remlen, int UNUSED(which))
+LOCAL int smb2_parse(MolochSession_t *session, SMBInfo_t *UNUSED(smb), BSB *bsb, char *state, uint32_t *remlen, int UNUSED(which))
 {
     unsigned char *start = BSB_WORK_PTR(*bsb);
 
@@ -375,7 +390,6 @@ int smb2_parse(MolochSession_t *session, SMBInfo_t *UNUSED(smb), BSB *bsb, char 
 
         if (BSB_REMAINING(*bsb) < 64) {
             return 1;
-            break;
         }
         BSB_IMPORT_skip(*bsb, 12);
         BSB_LIMPORT_u16(*bsb, cmd);
@@ -398,7 +412,7 @@ int smb2_parse(MolochSession_t *session, SMBInfo_t *UNUSED(smb), BSB *bsb, char 
             *state = SMB_SKIP;
         }
 #ifdef SMBDEBUG
-        LOG("%d cmd: %x flags: %x newstate: %d remlen: %d", which, cmd, flags, *state, *remlen);
+        LOG("%d cmd: %x flags: %x newstate: %d remlen: %u", which, cmd, flags, *state, *remlen);
 #endif
         *remlen -= (BSB_WORK_PTR(*bsb) - start);
         break;
@@ -460,7 +474,7 @@ int smb2_parse(MolochSession_t *session, SMBInfo_t *UNUSED(smb), BSB *bsb, char 
     return 0;
 }
 /******************************************************************************/
-int smb_parser(MolochSession_t *session, void *uw, const unsigned char *data, int remaining, int which)
+LOCAL int smb_parser(MolochSession_t *session, void *uw, const unsigned char *data, int remaining, int which)
 {
     SMBInfo_t            *smb          = uw;
     char                 *state        = &smb->state[which];
@@ -469,7 +483,7 @@ int smb_parser(MolochSession_t *session, void *uw, const unsigned char *data, in
     uint32_t             *remlen       = &smb->remlen[which];
 
 #ifdef SMBDEBUG
-    LOG("ENTER: remaining: %d state: %d buflen: %d remlen: %d", remaining, *state, *buflen, *remlen);
+    LOG("ENTER: remaining: %d state: %d buflen: %d remlen: %u", remaining, *state, *buflen, *remlen);
 #endif
 
     while (remaining > 0) {
@@ -490,14 +504,16 @@ int smb_parser(MolochSession_t *session, void *uw, const unsigned char *data, in
         }
 
         if (*state != SMB_SKIP && *remlen > MAX_SMB_BUFFER) {
-            LOG("ERROR - Not enough room for SMB packet %d", *remlen);
+#ifndef FUZZLOCH
+            LOG("ERROR - Not enough room for SMB packet %u", *remlen);
+#endif
             moloch_parsers_unregister(session, smb);
             return 0;
         }
 
         while (!done && BSB_REMAINING(bsb) > 0) {
 #ifdef SMBDEBUG
-            LOG(" S: bsbremaining: %ld remaining: %d state: %d buflen: %d remlen: %d done: %d", BSB_REMAINING(bsb), remaining, *state, *buflen, *remlen, done);
+            LOG(" S: bsbremaining: %ld remaining: %d state: %d buflen: %d remlen: %u done: %d", BSB_REMAINING(bsb), remaining, *state, *buflen, *remlen, done);
 #endif
             switch (*state) {
             case SMB_NETBIOS:
@@ -531,7 +547,7 @@ int smb_parser(MolochSession_t *session, void *uw, const unsigned char *data, in
             }
 
 #ifdef SMBDEBUG
-            LOG(" E: bsbremaining: %ld remaining: %d state: %d buflen: %d remlen: %d done: %d", BSB_REMAINING(bsb), remaining, *state, *buflen, *remlen, done);
+            LOG(" E: bsbremaining: %ld remaining: %d state: %d buflen: %d remlen: %u done: %d", BSB_REMAINING(bsb), remaining, *state, *buflen, *remlen, done);
 #endif
         }
 
@@ -542,7 +558,7 @@ int smb_parser(MolochSession_t *session, void *uw, const unsigned char *data, in
 
         if (BSB_REMAINING(bsb) > 0 && BSB_WORK_PTR(bsb) != (unsigned char *)buf) {
 #ifdef SMBDEBUG
-            LOG("  Moving data %ld %s", BSB_REMAINING(bsb), moloch_session_id_string(session->protocol, session->addr1, session->port1, session->addr2, session->port2));
+            //LOG("  Moving data %ld %s", BSB_REMAINING(bsb), moloch_session_id_string(session->protocol, session->addr1, session->port1, session->addr2, session->port2));
 #endif
             if (BSB_REMAINING(bsb) > MAX_SMB_BUFFER) {
                 LOG("ERROR - Not enough room for SMB packet %ld", BSB_REMAINING(bsb));
@@ -556,14 +572,14 @@ int smb_parser(MolochSession_t *session, void *uw, const unsigned char *data, in
     return 0;
 }
 /******************************************************************************/
-void smb_free(MolochSession_t UNUSED(*session), void *uw)
+LOCAL void smb_free(MolochSession_t UNUSED(*session), void *uw)
 {
     SMBInfo_t            *smb          = uw;
 
     MOLOCH_TYPE_FREE(SMBInfo_t, smb);
 }
 /******************************************************************************/
-void smb_classify(MolochSession_t *session, const unsigned char *data, int UNUSED(len), int UNUSED(which), void *UNUSED(uw))
+LOCAL void smb_classify(MolochSession_t *session, const unsigned char *data, int UNUSED(len), int UNUSED(which), void *UNUSED(uw))
 {
     if (data[4] != 0xff && data[4] != 0xfe)
         return;
@@ -581,48 +597,56 @@ void smb_classify(MolochSession_t *session, const unsigned char *data, int UNUSE
 void moloch_parser_init()
 {
     shareField =moloch_field_define("smb", "termfield",
-        "smb.share", "Share", "smbsh",
+        "smb.share", "Share", "smb.share",
         "SMB shares connected to",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
-        NULL);
+        (char *)NULL);
 
     fnField = moloch_field_define("smb", "termfield",
-        "smb.fn", "Filename", "smbfn",
+        "smb.fn", "Filename", "smb.filename",
         "SMB files opened, created, deleted",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
-        NULL);
+        (char *)NULL);
 
     osField = moloch_field_define("smb", "termfield",
-        "smb.os", "OS", "smbos",
+        "smb.os", "OS", "smb.os",
         "SMB OS information",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
-        NULL);
+        (char *)NULL);
 
     domainField = moloch_field_define("smb", "termfield",
-        "smb.domain", "Domain", "smbdm",
+        "smb.domain", "Domain", "smb.domain",
         "SMB domain",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
-        NULL);
+        (char *)NULL);
 
     verField = moloch_field_define("smb", "termfield",
-        "smb.ver", "Version", "smbver",
+        "smb.ver", "Version", "smb.version",
         "SMB Version information",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
-        NULL);
+        (char *)NULL);
 
     userField = moloch_field_define("smb", "termfield",
-        "smb.user", "User", "smbuser",
+        "smb.user", "User", "smb.user",
         "SMB User",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
         "category", "user",
-        NULL);
+        (char *)NULL);
 
     hostField = moloch_field_define("smb", "termfield",
-        "host.smb", "Hostname", "smbho",
+        "host.smb", "Hostname", "smb.host",
         "SMB Host name",
         MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_CNT,
         "category", "host",
-        "aliases", "[\"smb.host\"]", NULL);
+        "aliases", "[\"smb.host\"]",
+        (char *)NULL);
+
+    moloch_field_define("smb", "lotextfield",
+        "host.smb", "Hostname Tokens", "smb.hostTokens",
+        "SMB Host Tokens",
+        MOLOCH_FIELD_TYPE_STR_HASH,  MOLOCH_FIELD_FLAG_FAKE,
+        "aliases", "[\"smb.host.tokens\"]",
+        (char *)NULL);
 
     if (config.parseSMB) {
         moloch_parsers_classifier_register_tcp("smb", NULL, 5, (unsigned char*)"SMB", 3, smb_classify);
