@@ -27,17 +27,17 @@ typedef struct socksinfo {
     uint8_t   state5[2];
 } SocksInfo_t;
 
-static int ipField;
-static int portField;
-static int userField;
-static int hostField;
+LOCAL  int ipField;
+LOCAL  int portField;
+LOCAL  int userField;
+LOCAL  int hostField;
 
 //#define SOCKSDEBUG
 
 /******************************************************************************/
 #define SOCKS4_STATE_REPLY        0
 #define SOCKS4_STATE_DATA         1
-int socks4_parser(MolochSession_t *session, void *uw, const unsigned char *data, int remaining, int which)
+LOCAL int socks4_parser(MolochSession_t *session, void *uw, const unsigned char *data, int remaining, int which)
 {
     SocksInfo_t            *socks          = uw;
 
@@ -47,7 +47,7 @@ int socks4_parser(MolochSession_t *session, void *uw, const unsigned char *data,
             return 0;
         if (remaining >= 8 && data[0] == 0 && data[1] >= 0x5a && data[1] <= 0x5d) {
             if (socks->ip)
-                moloch_field_int_add(ipField, session, socks->ip);
+                moloch_field_ip4_add(ipField, session, socks->ip);
             moloch_field_int_add(portField, session, socks->port);
             moloch_session_add_protocol(session, "socks");
 
@@ -86,7 +86,7 @@ int socks4_parser(MolochSession_t *session, void *uw, const unsigned char *data,
 #define SOCKS5_STATE_CONN_REQUEST   5
 #define SOCKS5_STATE_CONN_REPLY     6
 #define SOCKS5_STATE_CONN_DATA      7
-int socks5_parser(MolochSession_t *session, void *uw, const unsigned char *data, int remaining, int which)
+LOCAL int socks5_parser(MolochSession_t *session, void *uw, const unsigned char *data, int remaining, int which)
 {
     SocksInfo_t            *socks          = uw;
     int                     consumed;
@@ -96,6 +96,11 @@ int socks5_parser(MolochSession_t *session, void *uw, const unsigned char *data,
 
     switch(socks->state5[which]) {
     case SOCKS5_STATE_VER_REQUEST:
+        if (remaining < 3) {
+            moloch_parsers_unregister(session, uw);
+            return 0;
+        }
+
         if (data[2] == 0) {
             socks->state5[which] = SOCKS5_STATE_CONN_REQUEST;
         } else {
@@ -128,7 +133,7 @@ int socks5_parser(MolochSession_t *session, void *uw, const unsigned char *data,
 
         return 2;
     case SOCKS5_STATE_USER_REQUEST:
-        if ((2 + data[1] > (int)remaining) || (2 + data[1] + 1 + data[data[1]+2]  > (int)remaining)) {
+        if (remaining < 2 || (3 + data[1] > (int)remaining) || (2 + data[1] + 1 + data[data[1]+2]  > (int)remaining)) {
             moloch_parsers_unregister(session, uw);
             return 0;
         }
@@ -150,15 +155,13 @@ int socks5_parser(MolochSession_t *session, void *uw, const unsigned char *data,
         if (data[3] == 1) { // IPV4
             socks->port = (data[8]&0xff) << 8 | (data[9]&0xff);
             memcpy(&socks->ip, data+4, 4);
-            moloch_field_int_add(ipField, session, socks->ip);
+            moloch_field_ip4_add(ipField, session, socks->ip);
             moloch_field_int_add(portField, session, socks->port);
             consumed = 4 + 4 + 2;
         } else if (data[3] == 3) { // Domain Name
             socks->port = (data[5+data[4]]&0xff) << 8 | (data[6+data[4]]&0xff);
-            char *lower = g_ascii_strdown((char*)data+5, data[4]);
-            if (!moloch_field_string_add(hostField, session, lower, data[4], FALSE)) {
-                g_free(lower);
-            }
+
+            moloch_field_string_add_lower(hostField, session, (char *)data+5, data[4]);
             moloch_field_int_add(portField, session, socks->port);
             consumed = 4 + 1 + data[4] + 2;
         } else if (data[3] == 4) { // IPV6
@@ -200,7 +203,7 @@ int socks5_parser(MolochSession_t *session, void *uw, const unsigned char *data,
 }
 
 /******************************************************************************/
-void socks_free(MolochSession_t UNUSED(*session), void *uw)
+LOCAL void socks_free(MolochSession_t UNUSED(*session), void *uw)
 {
     SocksInfo_t            *socks          = uw;
 
@@ -211,47 +214,48 @@ void socks_free(MolochSession_t UNUSED(*session), void *uw)
     MOLOCH_TYPE_FREE(SocksInfo_t, socks);
 }
 /******************************************************************************/
-void socks4_classify(MolochSession_t *session, const unsigned char *data, int len, int which, void *UNUSED(uw))
+LOCAL void socks4_classify(MolochSession_t *session, const unsigned char *data, int len, int which, void *UNUSED(uw))
 {
 #ifdef SOCKSDEBUG
     LOG("SOCKSDEBUG: enter %d %d", data[0], len);
 #endif
 
-    if (data[len-1] == 0)  {
-        SocksInfo_t *socks;
+    if (len < 8 || data[len-1] != 0)
+        return;
 
-        socks = MOLOCH_TYPE_ALLOC0(SocksInfo_t);
-        socks->which = which;
-        socks->port = (data[2]&0xff) << 8 | (data[3]&0xff);
-        if (data[4] == 0 && data[5] == 0 && data[6] == 0 && data[7] != 0) {
-            socks->ip = 0;
-        } else {
-            memcpy(&socks->ip, data+4, 4);
-        }
+    SocksInfo_t *socks;
 
-        int i;
-        for(i = 8; i < len && data[i]; i++);
-        if (i > 8 && i != len ) {
-            socks->user = g_strndup((char *)data+8, i-8);
-            socks->userlen = i - 8;
-        }
-
-        if (socks->ip == 0) {
-            i++;
-            int start;
-            for(start = i; i < len && data[i]; i++);
-            if (i > start && i != len ) {
-                socks->hostlen = i-start;
-                socks->host = g_ascii_strdown((char*)data+start, i-start);
-            }
-        }
-
-        moloch_parsers_register(session, socks4_parser, socks, socks_free);
+    socks = MOLOCH_TYPE_ALLOC0(SocksInfo_t);
+    socks->which = which;
+    socks->port = (data[2]&0xff) << 8 | (data[3]&0xff);
+    if (data[4] == 0 && data[5] == 0 && data[6] == 0 && data[7] != 0) {
+        socks->ip = 0;
+    } else {
+        memcpy(&socks->ip, data+4, 4);
     }
+
+    int i;
+    for(i = 8; i < len && data[i]; i++);
+    if (i > 8 && i != len ) {
+        socks->user = g_strndup((char *)data+8, i-8);
+        socks->userlen = i - 8;
+    }
+
+    if (socks->ip == 0) {
+        i++;
+        int start;
+        for(start = i; i < len && data[i]; i++);
+        if (i > start && i != len ) {
+            socks->hostlen = i-start;
+            socks->host = g_ascii_strdown((char*)data+start, i-start);
+        }
+    }
+
+    moloch_parsers_register(session, socks4_parser, socks, socks_free);
 }
 
 /******************************************************************************/
-void socks5_classify(MolochSession_t *session, const unsigned char *data, int len, int which, void *UNUSED(uw))
+LOCAL void socks5_classify(MolochSession_t *session, const unsigned char *data, int len, int which, void *UNUSED(uw))
 {
 #ifdef SOCKSDEBUG
     LOG("SOCKSDEBUG: enter %d %d", data[0], len);
@@ -272,37 +276,44 @@ void socks5_classify(MolochSession_t *session, const unsigned char *data, int le
 void moloch_parser_init()
 {
     ipField = moloch_field_define("socks", "ip",
-        "ip.socks", "IP", "socksip",
+        "ip.socks", "IP", "socks.ip",
         "SOCKS destination IP",
-        MOLOCH_FIELD_TYPE_IP, MOLOCH_FIELD_FLAG_IPPRE, 
-        "aliases", "[\"ip.socks\"]",
-        "portField", "sockspo", 
-        "category", "user",
-        NULL);
+        MOLOCH_FIELD_TYPE_IP, MOLOCH_FIELD_FLAG_IPPRE,
+        "aliases", "[\"socks.ip\"]",
+        "portField", "sockspo",
+        "portField2", "socks.port",
+        (char *)NULL);
 
     hostField = moloch_field_define("socks", "lotermfield",
-        "host.socks", "Host", "socksho",
+        "host.socks", "Host", "socks.host",
         "SOCKS destination host",
-        MOLOCH_FIELD_TYPE_STR,       0, 
-        "aliases", "[\"socks.host\"]", 
+        MOLOCH_FIELD_TYPE_STR,       0,
+        "aliases", "[\"socks.host\"]",
         "category", "host",
-        NULL);
+        (char *)NULL);
+
+    moloch_field_define("socks", "lotextfield",
+        "host.socks.tokens", "Hostname Tokens", "socks.hostTokens",
+        "SOCKS Hostname Tokens",
+        MOLOCH_FIELD_TYPE_STR,       MOLOCH_FIELD_FLAG_FAKE,
+        "aliases", "[\"socks.host.tokens\"]",
+        (char *)NULL);
 
     portField = moloch_field_define("socks", "integer",
-        "port.socks", "Port", "sockspo",
+        "port.socks", "Port", "socks.port",
         "SOCKS destination port",
-        MOLOCH_FIELD_TYPE_INT,       0, 
-        "aliases", "[\"socks.port\"]", 
+        MOLOCH_FIELD_TYPE_INT,       0,
+        "aliases", "[\"socks.port\"]",
         "category", "port",
-        NULL);
+        (char *)NULL);
 
     userField = moloch_field_define("socks", "termfield",
-        "socks.user", "User", "socksuser",
+        "socks.user", "User", "socks.user",
         "SOCKS authenticated user",
-        MOLOCH_FIELD_TYPE_STR,     0, 
-        "aliases", "[\"socksuser\"]", 
+        MOLOCH_FIELD_TYPE_STR,     0,
+        "aliases", "[\"socksuser\"]",
         "category", "user",
-        NULL);
+        (char *)NULL);
 
     moloch_parsers_classifier_register_tcp("socks5", NULL, 0, (unsigned char*)"\005", 1, socks5_classify);
     moloch_parsers_classifier_register_tcp("socks4", NULL, 0, (unsigned char*)"\004\000", 2, socks4_classify);
