@@ -20,11 +20,11 @@ typedef struct {
     int            pos[2];
 } TDSInfo_t;
 
-static int userField;
-
 extern MolochConfig_t        config;
+LOCAL  int userField;
+
 /******************************************************************************/
-int tds_parser(MolochSession_t *session, void *uw, const unsigned char *data, int remaining, int which)
+LOCAL int tds_parser(MolochSession_t *session, void *uw, const unsigned char *data, int remaining, int which)
 {
     TDSInfo_t *tds = uw;
 
@@ -46,21 +46,21 @@ int tds_parser(MolochSession_t *session, void *uw, const unsigned char *data, in
                 tds->data[0][480], tds->data[0] + 470
                 );
 #endif
-        moloch_field_string_add(userField, session, (const char *)tds->data[0] + 39, tds->data[0][69], TRUE);
+        moloch_field_string_add_lower(userField, session, (const char *)tds->data[0] + 39, tds->data[0][69]);
         moloch_parsers_unregister(session, uw);
     }
 
     return 0;
 }
 /******************************************************************************/
-void tds_free(MolochSession_t UNUSED(*session), void *uw)
+LOCAL void tds_free(MolochSession_t UNUSED(*session), void *uw)
 {
     TDSInfo_t            *tds          = uw;
 
     MOLOCH_TYPE_FREE(TDSInfo_t, tds);
 }
 /******************************************************************************/
-void tds_classify(MolochSession_t *session, const unsigned char *UNUSED(data), int len, int which, void *UNUSED(uw))
+LOCAL void tds_classify(MolochSession_t *session, const unsigned char *UNUSED(data), int len, int which, void *UNUSED(uw))
 {
     if (which != 0 || len < 512 || moloch_session_has_protocol(session, "tds"))
         return;
