@@ -58,16 +58,21 @@ DAQ_Verdict reader_daq_packet_cb(void *batch, const DAQ_PktHdr_t *h, const uint8
     packet->pkt           = (u_char *)data;
     packet->ts            = h->ts;
     packet->pktlen        = h->pktlen;
+    packet->readerPos     = ((MolochPacketBatch_t *)batch)->readerPos;
 
     moloch_packet_batch((MolochPacketBatch_t *)batch, packet);
     return DAQ_VERDICT_PASS;
 }
 /******************************************************************************/
-static void *reader_daq_thread(gpointer handle)
+LOCAL void *reader_daq_thread(gpointer posv)
 {
+    long pos = (long)posv;
+    gpointer handle = handles[pos];
+
+    MolochPacketBatch_t   batch;
+    moloch_packet_batch_init(&batch);
+    batch.readerPos = pos;
     while (1) {
-        MolochPacketBatch_t   batch;
-        moloch_packet_batch_init(&batch);
         int r = daq_acquire(module, handle, 10000, reader_daq_packet_cb, &batch);
         moloch_packet_batch_flush(&batch);
 
@@ -106,7 +111,7 @@ void reader_daq_start() {
 
         char name[100];
         snprintf(name, sizeof(name), "moloch-daq%d", i);
-        g_thread_new(name, &reader_daq_thread, NULL);
+        g_thread_unref(g_thread_new(name, &reader_daq_thread, NULL));
     }
 }
 /******************************************************************************/
