@@ -12,20 +12,8 @@
 #include <arpa/inet.h>
 #include "patricia.h"
 
-/*
- * prefix_tochar convert prefix information to bytes 
- */
-u_char         *
-prefix_tochar(prefix_t * prefix)
-{
-    if (prefix == NULL)
-        return (NULL);
-
-    return ((u_char *) & prefix->add.sin);
-}
-
 static inline int
-comp_with_mask(void *addr, void *dest, u_int mask)
+comp_with_mask(void *addr, const void *dest, u_int mask)
 {
 
     if ( /* mask/8 == 0 || */ memcmp(addr, dest, mask / 8) == 0) {
@@ -505,6 +493,52 @@ patricia_search_best2(patricia_tree_t * patricia, prefix_t * prefix,
     return (NULL);
 }
 
+patricia_node_t *
+patricia_search_best3(patricia_tree_t * patricia, const u_char *addr, int bitlen)
+{
+    patricia_node_t *node;
+    patricia_node_t *stack[PATRICIA_MAXBITS + 1];
+    int             cnt = 0;
+
+    if (!patricia || !addr)
+	return NULL;
+
+    if (patricia->head == NULL)
+        return (NULL);
+
+    node = patricia->head;
+
+    while (node->bit < bitlen) {
+
+        if (node->prefix) {
+            stack[cnt++] = node;
+        }
+
+        if (BIT_TEST(addr[node->bit >> 3], 0x80 >> (node->bit & 0x07))) {
+            node = node->r;
+        } else {
+            node = node->l;
+        }
+
+        if (node == NULL)
+            break;
+    }
+
+    if (node && node->prefix)
+        stack[cnt++] = node;
+
+    if (cnt <= 0)
+        return (NULL);
+
+    while (--cnt >= 0) {
+        node = stack[cnt];
+        if (comp_with_mask(prefix_touchar(node->prefix), addr, node->prefix->bitlen)) {
+            return (node);
+        }
+    }
+    return (NULL);
+}
+
 /*
  * if inclusive != 0, "best" may be the given prefix itself 
  */
@@ -544,6 +578,43 @@ patricia_search_all(patricia_tree_t * patricia, prefix_t * prefix, int inclusive
     if (inclusive && node->prefix && node->data &&
         comp_with_mask(prefix_tochar(node->prefix),
                        prefix_tochar(prefix), node->prefix->bitlen)) {
+        results[cnt++] = node;
+    }
+
+    return cnt;
+}
+
+int
+patricia_search_all2(patricia_tree_t * patricia, u_char *addr, int bitlen, patricia_node_t **results, int resultsize)
+{
+    patricia_node_t *node;
+    int             cnt = 0;
+
+    node = patricia->head;
+
+    if (node == NULL)
+        return 0;
+
+    while (node->bit < bitlen && cnt < resultsize) {
+
+        if (node->prefix && node->data &&
+            comp_with_mask(prefix_tochar(node->prefix),
+                           addr, node->prefix->bitlen)) {
+            results[cnt++] = node;
+        }
+
+        if (BIT_TEST(addr[node->bit >> 3], 0x80 >> (node->bit & 0x07))) {
+            node = node->r;
+        } else {
+            node = node->l;
+        }
+
+        if (!node)
+            return cnt;
+    }
+
+    if (node->prefix && node->data &&
+        comp_with_mask(prefix_touchar(node->prefix), addr, node->prefix->bitlen)) {
         results[cnt++] = node;
     }
 
