@@ -30,6 +30,7 @@
 #endif
 
 extern MolochConfig_t        config;
+extern MolochPcapFileHdr_t   pcapFileHeader;
 
 
 typedef struct moloch_output {
@@ -45,42 +46,42 @@ typedef struct moloch_output {
 } MolochDiskOutput_t;
 
 
-static MolochDiskOutput_t   *output;
-static MOLOCH_LOCK_DEFINE(output);
+LOCAL  MolochDiskOutput_t   *output;
+LOCAL  MOLOCH_LOCK_DEFINE(output);
 
-static MolochDiskOutput_t    outputQ;
-static MOLOCH_LOCK_DEFINE(outputQ);
-static MOLOCH_COND_DEFINE(outputQ);
+LOCAL  MolochDiskOutput_t    outputQ;
+LOCAL  MOLOCH_LOCK_DEFINE(outputQ);
+LOCAL  MOLOCH_COND_DEFINE(outputQ);
 
-static MolochIntHead_t       freeOutputBufs;
-static MOLOCH_LOCK_DEFINE(freeOutputBufs);
+LOCAL  MolochIntHead_t       freeOutputBufs;
+LOCAL  MOLOCH_LOCK_DEFINE(freeOutputBufs);
 
-static uint32_t              outputId;
-static char                 *outputFileName;
-static uint64_t              outputFilePos = 0;
-static struct timeval        outputFileTime;
+LOCAL  uint32_t              outputId;
+LOCAL  char                 *outputFileName;
+LOCAL  uint64_t              outputFilePos = 0;
+LOCAL  struct timespec       outputFileTime;
 
 #define MOLOCH_WRITE_NORMAL 0x00
-#define MOLOCH_WRITE_DIRECT 0x01 
+#define MOLOCH_WRITE_DIRECT 0x01
 #define MOLOCH_WRITE_MMAP   0x02
 #define MOLOCH_WRITE_THREAD 0x04
 
-static int                   writeMethod;
-static int                   pageSize;
+LOCAL  int                   writeMethod;
+LOCAL  int                   pageSize;
 
 /******************************************************************************/
-uint32_t writer_disk_queue_length_thread()
+LOCAL uint32_t writer_disk_queue_length_thread()
 {
     int count = DLL_COUNT(mo_, &outputQ);
     return count;
 }
 /******************************************************************************/
-uint32_t writer_disk_queue_length_nothread()
+LOCAL uint32_t writer_disk_queue_length_nothread()
 {
     return DLL_COUNT(mo_, &outputQ);
 }
 /******************************************************************************/
-void writer_disk_alloc_buf(MolochDiskOutput_t *out)
+LOCAL void writer_disk_alloc_buf(MolochDiskOutput_t *out)
 {
     if (writeMethod & MOLOCH_WRITE_THREAD)
         MOLOCH_LOCK(freeOutputBufs);
@@ -100,7 +101,7 @@ void writer_disk_alloc_buf(MolochDiskOutput_t *out)
         MOLOCH_UNLOCK(freeOutputBufs);
 }
 /******************************************************************************/
-void writer_disk_free_buf(MolochDiskOutput_t *out)
+LOCAL void writer_disk_free_buf(MolochDiskOutput_t *out)
 {
     if (writeMethod & MOLOCH_WRITE_THREAD)
         MOLOCH_LOCK(freeOutputBufs);
@@ -119,7 +120,7 @@ void writer_disk_free_buf(MolochDiskOutput_t *out)
 /******************************************************************************/
 /* Only used in non thread mode to write out data
  */
-gboolean writer_disk_output_cb(gint fd, GIOCondition UNUSED(cond), gpointer UNUSED(data))
+LOCAL gboolean writer_disk_output_cb(gint fd, GIOCondition UNUSED(cond), gpointer UNUSED(data))
 {
     if (fd && config.quitting)
         return FALSE;
@@ -176,7 +177,8 @@ gboolean writer_disk_output_cb(gint fd, GIOCondition UNUSED(cond), gpointer UNUS
     // The last write for this fd
     if (out->close) {
         if (filelen) {
-            (void)ftruncate(outputFd, filelen);
+            if (ftruncate(outputFd, filelen) < 0 && config.debug)
+                LOG("Truncate failed");
         } else {
             filelen = lseek(outputFd, 0, SEEK_CUR);
         }
@@ -200,9 +202,10 @@ gboolean writer_disk_output_cb(gint fd, GIOCondition UNUSED(cond), gpointer UNUS
     return DLL_COUNT(mo_, &outputQ) > 0;
 }
 /******************************************************************************/
-void *writer_disk_output_thread(void *UNUSED(arg))
+LOCAL void *writer_disk_output_thread(void *UNUSED(arg))
 {
-    LOG("THREAD %p", (gpointer)pthread_self());
+    if (config.debug)
+        LOG("THREAD %p", (gpointer)pthread_self());
 
     MolochDiskOutput_t *out;
     int outputFd = 0;
@@ -246,7 +249,9 @@ void *writer_disk_output_thread(void *UNUSED(arg))
 
         if (out->close) {
             if (filelen) {
-                (void)ftruncate(outputFd, filelen);
+                ftruncate(outputFd, filelen);
+                if (ftruncate(outputFd, filelen) < 0 && config.debug)
+                    LOG("Truncate failed");
             } else {
                 filelen = lseek(outputFd, 0, SEEK_CUR);
             }
@@ -260,7 +265,7 @@ void *writer_disk_output_thread(void *UNUSED(arg))
     }
 }
 /******************************************************************************/
-void writer_disk_flush(gboolean all)
+LOCAL void writer_disk_flush(gboolean all)
 {
     if (unlikely(config.dryRun || !output)) {
         return;
@@ -308,7 +313,7 @@ void writer_disk_flush(gboolean all)
     output = noutput;
 }
 /******************************************************************************/
-void writer_disk_exit()
+LOCAL void writer_disk_exit()
 {
     writer_disk_flush(TRUE);
     outputFileName = 0;
@@ -324,8 +329,7 @@ void writer_disk_exit()
     }
 }
 /******************************************************************************/
-extern MolochPcapFileHdr_t pcapFileHeader;
-void writer_disk_create(MolochPacket_t * const packet)
+LOCAL void writer_disk_create(MolochPacket_t * const packet)
 {
     outputFileName = moloch_db_create_file(packet->ts.tv_sec, NULL, 0, 0, &outputId);
     outputFilePos = 24;
@@ -334,7 +338,7 @@ void writer_disk_create(MolochPacket_t * const packet)
     output->max = config.pcapWriteSize;
     writer_disk_alloc_buf(output);
     output->pos = 24;
-    gettimeofday(&outputFileTime, 0);
+    clock_gettime(CLOCK_REALTIME_COARSE, &outputFileTime);
 
     output->fileId = outputId;
     memcpy(output->buf, &pcapFileHeader, 24);
@@ -349,8 +353,7 @@ struct pcap_sf_pkthdr {
     uint32_t caplen;		/* length of portion present */
     uint32_t pktlen;		/* length this packet (off wire) */
 };
-void
-writer_disk_write(const MolochSession_t * const UNUSED(session), MolochPacket_t * const packet)
+LOCAL void writer_disk_write(const MolochSession_t * const UNUSED(session), MolochPacket_t * const packet)
 {
     struct pcap_sf_pkthdr hdr;
 
@@ -384,13 +387,12 @@ writer_disk_write(const MolochSession_t * const UNUSED(session), MolochPacket_t 
     MOLOCH_UNLOCK(output);
 }
 /******************************************************************************/
-gboolean 
-writer_disk_file_time_gfunc (gpointer UNUSED(user_data))
+LOCAL gboolean writer_disk_file_time_gfunc (gpointer UNUSED(user_data))
 {
-    static struct timeval tv;
-    gettimeofday(&tv, 0);
+    struct timespec ts;
+    clock_gettime(CLOCK_REALTIME_COARSE, &ts);
 
-    if (outputFileName && outputFilePos > 24 && (tv.tv_sec - outputFileTime.tv_sec) >= config.maxFileTimeM*60) {
+    if (outputFileName && outputFilePos > 24 && (ts.tv_sec - outputFileTime.tv_sec) >= config.maxFileTimeM*60) {
         writer_disk_flush(TRUE);
         outputFileName = 0;
     }
@@ -423,7 +425,7 @@ void writer_disk_init(char *name)
 #endif
 
     if (writeMethod & MOLOCH_WRITE_THREAD) {
-        g_thread_new("moloch-output", &writer_disk_output_thread, NULL);
+        g_thread_unref(g_thread_new("moloch-output", &writer_disk_output_thread, NULL));
     }
 
     if ((writeMethod & MOLOCH_WRITE_DIRECT) && sizeof(off_t) == 4 && config.maxFileSizeG > 2)
