@@ -28,41 +28,45 @@
 extern MolochConfig_t        config;
 
 
-LOCAL GHashTable           *filePtr2Id;
-MOLOCH_LOCK_DEFINE(filePtr2Id);
+LOCAL MOLOCH_LOCK_DEFINE(filePtr2Id);
+
+extern char                *readerFileName[256];
+extern uint32_t             readerOutputIds[256];
 
 /******************************************************************************/
-uint32_t writer_inplace_queue_length()
+LOCAL uint32_t writer_inplace_queue_length()
 {
     return 0;
 }
 /******************************************************************************/
-void writer_inplace_flush(gboolean UNUSED(all))
+LOCAL void writer_inplace_exit()
 {
 }
 /******************************************************************************/
-void writer_inplace_exit()
-{
-}
-/******************************************************************************/
-long writer_inplace_create(MolochPacket_t * const packet)
+LOCAL long writer_inplace_create(MolochPacket_t * const packet)
 {
     struct stat st;
+    const char *readerName = readerFileName[packet->readerPos];
 
-    stat(packet->readerName, &st);
+    stat(readerName, &st);
 
     uint32_t outputId;
-    char *filename = moloch_db_create_file(packet->ts.tv_sec, packet->readerName, st.st_size, 1, &outputId);
-    g_free(filename);
-    g_hash_table_insert(filePtr2Id, packet->readerName, (gpointer)(long)outputId);
+    if (config.pcapReprocess) {
+        moloch_db_file_exists(readerName, &outputId);
+    } else {
+        char *filename = moloch_db_create_file(packet->ts.tv_sec, readerName, st.st_size, !config.noLockPcap, &outputId);
+        g_free(filename);
+    }
+    readerOutputIds[packet->readerPos] = outputId;
     return outputId;
 }
 
 /******************************************************************************/
-void writer_inplace_write(const MolochSession_t * const UNUSED(session), MolochPacket_t * const packet)
+LOCAL void writer_inplace_write(const MolochSession_t * const UNUSED(session), MolochPacket_t * const packet)
 {
+    // Need to lock since multiple packet threads for the same readerPos are running and only want to create once
     MOLOCH_LOCK(filePtr2Id);
-    long outputId = (long)g_hash_table_lookup(filePtr2Id, packet->readerName);
+    long outputId = readerOutputIds[packet->readerPos];
     if (!outputId)
         outputId = writer_inplace_create(packet);
     MOLOCH_UNLOCK(filePtr2Id);
@@ -71,7 +75,7 @@ void writer_inplace_write(const MolochSession_t * const UNUSED(session), MolochP
     packet->writerFilePos = packet->readerFilePos;
 }
 /******************************************************************************/
-void writer_inplace_write_dryrun(const MolochSession_t * const UNUSED(session), MolochPacket_t * const packet)
+LOCAL void writer_inplace_write_dryrun(const MolochSession_t * const UNUSED(session), MolochPacket_t * const packet)
 {
     packet->writerFilePos = packet->readerFilePos;
 }
@@ -84,6 +88,4 @@ void writer_inplace_init(char *UNUSED(name))
         moloch_writer_write    = writer_inplace_write_dryrun;
     else
         moloch_writer_write    = writer_inplace_write;
-
-    filePtr2Id = g_hash_table_new(g_direct_hash, g_direct_equal);
 }
