@@ -4,13 +4,13 @@
  * addUser.js <user id> <user friendly name> <password> [-noweb] [-admin]
  *
  * Copyright 2012-2016 AOL Inc. All rights reserved.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this Software except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -38,10 +38,12 @@ function help() {
   console.log("  --remove              Can remove data (scrub, delete tags)");
   console.log("  --webauth             Can auth using the web auth header or password");
   console.log("  --webauthonly         Can auth using the web auth header only, password ignored");
+  console.log("  --packetSearch        Can create a packet search job (hunt)");
   console.log("");
   console.log("Config Options:");
   console.log("  -c <config file>      Config file to use");
   console.log("  -n <node name>        Node name section to use in config file");
+  console.log("  --insecure            Allow insecure HTTPS");
 
   process.exit(0);
 }
@@ -63,6 +65,8 @@ function main() {
     emailSearch: false,
     createEnabled: false,
     removeEnabled: false,
+    packetSearch: false,
+    welcomeMsgNum: 0,
     settings: {}
   };
 
@@ -84,7 +88,7 @@ function main() {
     case "--apionly":
       nuser.webEnabled = false;
       break;
-      
+
     case "--webauthonly":
     case "-webauthonly":
       nuser.passStore = Config.pass2store(process.argv[2], crypto.randomBytes(48));
@@ -104,13 +108,18 @@ function main() {
       i++;
       break;
 
+    case "--packetSearch":
+    case "-packetSearch":
+      nuser.packetSearch = true;
+      break;
+
     default:
       console.log("Unknown option", process.argv[i]);
       help();
     }
   }
 
-  Db.setUser(process.argv[2], nuser, function(err, info) {
+  Db.setUser(process.argv[2], nuser, (err, info) => {
     if (err) {
       console.log("Elastic search error", err);
     } else {
@@ -126,5 +135,9 @@ if (process.argv.length < 5) {
 
 Db.initialize({host : escInfo,
                prefix: Config.get("prefix", ""),
+               esClientKey: Config.get("esClientKey", null),
+               esClientCert: Config.get("esClientCert", null),
+               esClientKeyPass: Config.get("esClientKeyPass", null),
+               insecure: Config.insecure,
                usersHost: Config.get("usersElasticsearch"),
                usersPrefix: Config.get("usersPrefix")}, main);
