@@ -27,22 +27,32 @@ try {
 var Config         = require('./config.js'),
     express        = require('express'),
     async          = require('async'),
-    sprintf        = require('./public/sprintf.js'),
-    os             = require('os'),
-    util           = require('util'),
     URL            = require('url'),
     ESC            = require('elasticsearch'),
     http           = require('http'),
-    https          = require('https');
+    https          = require('https'),
+    fs             = require('fs');
 } catch (e) {
   console.log ("ERROR - Couldn't load some dependancies, maybe need to 'npm update' inside viewer directory", e);
   process.exit(1);
 }
 
+var esSSLOptions = {rejectUnauthorized: !Config.insecure};
+var esClientKey = Config.get("esClientKey");
+var esClientCert = Config.get("esClientCert");
+if(esClientKey) {
+  esSSLOptions.key = fs.readFileSync(esClientKey);
+  esSSLOptions.cert = fs.readFileSync(esClientCert);
+  var esClientKeyPass = Config.get("esClientKeyPass");
+  if(esClientKeyPass) {
+     esSSLOptions.passphrase = esClientKeyPass;
+  }
+}
+
 var clients = {};
 var nodes = [];
 var httpAgent  =  new http.Agent({keepAlive: true, keepAliveMsecs:5000, maxSockets: 100});
-var httpsAgent =  new https.Agent({keepAlive: true, keepAliveMsecs:5000, maxSockets: 100});
+var httpsAgent =  new https.Agent(Object.assign({keepAlive: true, keepAliveMsecs:5000, maxSockets: 100}, esSSLOptions));
 
 function hasBody(req) {
   var encoding = 'transfer-encoding' in req.headers;
@@ -62,8 +72,8 @@ function saveBody (req, res, next) {
   // parse
   var buf = '';
   req.setEncoding('utf8');
-  req.on('data', function(chunk){ buf += chunk; });
-  req.on('end', function(){
+  req.on('data', (chunk) => { buf += chunk; });
+  req.on('end', () => {
     req.body = buf;
     next();
   });
@@ -89,8 +99,7 @@ app.use(function(req, res, next) {
 
 function node2Url(node) {
   var url = node.split(',')[0];
-  if (url.match(/^http/))
-    return url
+  if (url.match(/^http/)) {return url;}
   return "http://" + url;
 }
 
@@ -109,7 +118,7 @@ function node2Prefix(node) {
 }
 
 function simpleGather(req, res, bodies, doneCb) {
-  async.map(nodes, function (node, asyncCb) {
+  async.map(nodes, (node, asyncCb) => {
     var result = "";
     var url = node2Url(node) + req.url;
     var prefix = node2Prefix(node);
@@ -125,14 +134,14 @@ function simpleGather(req, res, bodies, doneCb) {
       info.agent  = httpAgent;
       client = http;
     }
-    var preq = client.request(info, function(pres) {
-      pres.on('data', function (chunk) {
+    var preq = client.request(info, (pres) => {
+      pres.on('data', (chunk) => {
         result += chunk.toString();
       });
-      pres.on('end', function () {
+      pres.on('end', () => {
         if (result.length) {
-          result = result.replace(new RegExp('(index":\s*|[,{]|  )"' + prefix + "(sessions|stats|tags|dstats|sequence|files|users)", "g"), "$1\"MULTIPREFIX_$2");
-          result = result.replace(new RegExp('(index":\s*)"' + prefix + "(fields_v1)\"", "g"), "$1\"MULTIPREFIX_$2\"");
+          result = result.replace(new RegExp('(index":\s*|[,{]|  )"' + prefix + "(sessions2|sessions|stats|dstats|sequence|files|users|history)", "g"), "$1\"MULTIPREFIX_$2");
+          result = result.replace(new RegExp('(index":\s*)"' + prefix + "(fields_v[1-4])\"", "g"), "$1\"MULTIPREFIX_$2\"");
           result = JSON.parse(result);
         } else {
           result = {};
@@ -141,6 +150,7 @@ function simpleGather(req, res, bodies, doneCb) {
         asyncCb(null, result);
       });
     });
+    preq.setHeader('content-type', "application/json");
     if (req._body) {
       if (bodies && bodies[node]) {
         preq.end(bodies[node]);
@@ -148,7 +158,7 @@ function simpleGather(req, res, bodies, doneCb) {
         preq.end(req.body);
       }
     }
-    preq.on('error', function (e) {
+    preq.on('error', (e) => {
       console.log("Request error with node", node, e);
     });
     preq.end();
@@ -161,10 +171,6 @@ function shallowCopy(obj1, obj2) {
   }
 }
 
-function deepCopy(obj) {
-  return JSON.parse(JSON.stringify(obj));
-}
-
 function shallowAdd(obj1, obj2) {
   for (var attrname in obj2) {
     if (typeof obj2[attrname] === "number") {
@@ -174,7 +180,7 @@ function shallowAdd(obj1, obj2) {
 }
 
 function simpleGatherCopy(req, res) {
-  simpleGather(req, res, null, function(err, results) {
+  simpleGather(req, res, null, (err, results) => {
     var obj = results[0];
     for (var i = 1; i < results.length; i++) {
       shallowCopy(obj.nodes, results[i].nodes);
@@ -184,7 +190,7 @@ function simpleGatherCopy(req, res) {
 }
 
 function simpleGatherAdd(req, res) {
-  simpleGather(req, res, null, function(err, results) {
+  simpleGather(req, res, null, (err, results) => {
     var obj = results[0];
     for (var i = 1; i < results.length; i++) {
       shallowAdd(obj, results[i]);
@@ -194,15 +200,23 @@ function simpleGatherAdd(req, res) {
   });
 }
 
+function simpleGatherFirst(req, res) {
+  simpleGather(req, res, null, (err, results) => {
+    res.send(results[0]);
+  });
+}
+
 app.get("/_cluster/nodes/stats", simpleGatherCopy);
+app.get("/_nodes", simpleGatherCopy);
 app.get("/_nodes/stats", simpleGatherCopy);
 app.get("/_nodes/stats/:kinds", simpleGatherCopy);
 app.get("/_cluster/health", simpleGatherAdd);
 
 app.get("/:index/_aliases", simpleGatherCopy);
+app.get("/:index/_alias", simpleGatherCopy);
 
-app.get("/:index/_status", function(req, res) {
-  simpleGather(req, res, null, function(err, results) {
+app.get("/:index/_status", (req, res) => {
+  simpleGather(req, res, null, (err, results) => {
     var obj = results[0];
     for (var i = 1; i < results.length; i++) {
       for (var index in results[i].indices) {
@@ -218,9 +232,9 @@ app.get("/:index/_status", function(req, res) {
   });
 });
 
-app.get("/:index/_stats", function(req, res) {
-  simpleGather(req, res, null, function(err, results) {
-    //console.log("DEBUG - _stats results", util.inspect(results, false, 50));
+app.get("/:index/_stats", (req, res) => {
+  simpleGather(req, res, null, (err, results) => {
+    //console.log("DEBUG - _stats results", JSON.stringify(results, null, 2));
     var obj = results[0];
     for (var i = 1; i < results.length; i++) {
       for (var index in results[i].indices) {
@@ -236,11 +250,13 @@ app.get("/:index/_stats", function(req, res) {
   });
 });
 
-app.get("/MULTIPREFIX_dstats/version/version", function(req, res) {
-  simpleGather(req, res, null, function(err, results) {
+app.get("/_template/MULTIPREFIX_sessions2_template", (req, res) => {
+  simpleGather(req, res, null, (err, results) => {
+    //console.log("DEBUG -", JSON.stringify(results, null, 2));
+
     var obj = results[0];
     for (var i = 1; i < results.length; i++) {
-      if (results[i]._source.version < obj._source.version) {
+      if (results[i].MULTIPREFIX_sessions2_template.mappings.session._meta.molochDbVersion < obj.MULTIPREFIX_sessions2_template.mappings.session._meta.molochDbVersion) {
         obj = results[i];
       }
     }
@@ -248,14 +264,20 @@ app.get("/MULTIPREFIX_dstats/version/version", function(req, res) {
   });
 });
 
-app.get("/users/user/:user", function(req, res) {
-  clients[nodes[0]].get({index: "users", type: "user", id: req.params.user}, function(err, result) {
+app.get("/users/user/:user", (req, res) => {
+  clients[nodes[0]].get({index: "users", type: "user", id: req.params.user}, (err, result) => {
     res.send(result);
   });
 });
 
-app.get("/:index/:type/_search", function(req, res) {
-  simpleGather(req, res, null, function(err, results) {
+app.post("/users/user/:user", (req, res) => {
+  clients[nodes[0]].index({index: "users", type: "user", id: req.params.user, body: req.body, refresh: true}, (err, result) => {
+    res.send(result);
+  });
+});
+
+app.get("/:index/:type/_search", (req, res) => {
+  simpleGather(req, res, null, (err, results) => {
     var obj = results[0];
     for (var i = 1; i < results.length; i++) {
       if (results[i].error) {
@@ -268,47 +290,8 @@ app.get("/:index/:type/_search", function(req, res) {
   });
 });
 
-app.get("/MULTIPREFIX_sessions-*/:type/:id", function(req, res) {
-  function fixTags(node, container, field, doneCb) {
-    if (!container || !container[field]) {
-      return doneCb(null);
-    }
-
-    async.map(container[field], function (item, cb) {
-      tagIdToName(node, item, function (name) {
-        cb(null, name);
-      });
-    },
-    function(err, results) {
-      container[field] = results;
-      doneCb(err);
-    });
-  }
-
-  simpleGather(req, res, null, function(err, results) {
-    for (var i = 0; i < results.length; i++) {
-      if (results[i].found) {
-        async.parallel([
-          function(parallelCb) {
-            fixTags(results[i]._node, results[i]._source, "ta", parallelCb);
-          },
-          function(parallelCb) {
-            fixTags(results[i]._node, results[i]._source, "hh1", parallelCb);
-          },
-          function(parallelCb) {
-            fixTags(results[i]._node, results[i]._source, "hh2", parallelCb);
-          }], function () {
-            return res.send(results[i]);
-          });
-        return;
-      }
-    }
-    res.send(results[0]);
-  });
-});
-
 app.get("/:index/:type/:id", function(req, res) {
-  simpleGather(req, res, null, function(err, results) {
+  simpleGather(req, res, null, (err, results) => {
     for (var i = 0; i < results.length; i++) {
       if (results[i].found) {
         return res.send(results[i]);
@@ -318,19 +301,23 @@ app.get("/:index/:type/:id", function(req, res) {
   });
 });
 
+app.get("/_cluster/settings", function(req, res) {
+  res.send({persistent: {}, transient: {}});
+});
+
 
 app.head(/^\/$/, function(req, res) {
   res.send("");
 });
 
 app.get(/^\/$/, function(req, res) {
-  simpleGather(req, res, null, function(err, results) {
+  simpleGather(req, res, null, (err, results) => {
     res.send(results[0]);
   });
 });
 
 app.get(/./, function(req, res) {
-  simpleGather(req, res, null, function(err, results) {
+  simpleGather(req, res, null, (err, results) => {
     console.log("UNKNOWN", req.method, req.url, results);
   });
 
@@ -355,9 +342,9 @@ function facet2Arr(facet, type) {
   }
 
   if (type === "histogram") {
-    arr = arr.sort(function(a,b) {return a.key - b.key;});
+    arr = arr.sort((a,b) => {return a.key - b.key;});
   } else {
-    arr = arr.sort(function(a,b) {return b.count - a.count;});
+    arr = arr.sort((a,b) => {return b.count - a.count;});
   }
   return arr;
 }
@@ -424,9 +411,9 @@ function agg2Arr(agg, type) {
   }
 
   if (type === "histogram") {
-    arr = arr.sort(function(a,b) {return a.key - b.key;});
+    arr = arr.sort((a,b) => {return a.key - b.key;});
   } else {
-    arr = arr.sort(function(a,b) {
+    arr = arr.sort((a,b) => {
       if (b.doc_count !== a.doc_count) {
         return b.doc_count - a.doc_count;
       }
@@ -470,42 +457,6 @@ function aggAdd(obj1, obj2) {
   }
 }
 
-// Tags Fun
-
-var tags = {};
-
-function tagNameToId(node, name, cb) {
-  if (tags[node].tagName2Id[name]) {
-    return setImmediate(cb, tags[node].tagName2Id[name]);
-  }
-
-  clients[node].get({index: 'tags', type: 'tag', id: name}, function(err, tdata) {
-    if (!err && (tdata.found)) {
-      tags[node].tagName2Id[name] = tdata._source.n;
-      tags[node].tagId2Name[tdata._source.n] = name;
-      return cb (tags[node].tagName2Id[name]);
-    }
-    return cb(-1);
-  });
-}
-
-function tagIdToName (node, id, cb) {
-  if (tags[node].tagId2Name[id]) {
-    return setImmediate(cb, tags[node].tagId2Name[id]);
-  }
-
-  var query = {query: {term: {n:id}}};
-  clients[node].search({index: node2Prefix(node) + 'tags', type: 'tag', body: query}, function(err, tdata) {
-    if (!err && tdata.hits.hits[0]) {
-      tags[node].tagId2Name[id] = tdata.hits.hits[0]._id;
-      tags[node].tagName2Id[tdata.hits.hits[0]._id] = id;
-      return cb(tags[node].tagId2Name[id]);
-    }
-
-    return cb(null);
-  });
-}
-
 function fixQuery(node, body, doneCb) {
   body = JSON.parse(body);
 
@@ -513,76 +464,19 @@ function fixQuery(node, body, doneCb) {
   if (body.size) {
     body.size = (+body.size) + ((+body.from) || 0);
   }
-  body.from = 0;
+
+  delete body.from;
 
   var outstanding = 0;
   var finished = 0;
   var err = null;
 
+  var convert;
+
   function process(parent, obj, item) {
     var query;
 
-    if (item.match(/^(ta|hh1|hh2)$/) && (typeof obj[item] === "string" || Array.isArray(obj[item]))) {
-      if (obj[item].indexOf("*") !== -1) {
-        delete parent.wildcard;
-        outstanding++;
-        if (item === "ta") {
-          query = {bool: {must: {wildcard: {_id: obj[item]}},
-                          must_not: {wildcard: {_id: "http:header:*"}}
-                         }
-                  };
-        } else {
-          query = {wildcard: {_id: "http:header:" + obj[item].toLowerCase()}};
-        }
-        clients[node].search({index: node2Prefix(node) + 'tags', type: 'tag', size:500, _source:["id", "n"], body: {query: query}}, function(err, result) {
-          var terms = [];
-          result.hits.hits.forEach(function (hit) {
-            terms.push(hit._source.n);
-          });
-          parent.terms = {};
-          parent.terms[item] = terms;
-          outstanding--;
-          if (finished && outstanding === 0) {
-            doneCb(err, body);
-          }
-        });
-      } else if (Array.isArray(obj[item])) {
-        outstanding++;
-
-        async.map(obj[item], function(str, cb) {
-          var tag = (item !== "ta" && str.indexOf("http:header:") !== 0?"http:header:" + str.toLowerCase():str);
-          tagNameToId(node, tag, function (id) {
-            if (id === null) {
-              cb(null, -1);
-            } else {
-              cb(null, id);
-            }
-          });
-        },
-        function (err, results) {
-          outstanding--;
-          obj[item] = results;
-          if (finished && outstanding === 0) {
-            doneCb(err, body);
-          }
-        });
-      } else {
-        outstanding++;
-        var tag = (item !== "ta" && obj[item].indexOf("http:header:") !== 0?"http:header:" + obj[item].toLowerCase():obj[item]);
-
-        tagNameToId(node, tag, function (id) {
-          outstanding--;
-          if (id === null) {
-            err = "Tag '" + tag + "' not found";
-          } else {
-            obj[item] = id;
-          }
-          if (finished && outstanding === 0) {
-            doneCb(err, body);
-          }
-        });
-      }
-    } else if (item === "fileand" && typeof obj[item] === "string") {
+    if (item === "fileand" && typeof obj[item] === "string") {
       var name = obj.fileand;
       delete obj.fileand;
       outstanding++;
@@ -594,11 +488,11 @@ function fixQuery(node, body, doneCb) {
       } else {
         query = {query: {term: {name: name}}};
       }
-      clients[node].search({index: node2Prefix(node) + 'files', type: 'file', size:500, body: query}, function(err, result) {
+      clients[node].search({index: node2Prefix(node) + 'files', type: 'file', size:500, body: query}, (err, result) => {
         outstanding--;
         obj.bool = {should: []};
-        result.hits.hits.forEach(function (file) {
-          obj.bool.should.push({bool: {must:[{term: {no: file._source.node}}, {term: {fs: file._source.num}}]}});
+        result.hits.hits.forEach((file) => {
+          obj.bool.should.push({bool: {must:[{term: {node: file._source.node}}, {term: {fileId: file._source.num}}]}});
         });
         if (obj.bool.should.length === 0) {
           err = "No matching files found";
@@ -612,11 +506,11 @@ function fixQuery(node, body, doneCb) {
     }
   }
 
-  function convert(parent, obj) {
+  convert = function (parent, obj) {
     for (var item in obj) {
       process(parent, obj, item);
     }
-  }
+  };
 
   convert(null, body);
   if (outstanding === 0) {
@@ -626,69 +520,6 @@ function fixQuery(node, body, doneCb) {
   finished = 1;
 }
 
-function fixResult(node, result, doField, doneCb) {
-  function facetTags(container, field, doneCb) {
-    if (!container || !container[field]) {
-      return doneCb(null);
-    }
-
-    async.map(container[field].terms, function (item, cb) {
-      tagIdToName(node, item.term, function (name) {
-        item.term = name;
-        cb(null, item);
-      });
-    },
-    function(err, tagsResults) {
-      container[field].terms = tagsResults;
-      doneCb(err);
-    });
-  }
-
-  function aggTags(container, field, doneCb) {
-    if (!container || !container[field]) {
-      return doneCb(null);
-    }
-
-    async.map(container[field].buckets, function (item, cb) {
-      tagIdToName(node, item.key, function (name) {
-        item.key = name;
-        cb(null, item);
-      });
-    },
-    function(err, tagsResults) {
-      container[field].buckets = tagsResults;
-      doneCb(err);
-    });
-  }
-
-  async.parallel([
-    function(parallelCb) {
-      facetTags(result.facets, "ta", parallelCb);
-    },
-    function(parallelCb) {
-      facetTags(result.facets, "hh1", parallelCb);
-    },
-    function(parallelCb) {
-      facetTags(result.facets, "hh2", parallelCb);
-    },
-    function(parallelCb) {
-      aggTags(result.aggregations, "ta", parallelCb);
-    },
-    function(parallelCb) {
-      aggTags(result.aggregations, "hh1", parallelCb);
-    },
-    function(parallelCb) {
-      aggTags(result.aggregations, "hh2", parallelCb);
-    },
-    function(parallelCb) {
-      if (!doField) {
-        return parallelCb();
-      }
-      aggTags(result.aggregations, "field", parallelCb);
-    }], function () {
-      return setImmediate(doneCb);
-    });
-}
 function combineResults(obj, result) {
   if (!result.hits) {
     console.log("NO RESULTS", result);
@@ -699,6 +530,9 @@ function combineResults(obj, result) {
   obj.hits.missing += result.hits.missing;
   obj.hits.other += result.hits.other;
   if (result.hits.hits) {
+    for (var i = 0; i < result.hits.hits.length; i++) {
+      result.hits.hits[i]._node = result._node;
+    }
     obj.hits.hits = obj.hits.hits.concat(result.hits.hits);
   }
   if (obj.facets) {
@@ -720,7 +554,7 @@ function sortResults(search, obj) {
       sortorder[i] = search.sort[i][Object.keys(search.sort[i])[0]].order === "asc"? 1:-1;
     }
 
-    obj.hits.hits = obj.hits.hits.sort(function(a, b) {
+    obj.hits.hits = obj.hits.hits.sort((a, b) => {
       for (var i = 0; i < a.sort.length; i++) {
         if (a.sort[i] === b.sort[i]) {
           continue;
@@ -765,39 +599,8 @@ function newResult(search) {
   return result;
 }
 
-// Only tags search is for auto complete so unique the results
-app.post("/MULTIPREFIX_tags/tag/_search", function(req, res) {
-  var search = JSON.parse(req.body);
-
-  simpleGather(req, res, null, function(err, results) {
-    async.each(results, function (result, asyncCb) {
-      fixResult(result._node, result, false, asyncCb);
-    }, function (err) {
-      var tags = {};
-      for (var i = 0; i < results.length; i++) {
-        if (results[i].error || !results[i].hits) {
-          console.log("Issue with tag results", results[i].error);
-          continue;
-        }
-        for (var h = 0; h < results[i].hits.hits.length; h++) {
-          tags[results[i].hits.hits[h]._id] = results[i].hits.hits[h];
-        }
-      }
-      var obj = results[0];
-      obj.hits.hits = [];
-
-      for (var tag in tags) {
-        obj.hits.hits.push(tags[tag]);
-      }
-
-      res.send(obj);
-    });
-  });
-
-});
-
 app.post("/MULTIPREFIX_fields/field/_search", function(req, res) {
-  simpleGather(req, res, null, function(err, results) {
+  simpleGather(req, res, null, (err, results) => {
     var obj = {
       hits: {
         total: 0,
@@ -829,43 +632,33 @@ app.post("/MULTIPREFIX_fields/field/_search", function(req, res) {
 app.post("/:index/:type/_search", function(req, res) {
   var bodies = {};
   var search = JSON.parse(req.body);
-  //console.log("DEBUG - INCOMING SEARCH", util.inspect(search, false, 50));
+  //console.log("DEBUG - INCOMING SEARCH", JSON.stringify(search, null, 2));
 
-  var doField = search.aggregations &&
-                search.aggregations.field &&
-                search.aggregations.field.terms &&
-                search.aggregations.field.terms.field.match(/^(ta|hh1|hh2)$/);
-
-  async.each(nodes, function (node, asyncCb) {
-    fixQuery(node, req.body, function(err, body) {
-      //console.log("DEBUG - OUTGOING SEARCH", node, util.inspect(body, false, 50));
+  async.each(nodes, (node, asyncCb) => {
+    fixQuery(node, req.body, (err, body) => {
+      //console.log("DEBUG - OUTGOING SEARCH", node, JSON.stringify(body, null, 2));
       bodies[node] = JSON.stringify(body);
       asyncCb(null);
     });
-  }, function (err) {
-    simpleGather(req, res, bodies, function(err, results) {
-      async.each(results, function (result, asyncCb) {
-        //console.log("DEBUG - RESULT", util.inspect(result, false, 50));
-        fixResult(result._node, result, doField, asyncCb);
-      }, function (err) {
-        var obj = newResult(search);
+  }, (err) => {
+    simpleGather(req, res, bodies, (err, results) => {
+      var obj = newResult(search);
 
-        for (var i = 0; i < results.length; i++) {
-          combineResults(obj, results[i]);
-        }
+      for (var i = 0; i < results.length; i++) {
+        combineResults(obj, results[i]);
+      }
 
-        if (obj.facets) {
-          facetConvert2Arr(obj.facets);
-        }
+      if (obj.facets) {
+        facetConvert2Arr(obj.facets);
+      }
 
-        if (obj.aggregations) {
-          aggConvert2Arr(obj.aggregations);
-        }
+      if (obj.aggregations) {
+        aggConvert2Arr(obj.aggregations);
+      }
 
-        sortResults(search, obj);
+      sortResults(search, obj);
 
-        res.send(obj);
-      });
+      res.send(obj);
     });
   });
 });
@@ -874,63 +667,77 @@ function msearch(req, res) {
   var lines = req.body.split(/[\r\n]/);
   var bodies = {};
 
-  async.each(nodes, function (node, nodeCb) {
+  async.each(nodes, (node, nodeCb) => {
     var nlines = [];
-    async.eachSeries(lines, function (line, lineCb) {
+    async.eachSeries(lines, (line, lineCb) => {
       if (line === "{}" || line === "") {
         nlines.push("{}");
         return lineCb();
       }
-      fixQuery(node, line, function(err, body) {
+      fixQuery(node, line, (err, body) => {
         nlines.push(JSON.stringify(body));
         lineCb();
       });
-    }, function(err) {
-      bodies[node] = nlines.join("\n");
+    }, (err) => {
+      bodies[node] = nlines.join("\n") + "\n";
       var prefix = node2Prefix(node);
       bodies[node] = bodies[node].replace(/MULTIPREFIX_/g, prefix);
       nodeCb();
     });
-  }, function(err) {
-    var responses = [];
-    simpleGather(req, res, bodies, function(err, results) {
-      async.eachSeries(results, function(result, resultCb) {
-        async.eachSeries(result.responses, function(response, responseCb) {
-          fixResult(result._node, response, false, responseCb);
-        }, function(err) {
-          resultCb();
-        });
-      }, function(err) {
-        var obj = {responses:[]};
-        for (var h = 0; h < results[0].responses.length; h++) {
-          obj.responses[h] = newResult(JSON.parse(lines[h*2+1]));
+  }, (err) => {
+    simpleGather(req, res, bodies, (err, results) => {
+      var obj = {responses:[]};
+      for (var h = 0; h < results[0].responses.length; h++) {
+        obj.responses[h] = newResult(JSON.parse(lines[h*2+1]));
 
-          for (var i = 0; i < results.length; i++) {
-            combineResults(obj.responses[h], results[i].responses[h]);
-          }
-
-          if (obj.responses[h].facets) {
-            facetConvert2Arr(obj.responses[h].facets);
-          }
-
-          if (obj.responses[h].aggregations) {
-            aggConvert2Arr(obj.responses[h].aggregations);
-          }
-
-          sortResults(JSON.parse(lines[h*2+1]), obj.responses[h]);
+        for (var i = 0; i < results.length; i++) {
+          combineResults(obj.responses[h], results[i].responses[h]);
         }
 
-        res.send(obj);
-      });
+        if (obj.responses[h].facets) {
+          facetConvert2Arr(obj.responses[h].facets);
+        }
+
+        if (obj.responses[h].aggregations) {
+          aggConvert2Arr(obj.responses[h].aggregations);
+        }
+
+        sortResults(JSON.parse(lines[h*2+1]), obj.responses[h]);
+      }
+
+      res.send(obj);
     });
   });
 }
+
+app.post("/:index/:type/:id/_update", function(req, res) {
+  var body = JSON.parse(req.body);
+  if (body._node && clients[body._node]) {
+
+    var node = body._node;
+    delete body._node;
+
+    var prefix = node2Prefix(node);
+    var index = req.params.index.replace(/MULTIPREFIX_/g, prefix);
+
+    clients[node].update({index: index, type: req.params.type, id: req.params.id, body: body}, (err, result) => {
+      return res.send(result);
+    });
+  } else {
+    console.log ('ERROR - body of the request does not contain _node field', req.method, req.url, req.body);
+    return res.end();
+  }
+});
+
+app.post("/:index/history", simpleGatherFirst);
 
 app.post("/:index/:type/_msearch", msearch);
 app.post("/_msearch", msearch);
 
 app.get("/:index/_count", simpleGatherAdd);
 app.post("/:index/_count", simpleGatherAdd);
+app.get("/:index/:type/_count", simpleGatherAdd);
+app.post("/:index/:type/_count", simpleGatherAdd);
 
 
 if (Config.get("regressionTests")) {
@@ -954,21 +761,25 @@ if (nodes.length === 0 || nodes[0] === "") {
   process.exit(1);
 }
 
-nodes.forEach(function(node) {
-  tags[node] = {tagName2Id: {}, tagId2Name: {}};
-
+// First connect
+nodes.forEach((node) => {
   clients[node] = new ESC.Client({
     host: node.split(",")[0],
-    apiVersion: "1.2",
-    requestTimeout: 300000
+    apiVersion: "6.6",
+    requestTimeout: 300000,
+    keepAlive: true,
+    ssl: esSSLOptions
   });
+});
 
-  clients[node].info(function(err,data) {
+// Now check version numbers
+nodes.forEach((node) => {
+  clients[node].info((err,data) => {
     if (err) {
       console.log(err);
     }
-    if (data.version.number.match(/^(1.[01]|0)/)) {
-      console.log("ES", data.version.number,"is not supported, upgrade to 1.2.x or 1.3.x:", node);
+    if (data.version.number.match(/^([012345])/)) {
+      console.log("ES", data.version.number, "is not supported, upgrade to >= 6.6.x:", node);
       process.exit();
     }
   });
@@ -977,4 +788,4 @@ nodes.forEach(function(node) {
 console.log(nodes);
 
 console.log("Listen on ", Config.get("multiESPort", "8200"));
-var server = http.createServer(app).listen(Config.get("multiESPort", "8200"), Config.get("multiESHost", undefined));
+http.createServer(app).listen(Config.get("multiESPort", "8200"), Config.get("multiESHost", undefined));
